The HPJ bias-corrected estimation of the moments

The `hpjmoment()` function enables to implement the HPJ bias-corrected estimation of the moments of the heterogeneous mean, the heterogeneous autocovariance, and the heterogeneous autocorrelation. The method is developed by Okui and Yanagi (2019). For more details, see the package vignette with `vignette("panelhetero")`.

Usage

hpjmoment(data, acov_order = 0, acor_order = 1, R = 1000)

Arguments

data

A matrix of panel data. Each row corresponds to individual time series.

acov_order

A non-negative integer of the order of autocovariance. Default is 0.

acor_order

A positive integer of the order of autocorrelation Default is 1.

R

A positive integer of the number of bootstrap repetitions. Default is 1000.

Value

A list that contains the following elements.

estimate

A vector of the parameter estimates

se

A vector of the standard errors

ci

A matrix of the 95 percent confidence intervals

quantity

A matrix of the estimated heterogeneous quantities

acov_order

The order of autocovariance

acor_order

The order of autocovariance

N

The number of cross-sectional units

S

The length of time series

R

The number of bootstrap repetitions

References

Okui, R. and Yanagi, T., 2019. Panel data analysis with heterogeneous dynamics. Journal of Econometrics, 212(2), pp.451-475.

Examples

data <- panelhetero::simulation(N = 300, S = 50)
panelhetero::hpjmoment(data = data)
#> $estimate
#>         E(mean)         E(acov)         E(acor)       var(mean)       var(acov) 
#>    -0.001669016     0.381211302     0.284337641     1.057968334     0.052445567 
#>       var(acor) cor(mean, acov) cor(mean, acor) cor(acov, acor) 
#>     0.020415568    -0.044154912     0.029575389    -0.442808471 
#> 
#> $se
#>         E(mean)         E(acov)         E(acor)       var(mean)       var(acov) 
#>     0.056114574     0.014310220     0.011962023     0.083356232     0.005978863 
#>       var(acor) cor(mean, acov) cor(mean, acor) cor(acov, acor) 
#>     0.003496228     0.067460493     0.075006044     0.070148019 
#> 
#> $ci
#>                 95% CI lower 95% CI upper
#> E(mean)          -0.11026037   0.10391502
#> E(acov)           0.35390532   0.40972134
#> E(acor)           0.26026784   0.30728142
#> var(mean)         0.89409297   1.22236887
#> var(acov)         0.04079657   0.06438198
#> var(acor)         0.01344761   0.02734982
#> cor(mean, acov)  -0.17872698   0.08781995
#> cor(mean, acor)  -0.11261040   0.18397299
#> cor(acov, acor)  -0.57129274  -0.29792043
#> 
#> $quantity
#>                mean autocovariance autocorrelation
#>   [1,] -0.914098398     0.96708653    0.3583601751
#>   [2,] -0.006622441     0.09572946    0.0854428258
#>   [3,]  1.850187070     0.17593677    0.1295365701
#>   [4,]  1.304323246     0.15858262    0.4207446866
#>   [5,] -0.874723730     0.21727485    0.3926297035
#>   [6,]  0.935263594     0.19131820    0.0660718385
#>   [7,] -1.618974303     0.66054105   -0.0475087654
#>   [8,] -0.407137856     0.18127151    0.4216791420
#>   [9,] -0.374847914     0.25849004    0.2133186381
#>  [10,]  2.320430627     0.26308279    0.2181710821
#>  [11,]  0.516772070     1.01894417    0.1598096145
#>  [12,] -0.012956857     0.55561105    0.0280487003
#>  [13,]  1.911729361     0.30869837    0.1021011195
#>  [14,]  1.380342769     0.54149404    0.4457760386
#>  [15,] -0.465823666     0.60693106    0.5115819985
#>  [16,] -0.917725371     0.08692383    0.5940939778
#>  [17,]  0.343778119     0.11464679    0.4606138475
#>  [18,]  1.096085442     0.83100784    0.2493607593
#>  [19,] -0.838024180     0.25004228    0.5985772878
#>  [20,]  0.260760744     0.85202464    0.1364515989
#>  [21,]  1.147039076     0.32741157    0.4703010439
#>  [22,] -0.337812603     0.32733984    0.0730577146
#>  [23,]  1.158785499     0.09121928    0.2789000446
#>  [24,]  0.943939074     0.13807866    0.1288069858
#>  [25,]  0.389557153     0.47399140    0.2151103364
#>  [26,] -0.221920094     0.11569690    0.4154763659
#>  [27,]  0.243833161     0.44022691   -0.0877383772
#>  [28,] -0.194305727     0.06594927    0.4328648052
#>  [29,]  0.050560395     0.33998764    0.3968991865
#>  [30,]  0.504116653     0.28393122    0.3885184463
#>  [31,] -0.466495329     0.20452726    0.2083780787
#>  [32,] -0.336119587     0.11268641    0.3872386156
#>  [33,] -1.473199254     0.45898831    0.3626567163
#>  [34,]  1.263300062     0.28458047    0.2838385193
#>  [35,] -0.661551837     0.10454037    0.2430528753
#>  [36,]  1.025245456     0.61443300    0.0831234977
#>  [37,] -0.006332139     0.64315844    0.4760160582
#>  [38,]  2.035414326     0.08348699    0.5765903069
#>  [39,] -0.056425189     0.18675273   -0.1869901825
#>  [40,]  0.244915812     0.21744078    0.3326776521
#>  [41,]  0.229479078     0.23933236    0.0837976142
#>  [42,]  2.659490239     0.29614039    0.4706761750
#>  [43,] -0.706523889     0.16088580    0.1607546135
#>  [44,]  2.058064166     0.49173388   -0.0800981564
#>  [45,]  2.708593174     0.10365659    0.4533998549
#>  [46,] -0.350857069     0.65231139    0.3633921012
#>  [47,] -0.624115140     0.17402576    0.0521503414
#>  [48,] -0.344443025     0.21124043    0.4126203976
#>  [49,] -0.131725718     0.18188287    0.0261414314
#>  [50,]  1.196768440     0.08411804    0.2574994312
#>  [51,]  1.288273635     0.09918986    0.4593314250
#>  [52,]  0.940967087     0.10056488    0.4511540750
#>  [53,] -0.992979326     0.16583061    0.1063280652
#>  [54,] -0.949958642     0.97096739    0.2701665122
#>  [55,] -0.340082681     0.43298043    0.7724130000
#>  [56,] -0.762688604     0.90061024   -0.0674990252
#>  [57,]  0.467165403     0.23282381   -0.1183288048
#>  [58,]  0.560536615     0.45085071    0.2186595248
#>  [59,] -2.541411884     0.33159758    0.0507815774
#>  [60,]  0.328237846     0.20942742   -0.0046527772
#>  [61,]  0.206365528     0.34979148    0.3047681695
#>  [62,] -1.558385809     0.54302945    0.3713414289
#>  [63,]  0.944225690     0.25237875    0.2939780092
#>  [64,] -0.018740039     0.32713495    0.0424934758
#>  [65,] -1.188065021     0.26078926    0.1577634993
#>  [66,] -1.234957337     0.15139252    0.5685482279
#>  [67,]  0.481623140     0.53187669    0.0992224288
#>  [68,] -0.025776420     0.70282257    0.4297193970
#>  [69,] -0.279370821     0.38599859    0.4389243546
#>  [70,]  1.283965437     0.21587652    0.3910102659
#>  [71,]  1.493404631     0.20957815    0.4885867070
#>  [72,]  1.030540692     0.39854834    0.4129542346
#>  [73,] -0.526443702     0.40524714    0.3693271994
#>  [74,]  1.502129510     0.36829758    0.2678031288
#>  [75,]  0.433983707     0.30007470    0.2592906722
#>  [76,]  1.553167689     0.25761469    0.2609789420
#>  [77,]  0.740932661     0.23801093   -0.1216580212
#>  [78,] -1.206090945     0.49341565    0.1705506310
#>  [79,]  1.739605040     0.32835911    0.1153572188
#>  [80,] -0.843340289     0.25336305    0.0372462188
#>  [81,] -0.266920455     0.34016808    0.4028420114
#>  [82,]  0.875261062     0.39796264   -0.0183214958
#>  [83,] -0.566118235     0.13029972    0.3479066410
#>  [84,] -1.309182055     0.62160630    0.1391500792
#>  [85,] -1.216300926     0.30513556    0.0774087868
#>  [86,]  1.079101130     0.29374681    0.0203513705
#>  [87,] -1.573454239     0.24004057   -0.2219770616
#>  [88,]  0.709278416     0.47886627    0.2644930373
#>  [89,]  0.191093125     0.12816855    0.3124566809
#>  [90,] -1.485251521     0.13902568    0.2953966571
#>  [91,] -0.197446947     0.32936233    0.1638577485
#>  [92,] -1.016910479     0.29641503    0.1754286008
#>  [93,] -0.166708574     0.32735137    0.4211440062
#>  [94,] -2.007462975     0.32639559    0.5914141272
#>  [95,]  0.437975853     0.17865417    0.4791148479
#>  [96,] -0.220146560     0.42885748    0.2343554515
#>  [97,] -1.760438211     0.11499264    0.1231621178
#>  [98,] -0.830714311     0.16127097    0.2280845239
#>  [99,] -0.805157375     0.26099301    0.2361071717
#> [100,]  0.814628198     0.61627123    0.2772950370
#> [101,]  1.808244243     0.94521813    0.1920787375
#> [102,]  0.265977096     0.41389651    0.3954187455
#> [103,]  0.238078860     0.47102809   -0.0043756414
#> [104,] -0.109212377     0.31525985    0.2809932258
#> [105,] -2.468670882     0.30755643    0.1553117893
#> [106,] -2.258461450     0.32942957    0.3134840725
#> [107,]  0.065622370     0.27738270    0.0927136860
#> [108,] -1.700064596     0.85045821   -0.0388615705
#> [109,]  0.535572995     0.16481716   -0.0418476743
#> [110,]  0.018382345     0.15161995    0.5873909353
#> [111,] -0.100820733     0.50793617    0.0514284167
#> [112,] -0.304701054     0.46349419    0.1038016908
#> [113,] -1.121048808     0.33370988    0.2777916892
#> [114,] -0.800019078     0.28444305    0.4755294711
#> [115,]  0.050685321     0.14918007   -0.1187076755
#> [116,] -0.013176196     0.31116339    0.1343710269
#> [117,]  0.048614294     0.83491520    0.3107468673
#> [118,]  0.093018640     0.21899375    0.4217692264
#> [119,]  0.363841273     0.10221860    0.1647859358
#> [120,]  0.090904518     0.18720123    0.3583960907
#> [121,] -0.665689011     0.74181584    0.2198926737
#> [122,] -0.334410295     0.79577348    0.0747941851
#> [123,] -0.483546949     0.55627790    0.3163148672
#> [124,]  0.396955952     0.43214717    0.5826815543
#> [125,]  0.153503508     0.97890179   -0.2552668303
#> [126,]  0.825235782     0.43515069    0.5737009637
#> [127,]  0.174305551     0.49527533    0.0697700849
#> [128,]  1.278509239     0.21335231    0.6218335703
#> [129,] -0.955491250     0.11760519    0.3501142821
#> [130,] -1.347129162     0.48855481    0.2975970785
#> [131,] -0.415899295     0.14129303    0.3895355664
#> [132,] -0.275354753     0.38200607    0.0487995461
#> [133,]  1.838398353     0.20378107    0.5169652215
#> [134,]  2.354094964     0.24579382    0.0364474512
#> [135,]  1.117161950     1.26797186   -0.1882516126
#> [136,]  0.467882055     0.36749698    0.2428660318
#> [137,] -0.256247573     0.35133873    0.1709473055
#> [138,] -1.688512179     0.74450568   -0.0813722154
#> [139,] -0.428398492     0.20753069    0.3173404981
#> [140,] -0.288704301     0.51560502    0.0163573127
#> [141,]  1.448202908     0.46503161    0.0690198802
#> [142,] -1.677240540     0.07477620    0.5200518214
#> [143,] -0.451793004     0.14486443    0.3213599847
#> [144,] -1.444970737     0.91692847   -0.0244975494
#> [145,]  0.477691573     0.23157399    0.3172871838
#> [146,] -0.763011353     0.27204812    0.2623930175
#> [147,]  0.906557433     0.32041019    0.1741524714
#> [148,] -0.615994409     0.26766176    0.0139499784
#> [149,]  1.274674659     0.40165758    0.3752684396
#> [150,] -1.165726926     0.08702516    0.3706295659
#> [151,] -1.505994869     0.25779498    0.3022654645
#> [152,] -0.685799467     0.56283992    0.0464598510
#> [153,]  1.236918150     0.19483382    0.4495316379
#> [154,] -0.989092935     0.23392204    0.2116315546
#> [155,] -1.018264559     0.10380712    0.4795010823
#> [156,] -0.996436698     0.16423086    0.5185217803
#> [157,]  0.307301638     0.21613308    0.1759727369
#> [158,]  0.144247590     0.92664149    0.1551612135
#> [159,]  1.989106491     0.71566532   -0.0651395264
#> [160,]  0.370850867     0.34040859    0.1548510906
#> [161,] -0.094570186     0.38083238   -0.1484786716
#> [162,]  0.933837803     0.31315801    0.0680883887
#> [163,]  0.757152859     0.12905207    0.4520284080
#> [164,]  1.326401667     0.20206845    0.0470416606
#> [165,] -0.698195116     0.16243874    0.4510373974
#> [166,] -0.443526019     0.36439709    0.0105692364
#> [167,]  2.161410598     0.25946480    0.1120899073
#> [168,] -0.228370246     0.13953008    0.4606885532
#> [169,]  0.102177882     0.14573070    0.5822364443
#> [170,] -0.097120672     0.30478192    0.2847366482
#> [171,] -0.080880278     0.25755196    0.4733525540
#> [172,] -1.193697148     0.15161988    0.5113272259
#> [173,]  0.266687636     0.11122932    0.5741190132
#> [174,]  0.643266603     0.32294927    0.2831608406
#> [175,] -2.936323105     0.10710257    0.2913371588
#> [176,] -0.620506650     0.27517144    0.4337632945
#> [177,]  0.317398672     0.47944965    0.2140145471
#> [178,]  1.716193562     0.31742470   -0.1032704580
#> [179,]  0.462342126     0.24184895    0.1872039794
#> [180,]  1.086460442     1.05479084    0.1264387751
#> [181,]  2.067504801     0.51294223    0.2834235342
#> [182,]  0.560626142     0.23171221   -0.0967745974
#> [183,] -0.379739606     0.37378518    0.1148092715
#> [184,] -0.384574996     0.22381765    0.1966022886
#> [185,] -0.006015274     0.28093050    0.5791314524
#> [186,]  1.564220128     0.77545053    0.1070667045
#> [187,] -0.064003478     0.10680630    0.5413719368
#> [188,]  1.064027843     0.38886471   -0.0757080732
#> [189,]  0.100593983     0.15963841    0.4479941333
#> [190,]  0.215671266     0.31144712    0.2179995573
#> [191,]  0.744727243     0.31556041    0.4930259157
#> [192,] -1.693722142     0.59501118    0.1751292875
#> [193,] -1.044569424     0.50598575   -0.0277084379
#> [194,] -0.602128833     0.18020653    0.3037218437
#> [195,] -1.806764156     0.16503690    0.0792208944
#> [196,] -0.599580655     0.91902963   -0.0415523745
#> [197,]  0.079542363     0.18095977    0.2906856479
#> [198,]  0.978572646     0.52959415    0.1717232416
#> [199,]  2.112278717     0.31095966    0.2957182551
#> [200,] -0.013049494     0.61829662    0.3279821082
#> [201,]  0.168396689     0.85140648    0.3115108674
#> [202,] -1.101812419     0.13136771    0.3396836315
#> [203,]  0.142661152     0.22516520    0.2549184930
#> [204,] -0.297215811     0.53042183   -0.1226576495
#> [205,] -1.264427131     0.12911884    0.3491672801
#> [206,] -0.424035817     0.24096403    0.5153671483
#> [207,]  0.757815288     0.21652347    0.5590228153
#> [208,]  1.530146721     0.34195792    0.1155543875
#> [209,]  0.678389930     0.64967031    0.3570491310
#> [210,] -0.212122387     0.30570890    0.5520041161
#> [211,] -0.420777826     0.12173316    0.1091330521
#> [212,]  0.657335544     0.22289089    0.4189747761
#> [213,]  0.783214489     0.77248861    0.1720524844
#> [214,]  1.104453137     0.14151129    0.2807117436
#> [215,]  0.148386152     0.28201924    0.1338947695
#> [216,] -0.402517618     0.21220050    0.3803759776
#> [217,] -2.563427729     0.45048855    0.4872359944
#> [218,]  0.155170942     0.42719214    0.0588860373
#> [219,]  0.279311708     0.26860627    0.4482023121
#> [220,]  0.087219568     0.78569202    0.1272424658
#> [221,] -1.034870973     0.45185527    0.2639852821
#> [222,]  0.615267761     0.48901193    0.4294727371
#> [223,]  1.610752793     0.26393758    0.2889435071
#> [224,] -1.593381256     0.19008861    0.0093613147
#> [225,] -1.329047418     0.51190021   -0.0376918617
#> [226,]  0.134931877     0.96643620    0.0918807525
#> [227,]  0.952499549     0.14092982    0.5916140557
#> [228,]  0.699501561     0.18563172    0.0939546360
#> [229,] -1.054306770     0.80484177    0.2278411067
#> [230,] -0.093152148     0.09600105    0.6325098537
#> [231,]  0.028826115     0.18027207    0.2808690455
#> [232,]  0.497702351     0.44097682    0.1932495921
#> [233,] -0.706357538     0.45557840    0.0557943748
#> [234,]  0.216089532     0.57938390    0.4297408551
#> [235,] -0.640080910     0.42854662    0.2862077605
#> [236,] -0.702779098     0.26926171    0.0244246810
#> [237,] -1.425853553     0.29302202    0.1936062565
#> [238,] -0.664174826     0.42161883    0.0901498004
#> [239,]  1.115345323     0.65624363    0.3120875325
#> [240,]  0.157963655     0.14570040    0.1480086845
#> [241,] -0.596648761     0.31829292    0.4919485665
#> [242,] -1.071404241     0.33725795    0.5644078197
#> [243,] -0.561452411     0.11778436    0.4473726650
#> [244,] -0.501561822     0.50522134    0.0494162320
#> [245,]  0.270310915     0.26821490    0.1669880977
#> [246,] -0.635258614     0.53277042    0.2485918884
#> [247,] -0.850094476     0.32997148    0.1208978041
#> [248,]  0.924507287     0.53529528    0.2602358334
#> [249,] -2.214441614     0.85460168    0.1905891828
#> [250,] -0.957015165     1.01915075    0.3431531553
#> [251,]  1.491000359     0.26129104    0.3000291435
#> [252,] -0.732063791     0.45389852    0.2170746489
#> [253,] -0.614335657     0.36103685    0.0250287991
#> [254,] -0.682405193     0.68043241    0.4742822683
#> [255,]  0.142159425     0.15023013    0.1903223802
#> [256,]  0.141394114     0.77045710    0.1043887800
#> [257,] -0.462845451     0.14835289    0.4354929805
#> [258,]  1.130400790     0.68882003    0.1033702971
#> [259,]  1.281288899     0.13362943    0.3623531844
#> [260,] -1.052278545     0.23495099    0.2851984374
#> [261,]  1.141257302     0.51506101   -0.1794882341
#> [262,]  0.208467257     0.25480143    0.2824331425
#> [263,] -1.471218660     0.91917205    0.1214131185
#> [264,] -0.719967592     0.12429392    0.3631826608
#> [265,] -0.622292255     0.46110237    0.4290734267
#> [266,] -1.087749869     0.50347889    0.2044482483
#> [267,]  1.979022669     0.13694691    0.2889968525
#> [268,]  0.473082886     0.29033983    0.1953204210
#> [269,] -0.335608149     0.23188347    0.2706554423
#> [270,]  0.327439118     0.29437021    0.2557932638
#> [271,] -0.358984501     0.31557200    0.3948177825
#> [272,] -0.735419795     0.46850751    0.4950203012
#> [273,]  0.037030931     0.17656967    0.2183373937
#> [274,]  0.665639208     0.38147343    0.3505164804
#> [275,] -0.787200643     0.29495447    0.1226637287
#> [276,] -0.962200768     0.39723955    0.5737195522
#> [277,]  1.419612393     0.29387047    0.4648639877
#> [278,]  0.045003875     0.17100640    0.3370782648
#> [279,] -0.631089038     0.26880685    0.4894606450
#> [280,]  0.627009947     0.48447690    0.0987311708
#> [281,] -1.081141662     0.44101481    0.2287177964
#> [282,]  0.442298123     0.19273820    0.3132783756
#> [283,] -1.369990501     0.20067695    0.3921469219
#> [284,]  0.215914755     0.29031893    0.2721691828
#> [285,] -1.143606667     0.14904921    0.6040235813
#> [286,]  0.330846932     0.28832017    0.5493135590
#> [287,] -0.981206545     0.18706426   -0.2265064462
#> [288,]  1.610929152     1.10096642    0.1965903588
#> [289,]  0.465260643     0.40460013    0.2165054404
#> [290,]  0.357150081     0.51218232   -0.0510186314
#> [291,]  1.738595695     0.44562653    0.3648435460
#> [292,]  0.844580933     0.25018976   -0.0991958252
#> [293,]  0.774089026     0.29276454    0.1149294679
#> [294,]  0.863307441     0.69117770    0.3400958571
#> [295,] -1.078970276     0.76000047    0.2927783389
#> [296,] -0.383243410     0.29784760    0.0160771507
#> [297,]  2.067985242     0.24339749    0.3624557968
#> [298,] -2.217043624     0.52658718   -0.0007637109
#> [299,] -0.060984142     0.23536953    0.3849937044
#> [300,] -2.079992445     1.18243875   -0.2905247613
#> 
#> $acov_order
#> [1] 0
#> 
#> $acor_order
#> [1] 1
#> 
#> $N
#> [1] 300
#> 
#> $S
#> [1] 50
#> 
#> $R
#> [1] 1000
#>