The naive estimation of the moments

The `nemoment()` function enables to implement the naive estimation of the moments of the heterogeneous mean, the heterogeneous autocovariance, and the heterogeneous autocorrelation. The method is developed by Okui and Yanagi (2019). For more details, see the package vignette with `vignette("panelhetero")`.

Usage

nemoment(data, acov_order = 0, acor_order = 1, R = 1000)

Arguments

data

A matrix of panel data. Each row corresponds to individual time series.

acov_order

A non-negative integer of the order of autocovariance. Default is 0.

acor_order

A positive integer of the order of autocorrelation Default is 1.

R

A positive integer of the number of bootstrap repetitions. Default is 1000.

Value

A list that contains the following elements.

estimate

A vector of the parameter estimates

se

A vector of the standard errors

ci

A matrix of the 95 percent confidence intervals

quantity

A matrix of the estimated heterogeneous quantities

acov_order

The order of autocovariance

acor_order

The order of autocovariance

N

The number of cross-sectional units

S

The length of time series

R

The number of bootstrap repetitions

References

Okui, R. and Yanagi, T., 2019. Panel data analysis with heterogeneous dynamics. Journal of Econometrics, 212(2), pp.451-475.

Examples

data <- panelhetero::simulation(N = 300, S = 50)
panelhetero::nemoment(data = data)
#> $estimate
#>         E(mean)         E(acov)         E(acor)       var(mean)       var(acov) 
#>      0.00946270      0.35203151      0.23465957      1.06033774      0.04600905 
#>       var(acor) cor(mean, acov) cor(mean, acor) cor(acov, acor) 
#>      0.03634836     -0.05057665     -0.01313619     -0.29666390 
#> 
#> $se
#>         E(mean)         E(acov)         E(acor)       var(mean)       var(acov) 
#>     0.060409890     0.012151567     0.011028968     0.085217561     0.005547262 
#>       var(acor) cor(mean, acov) cor(mean, acor) cor(acov, acor) 
#>     0.002595378     0.057469443     0.056890809     0.045803600 
#> 
#> $ci
#>                 95% CI lower 95% CI upper
#> E(mean)          -0.10808892   0.12539402
#> E(acov)           0.32905660   0.37535883
#> E(acor)           0.21307802   0.25625469
#> var(mean)         0.89701477   1.23745968
#> var(acov)         0.03567220   0.05726507
#> var(acor)         0.03141799   0.04163332
#> cor(mean, acov)  -0.16770168   0.06279481
#> cor(mean, acor)  -0.12348267   0.09585648
#> cor(acov, acor)  -0.38474061  -0.20878939
#> 
#> $quantity
#>                mean autocovariance autocorrelation
#>   [1,]  0.663666842     0.29839369    0.3177999794
#>   [2,]  0.288922106     0.29941846    0.0371763394
#>   [3,] -0.486215203     0.11776498    0.5410312670
#>   [4,] -0.183472984     0.21430025    0.1429666295
#>   [5,] -1.613822612     0.10240334    0.0872303253
#>   [6,]  1.602381856     0.10434800    0.1223784456
#>   [7,] -1.510183143     1.21324129    0.2644782895
#>   [8,]  2.730683251     0.08116850    0.6201828313
#>   [9,] -2.151767564     0.35237095    0.1425450641
#>  [10,]  0.482899027     0.33213961    0.0018286394
#>  [11,] -2.920531287     0.14064366    0.5265130670
#>  [12,] -1.591802249     0.52909493    0.0668981748
#>  [13,] -2.819378638     0.36550244    0.2335938046
#>  [14,] -0.939949987     0.12622317    0.4850317961
#>  [15,] -0.970184437     0.29745197    0.0592558529
#>  [16,]  1.203835393     0.12899020    0.3224529671
#>  [17,]  1.061947324     0.35497508   -0.0029406419
#>  [18,] -0.389156153     0.27670513   -0.0824617180
#>  [19,] -0.944900867     0.31937133    0.3644361572
#>  [20,] -0.352042075     0.41351007    0.3468732848
#>  [21,]  0.513206972     0.56484371    0.4368072587
#>  [22,]  0.892732885     0.83926970    0.0375604831
#>  [23,]  0.877560075     0.26763074    0.2222414350
#>  [24,]  1.103281162     0.22939644    0.2150775710
#>  [25,] -1.125027369     0.22399775    0.3447247579
#>  [26,]  0.763388364     0.44796574    0.5272739979
#>  [27,]  1.265205464     0.49233074    0.4479484361
#>  [28,]  0.392143361     0.20481803    0.2375160111
#>  [29,]  0.080157155     1.06119942   -0.0003285979
#>  [30,]  1.356138307     0.23995201   -0.0356186432
#>  [31,] -0.986033706     0.20546253    0.2902222037
#>  [32,] -0.028968149     0.18362501   -0.0689371518
#>  [33,]  0.434507692     0.17703359    0.3257691680
#>  [34,] -0.012650740     0.73257339    0.1927255376
#>  [35,] -0.004006884     0.21866303    0.2788294409
#>  [36,]  0.270092044     0.35870184    0.1059736361
#>  [37,]  0.172228081     0.56437388    0.0200954088
#>  [38,] -1.071315797     0.69238175    0.2307368343
#>  [39,] -0.458489291     0.21715590    0.4693821990
#>  [40,] -1.808614147     0.34737536    0.3612188558
#>  [41,]  0.694455620     0.11972000    0.4916535374
#>  [42,] -0.470815073     0.22578639    0.5121485953
#>  [43,]  0.858411010     0.23280911    0.0511271476
#>  [44,] -0.928693905     0.37108137    0.4957908070
#>  [45,]  0.704370172     0.30201383    0.3776438335
#>  [46,] -0.221399879     0.22281770    0.1319467997
#>  [47,]  0.332496628     0.22202849    0.1981418844
#>  [48,]  1.748941305     0.35065826    0.4703544970
#>  [49,] -0.683869150     0.10782495    0.6379295499
#>  [50,] -0.370767474     0.32805270   -0.1994735054
#>  [51,]  1.319227328     0.57329329    0.0177714260
#>  [52,] -0.303268905     0.19686554    0.5297406249
#>  [53,] -0.310205094     0.32337123    0.2023333509
#>  [54,] -1.708270865     0.49249862    0.1233368739
#>  [55,] -0.442756595     0.30842919    0.0881744593
#>  [56,]  0.822609314     0.46504493    0.2835977731
#>  [57,] -0.362861410     0.23104772    0.1328435364
#>  [58,] -0.711695905     0.29831438    0.4884752812
#>  [59,]  1.571417988     0.07253951    0.5330882208
#>  [60,]  1.185533324     0.25502210    0.1995984205
#>  [61,] -0.944911955     0.16839252    0.1268269371
#>  [62,] -0.977750762     0.56031165   -0.0524287997
#>  [63,]  0.868142874     0.09817629    0.6450328018
#>  [64,] -1.446054648     0.14088909    0.1245077082
#>  [65,] -0.019220162     0.54237817    0.5404913730
#>  [66,] -0.510766295     0.16935829    0.1327744574
#>  [67,] -0.088963198     0.36367138    0.3170185364
#>  [68,] -1.149464326     0.53257558    0.2162091325
#>  [69,] -0.049312624     0.19004888   -0.0275886726
#>  [70,] -1.503906827     0.54470391   -0.0062355289
#>  [71,]  0.490440819     0.17117902    0.5440148179
#>  [72,]  1.074663197     0.05649078    0.3181176331
#>  [73,]  0.002485644     0.09814991    0.4606705728
#>  [74,]  0.316172009     0.09923003    0.4635092275
#>  [75,]  0.204930971     0.54986450    0.1507575002
#>  [76,]  0.176529842     0.33104813    0.5814745663
#>  [77,] -0.274695980     0.67585316   -0.0451045448
#>  [78,]  0.374456573     0.35965250    0.3049587585
#>  [79,]  0.098838010     0.22116058    0.3567106465
#>  [80,]  0.097673592     0.27213616    0.1414575681
#>  [81,]  0.280790516     0.21061795    0.1569941371
#>  [82,]  1.493375807     0.22426528    0.4837951102
#>  [83,] -0.854560482     0.80320634    0.0907245519
#>  [84,]  0.230489368     0.24239422    0.2688831100
#>  [85,]  0.935053386     0.15536987    0.3042701501
#>  [86,]  0.522781575     0.48554625    0.3861460998
#>  [87,]  1.628182439     0.12165560    0.5070072324
#>  [88,] -1.033179871     0.13085947    0.2903834714
#>  [89,]  0.212299493     0.06210654    0.2652539207
#>  [90,]  0.308466441     0.26746465    0.4531577995
#>  [91,]  0.234228815     0.16709785    0.4186989137
#>  [92,] -0.919894355     0.09316104    0.2267469087
#>  [93,]  0.130195726     0.19412414    0.4768952930
#>  [94,] -0.376360568     0.27506987    0.0380661568
#>  [95,] -1.268271889     0.38100995    0.3805060705
#>  [96,] -0.245268558     0.19294807    0.2478312718
#>  [97,] -0.796495038     0.34476246    0.2985620219
#>  [98,]  0.817536245     0.36674196    0.1543894512
#>  [99,] -0.797237970     0.43281678   -0.2499980154
#> [100,] -0.259167180     0.40541173    0.4000129935
#> [101,] -0.363944169     0.25932059    0.1402267212
#> [102,] -0.977914495     0.59438569    0.1659743979
#> [103,] -1.294132607     0.18420228    0.4062990365
#> [104,]  1.262432833     0.30680408    0.4324656629
#> [105,]  1.063025161     0.12725507    0.1613125577
#> [106,] -2.043102180     0.55880555    0.1321827407
#> [107,] -0.739128453     0.37270569    0.1266720352
#> [108,]  1.387625767     0.29449826    0.1034747526
#> [109,] -1.324673497     0.29964294    0.1180622549
#> [110,] -0.459982667     0.17646418    0.3752835362
#> [111,] -0.663742637     0.37804344    0.2007615493
#> [112,] -0.432671987     0.43157882   -0.0171176211
#> [113,]  0.041019453     0.87517993    0.1161825493
#> [114,]  0.671229110     0.34777098   -0.1226920641
#> [115,] -1.121903425     0.31027554    0.1022793989
#> [116,]  1.091912095     0.35686534    0.2305954573
#> [117,]  0.002045994     0.32623201    0.0204338972
#> [118,] -0.169812701     0.22548958    0.1095541531
#> [119,] -0.616187331     0.75957707    0.0192440370
#> [120,]  0.335581740     0.10702651    0.5482262386
#> [121,]  1.153931180     1.08052293    0.3044834447
#> [122,]  1.312975130     0.52500693   -0.0271936199
#> [123,]  1.343572784     0.19015214    0.3483801246
#> [124,]  1.912138781     0.21529186    0.1484718409
#> [125,]  1.816958540     0.17284405    0.0116992348
#> [126,] -1.995540451     0.10037398    0.5118498942
#> [127,] -0.784949245     0.47462087    0.0892597979
#> [128,]  0.067382003     0.41806266    0.0750004098
#> [129,]  1.167844569     0.08925308    0.3229658089
#> [130,]  1.777618186     0.40372444    0.1639816343
#> [131,]  0.026366275     0.19604109    0.6769942875
#> [132,]  0.656536644     0.21644889    0.1666543919
#> [133,] -1.367620841     0.56196513    0.4977227929
#> [134,]  0.873939952     0.26079051    0.0660735563
#> [135,] -0.348475764     0.34288014    0.0202693769
#> [136,] -0.988434454     0.40621559    0.4620832297
#> [137,] -1.319046883     0.15007478    0.1295272429
#> [138,]  0.074184972     0.19280784    0.5449168220
#> [139,] -0.529745525     0.55191224    0.2850728472
#> [140,] -0.057586818     0.36545333    0.0756250656
#> [141,] -0.157204871     0.13452670   -0.0376161119
#> [142,] -1.089759435     0.30145638    0.0606400592
#> [143,] -0.890024413     0.16816494    0.2628999164
#> [144,]  0.230868074     0.24823590    0.0732988961
#> [145,] -1.245609143     0.14457359    0.2992609476
#> [146,] -1.151764917     0.37928993    0.4643541145
#> [147,]  0.012681703     0.91278987   -0.0735415164
#> [148,]  0.568367123     0.17840085    0.1799478392
#> [149,] -0.169687351     0.49759817    0.2354169572
#> [150,] -0.580490867     0.80157861    0.2295084675
#> [151,] -2.103807548     0.48516539    0.3115023971
#> [152,]  2.275242906     0.41714778    0.1308048376
#> [153,] -0.240847894     0.40788001    0.3377224461
#> [154,] -0.430012622     0.55431108    0.0665838460
#> [155,]  1.237941494     0.43120314   -0.0515958682
#> [156,]  0.681602041     0.80007667    0.3431031923
#> [157,]  0.207907528     0.34138751    0.0326286930
#> [158,]  0.230470078     0.22648152   -0.0228486725
#> [159,]  0.475882396     0.57870234    0.0179172974
#> [160,] -2.227704978     0.45805760    0.4420030244
#> [161,] -0.542406323     0.31929875    0.0788496372
#> [162,]  0.215461917     0.36549483    0.1210929233
#> [163,] -0.834036120     0.24120410    0.1886868302
#> [164,] -1.749547567     0.27413435    0.1438837021
#> [165,] -0.391098036     0.28531176    0.1362178269
#> [166,]  0.321166993     0.62298684   -0.0666149441
#> [167,] -0.005634056     0.57832828    0.0217744101
#> [168,]  0.466547308     0.21181654    0.2819495333
#> [169,]  0.438104785     0.95739587    0.0678560942
#> [170,] -0.138623058     0.33811366   -0.0460138616
#> [171,] -0.258550470     0.33216122    0.2855055518
#> [172,]  1.112894428     0.53528728    0.0095227797
#> [173,]  1.649920640     0.30015605    0.1653451632
#> [174,]  1.136974302     0.21072227    0.0730644403
#> [175,]  0.644054637     0.30979599    0.2752778544
#> [176,] -1.951154279     0.35214211    0.2700632180
#> [177,]  1.250953718     0.52766957    0.3523273168
#> [178,]  0.269069868     0.22096598    0.3203904380
#> [179,]  2.080956233     0.52106314    0.0261364026
#> [180,]  0.581601178     0.22733556    0.3430728944
#> [181,] -0.222630431     0.46939606    0.1545721096
#> [182,]  0.193554024     0.25832402    0.5020828482
#> [183,] -1.470467317     0.46996972    0.0791663049
#> [184,]  1.151795198     0.34283513    0.0219624072
#> [185,] -0.887945974     0.17662531    0.1245815618
#> [186,]  1.067208087     0.45098142    0.3786993395
#> [187,] -0.112642824     0.28057182    0.4112897465
#> [188,]  1.290732103     0.29127874    0.4924664611
#> [189,]  1.199700676     0.27621940    0.0754891477
#> [190,]  1.532708050     0.31678295    0.1838750231
#> [191,]  0.787980577     0.07683586    0.1760092309
#> [192,]  0.265827077     0.13971793    0.1227109686
#> [193,] -0.046338739     0.19074021    0.4628739099
#> [194,]  1.282090484     0.24224917    0.2591624112
#> [195,]  0.701690715     0.20456305    0.2562585118
#> [196,]  0.094078225     0.11947828    0.5066573891
#> [197,]  0.840359833     0.19909893    0.4118412148
#> [198,]  1.067389207     0.31102110    0.0720747783
#> [199,]  0.056565709     0.48886514    0.0742494999
#> [200,]  1.551856600     0.51623531   -0.0418481928
#> [201,]  0.459300652     0.09151258    0.7477174504
#> [202,]  0.365727341     0.48421015    0.1799989383
#> [203,] -0.924938016     0.27334760    0.2743424487
#> [204,]  1.835079015     0.17111259    0.0742028107
#> [205,] -0.462446066     0.57344166    0.1167511292
#> [206,] -1.964559814     0.43625327    0.2016636494
#> [207,] -0.412750561     0.06614148    0.4465054708
#> [208,] -0.212428028     0.43848318    0.0092425967
#> [209,] -1.062878627     0.28250628    0.1823450679
#> [210,]  0.647572448     0.29378066    0.2292749013
#> [211,]  1.640188035     0.92572923    0.1294107662
#> [212,]  0.104473474     0.24146654    0.0121708753
#> [213,] -1.179755216     0.53590071    0.3074657057
#> [214,]  0.900508040     0.71763416    0.1266723349
#> [215,] -2.344703011     1.05858289    0.0006632426
#> [216,] -0.958559139     0.31040593    0.4458956720
#> [217,] -0.628891756     0.31740632    0.0502827979
#> [218,] -2.016651463     0.29234293    0.4910319393
#> [219,] -1.069055962     0.31004752   -0.0442929649
#> [220,] -0.895600798     0.33053896    0.2036157878
#> [221,] -1.092681578     0.72889628   -0.1159645312
#> [222,]  0.896336938     0.41718885   -0.0727802304
#> [223,] -0.077366234     0.08474412    0.3489796199
#> [224,]  1.105168142     0.38958959    0.3359330647
#> [225,]  0.807488989     0.11178176    0.3920206450
#> [226,]  0.989767992     0.44892537   -0.2346905376
#> [227,] -2.741722087     0.37254476    0.1169079002
#> [228,] -0.932493036     0.13056738    0.3304057685
#> [229,]  0.605985702     0.34300822    0.3129588360
#> [230,]  0.949881689     0.37726594    0.0903164724
#> [231,]  0.519324849     0.39554234    0.2080330713
#> [232,]  0.633415414     0.21144084    0.1706570425
#> [233,] -0.561153141     0.36610775    0.0957236627
#> [234,]  0.877076204     0.14690548    0.1620300898
#> [235,] -0.958188826     0.18544734    0.2554129426
#> [236,]  1.481288022     0.60522658    0.1710330980
#> [237,] -1.126540821     0.36714802    0.2544274228
#> [238,]  1.240864519     0.40629703   -0.1292704555
#> [239,] -0.587799982     0.67043536    0.2666241035
#> [240,]  0.882610492     0.50426164    0.3010587418
#> [241,] -0.700583981     0.23429543    0.2568012780
#> [242,] -0.902616042     0.26547033    0.2021542352
#> [243,] -0.331061015     0.52917399    0.2139749261
#> [244,]  0.588622407     1.25488843    0.2772096067
#> [245,]  1.131890058     0.44361602    0.2696973494
#> [246,] -1.144521798     0.12758030    0.4227083729
#> [247,] -0.938504081     0.93068645    0.2830073300
#> [248,] -0.594223047     0.57607351    0.1466689714
#> [249,]  0.891485618     1.05034656    0.2833808054
#> [250,] -0.172642020     0.47294557    0.2281038971
#> [251,] -0.207773230     0.50798528    0.4032690391
#> [252,] -0.938078431     0.11168291    0.3027385510
#> [253,]  0.426901170     0.46752969    0.2239372440
#> [254,]  2.194181965     0.23514521    0.2702441705
#> [255,]  1.497082274     0.36699787    0.4438684108
#> [256,] -0.381264043     0.33602374    0.1295442490
#> [257,] -0.170406231     0.10608189    0.5148832053
#> [258,] -0.887433791     0.27586464    0.0288833977
#> [259,] -1.123535940     0.08553217    0.3028803702
#> [260,] -0.968885106     0.52079728    0.4349568712
#> [261,]  0.276769457     0.34692520    0.1437817683
#> [262,] -0.109903528     0.30444333    0.2336703207
#> [263,] -0.813851769     0.16609835    0.5448045758
#> [264,] -0.162562388     0.28705149    0.6026439658
#> [265,] -0.665741085     0.38203943    0.2988123762
#> [266,] -0.573303192     0.60612304    0.4698789789
#> [267,]  0.973371706     0.24653770    0.1933399144
#> [268,] -1.474458832     0.20803683    0.3820672727
#> [269,] -1.196664717     0.04382946    0.7342559563
#> [270,] -0.423405675     0.58417344    0.3292146733
#> [271,]  0.126366711     0.46281020    0.0781867863
#> [272,] -0.212604611     0.40874097    0.1234841508
#> [273,]  1.402251254     0.32774519    0.0389661261
#> [274,]  2.061065969     0.36930645    0.6778644562
#> [275,]  0.713077999     0.22685743    0.1113596516
#> [276,] -0.014851960     0.71224642    0.2164405925
#> [277,]  1.529686925     0.21606175   -0.0520237529
#> [278,] -0.239377546     0.25608172    0.4157797036
#> [279,]  0.705737157     0.45898812    0.2138889384
#> [280,]  0.419669318     0.16152406    0.3973076435
#> [281,] -0.509834151     0.48967865    0.3859929652
#> [282,] -1.109194285     0.54218782   -0.0377380621
#> [283,] -0.472691583     0.32101180    0.1665094478
#> [284,]  0.270425189     0.20735205    0.0938256426
#> [285,] -0.574045161     0.19482992    0.1728136243
#> [286,]  2.998957710     0.29391621    0.3547055665
#> [287,]  0.441563351     0.23825450    0.4527801170
#> [288,]  2.498741692     0.33352477    0.4968956677
#> [289,]  0.084692172     0.20341133    0.3171503454
#> [290,]  0.241379913     0.35755070    0.1305531545
#> [291,] -0.148037882     0.12101790    0.4102437403
#> [292,]  0.177185563     0.24580528    0.5310871369
#> [293,]  0.490707037     0.09649358    0.4734615501
#> [294,]  0.407131577     0.95353137    0.1162555442
#> [295,]  1.014192212     0.20761196    0.2712039540
#> [296,] -0.717837061     0.27906460    0.7110580129
#> [297,] -0.408441076     0.11223047    0.0673402079
#> [298,]  1.027680407     0.20357109    0.6298292687
#> [299,] -2.043372011     0.30674191    0.2245869562
#> [300,]  0.643291655     0.84850538   -0.0201235601
#> 
#> $acov_order
#> [1] 0
#> 
#> $acor_order
#> [1] 1
#> 
#> $N
#> [1] 300
#> 
#> $S
#> [1] 50
#> 
#> $R
#> [1] 1000
#>