The TOJ bias-corrected estimation of the moments

The `tojmoment()` function enables to implement the TOJ bias-corrected estimation of the moments of the heterogeneous mean, the heterogeneous autocovariance, and the heterogeneous autocorrelation. The method is developed by Okui and Yanagi (2019). For more details, see the package vignette with `vignette("panelhetero")`.

Usage

tojmoment(data, acov_order = 0, acor_order = 1, R = 1000)

Arguments

data

A matrix of panel data. Each row corresponds to individual time series.

acov_order

A non-negative integer of the order of autocovariance. Default is 0.

acor_order

A positive integer of the order of autocorrelation Default is 1.

R

A positive integer of the number of bootstrap repetitions. Default is 1000.

Value

A list that contains the following elements.

estimate

A vector of the parameter estimates

se

A vector of the standard errors

ci

A matrix of the 95 percent confidence intervals

quantity

A matrix of the estimated heterogeneous quantities

acov_order

The order of autocovariance

acor_order

The order of autocovariance

N

The number of cross-sectional units

S

The length of time series

R

The number of bootstrap repetitions

References

Okui, R. and Yanagi, T., 2019. Panel data analysis with heterogeneous dynamics. Journal of Econometrics, 212(2), pp.451-475.

Examples

data <- panelhetero::simulation(N = 300, S = 50)
panelhetero::tojmoment(data = data)
#> $estimate
#>         E(mean)         E(acov)         E(acor)       var(mean)       var(acov) 
#>     -0.07822102      0.37112848      0.27363736      1.00249435      0.02984209 
#>       var(acor) cor(mean, acov) cor(mean, acor) cor(acov, acor) 
#>      0.01629745     -0.03641749      0.02002358     -0.55039317 
#> 
#> $se
#>         E(mean)         E(acov)         E(acor)       var(mean)       var(acov) 
#>     0.058169984     0.012250640     0.015765843     0.077739460     0.005384604 
#>       var(acor) cor(mean, acov) cor(mean, acor) cor(acov, acor) 
#>     0.007264171     0.072466783     0.087371706     0.113858253 
#> 
#> $ci
#>                 95% CI lower 95% CI upper
#> E(mean)         -0.187101241   0.03674193
#> E(acov)          0.347405176   0.39550639
#> E(acor)          0.243082003   0.30408058
#> var(mean)        0.849718581   1.15646572
#> var(acov)        0.019796343   0.04169034
#> var(acor)        0.002247451   0.03078345
#> cor(mean, acov) -0.173242169   0.10459012
#> cor(mean, acor) -0.147890287   0.19132353
#> cor(acov, acor) -0.753385115  -0.31461049
#> 
#> $quantity
#>                mean autocovariance autocorrelation
#>   [1,]  1.113991017     0.16684078     0.367664196
#>   [2,] -1.563463978     0.07313756     0.642720409
#>   [3,] -0.103503523     0.62912998     0.230518240
#>   [4,]  1.204735844     0.31719145     0.444943384
#>   [5,]  0.870703974     0.10129646     0.288148201
#>   [6,] -0.262315468     0.13796982     0.155340291
#>   [7,]  1.431711598     0.40716802     0.438604294
#>   [8,]  0.143756118     0.63079176     0.279935773
#>   [9,]  1.548656260     0.25730644     0.067994610
#>  [10,] -2.397682323     0.32286306     0.254834734
#>  [11,]  1.618830402     0.40560057     0.193018436
#>  [12,]  0.254822601     0.15585557     0.335845343
#>  [13,]  1.030728884     0.10866534     0.533737516
#>  [14,]  1.086069443     0.23540491    -0.020014452
#>  [15,] -0.443514080     0.26612402     0.165938045
#>  [16,]  0.301495046     0.18738218     0.424106839
#>  [17,]  0.620681008     0.17622937     0.339302865
#>  [18,] -0.024811063     0.22570480    -0.025107723
#>  [19,] -0.521614960     0.22178722     0.247484407
#>  [20,] -2.134613904     0.31798468     0.469686450
#>  [21,] -0.309533975     0.44356748     0.066902901
#>  [22,]  1.068218430     0.22211215     0.116447535
#>  [23,]  0.821520413     0.14952879     0.382772169
#>  [24,] -1.082880404     0.28029549     0.592477874
#>  [25,] -0.059226283     0.59316036     0.214856316
#>  [26,] -1.713491599     0.50035168     0.312818113
#>  [27,] -0.667749681     0.40872966    -0.070532970
#>  [28,]  1.549537920     0.33781703     0.281775955
#>  [29,] -1.155501256     0.15282424     0.497356045
#>  [30,]  1.173010276     0.66101548     0.163282545
#>  [31,]  0.563515254     0.21447021     0.155653418
#>  [32,]  0.008849323     0.34623464     0.485344662
#>  [33,] -0.407500133     0.13116965     0.406636023
#>  [34,]  0.739744795     0.36603556     0.093342349
#>  [35,]  0.221499658     0.19793254     0.096273983
#>  [36,] -0.840760585     0.25517609     0.182388453
#>  [37,]  1.241490834     0.80727203    -0.076469281
#>  [38,]  0.604631344     0.30210603     0.545065835
#>  [39,]  1.286359532     0.41263469     0.266220380
#>  [40,] -1.149542679     0.44022711     0.199330386
#>  [41,]  0.921847788     0.39243357    -0.055049367
#>  [42,] -0.749595457     0.18684220     0.493874866
#>  [43,] -0.453460635     0.27645563    -0.050287700
#>  [44,] -0.433724637     0.41138613     0.221477720
#>  [45,] -1.175838275     0.52236225     0.231405407
#>  [46,] -0.292631395     0.34128030     0.428738267
#>  [47,] -0.636082390     0.54793295     0.070178741
#>  [48,]  0.346544426     0.13612790     0.445864321
#>  [49,] -0.672494517     0.12000313     0.268594271
#>  [50,] -2.489502370     0.33104828     0.442888657
#>  [51,] -1.146697357     0.16180905     0.316232744
#>  [52,] -1.078644187     0.15813770     0.090765486
#>  [53,] -1.505519931     0.46657044    -0.037599712
#>  [54,] -0.034053782     0.28008268     0.092916261
#>  [55,]  0.645482835     0.43451227     0.088014322
#>  [56,]  1.194486458     0.82218376     0.507758862
#>  [57,]  0.023029400     0.17100702     0.329077578
#>  [58,]  1.901686234     0.11415357     0.343198850
#>  [59,]  1.243091786     0.22498641     0.208113578
#>  [60,]  0.633846386     0.55477510     0.299630719
#>  [61,] -0.414451607     0.29079176     0.069953777
#>  [62,]  1.324789425     0.21317912     0.107471781
#>  [63,] -0.377645564     0.23325342     0.065650184
#>  [64,]  0.545856321     0.24750642     0.149431805
#>  [65,]  0.310213679     0.34591841     0.231546224
#>  [66,] -0.892979232     0.30432264     0.383835717
#>  [67,] -0.167431828     0.60853596     0.174314766
#>  [68,] -1.065373246     0.69185459     0.128136154
#>  [69,]  1.616998577     0.11869091     0.502512818
#>  [70,] -2.256726452     0.40891999     0.106513480
#>  [71,] -2.346062769     0.14921468     0.276055315
#>  [72,] -0.482819178     0.39435322     0.173714349
#>  [73,]  0.139222422     0.57205068     0.345634652
#>  [74,] -0.783785384     0.14435040     0.346931069
#>  [75,]  1.411890792     0.08563841     0.710660972
#>  [76,] -1.490305714     1.07805339    -0.127917331
#>  [77,]  0.554470908     0.43866129     0.285707689
#>  [78,] -0.014591138     0.41295281     0.440510657
#>  [79,] -1.875787006     0.38983928     0.044318904
#>  [80,] -0.440014650     0.40357041     0.221236425
#>  [81,]  0.159468007     0.20542814     0.120274916
#>  [82,]  0.199287021     0.14110105     0.519031004
#>  [83,] -0.975002757     0.42781589    -0.008648031
#>  [84,] -2.797556805     0.38040434     0.254338305
#>  [85,]  0.921011268     0.50898991    -0.085174387
#>  [86,] -1.265291683     0.48971407     0.152642212
#>  [87,]  0.324804004     0.22063810     0.101243541
#>  [88,] -0.489220039     0.17510538     0.019406125
#>  [89,]  0.336216654     0.14980942     0.510384805
#>  [90,] -0.510776422     0.36318890    -0.229405765
#>  [91,] -0.260845275     0.41868429     0.239577036
#>  [92,] -1.555220012     0.35437297     0.259758446
#>  [93,] -1.443147738     0.14508042     0.495933346
#>  [94,]  0.715959755     0.39977872    -0.074011607
#>  [95,]  1.072501918     0.12910132     0.506763162
#>  [96,] -0.474613825     0.30939049     0.117955644
#>  [97,] -0.409200228     0.05450526     0.535021115
#>  [98,]  0.547834343     0.14060506     0.208057240
#>  [99,] -0.689735609     0.58368953     0.152418977
#> [100,] -1.253285083     0.22349517    -0.057886607
#> [101,] -0.594725402     0.19240265     0.240856726
#> [102,] -0.391598401     0.15282268     0.430708041
#> [103,] -0.280381188     0.39971758     0.353891988
#> [104,]  0.939605964     0.54204221     0.243808189
#> [105,] -0.903064763     0.07367848     0.151885008
#> [106,]  1.123677043     0.25695129     0.254069806
#> [107,]  0.431514401     0.22876757     0.280743693
#> [108,] -1.208613188     0.43515871     0.156240458
#> [109,]  0.353552376     0.18177673     0.117894006
#> [110,]  0.377672489     0.40440222     0.054829300
#> [111,]  1.296172098     0.41411584     0.112404127
#> [112,] -0.074485483     0.32454610    -0.067129597
#> [113,] -0.332849649     0.12881580     0.445256454
#> [114,] -0.774595289     0.33645002     0.430326002
#> [115,] -0.140240616     0.18038584     0.055083156
#> [116,]  0.123324705     0.25817114     0.075520373
#> [117,]  1.826654960     0.35531745     0.273130618
#> [118,] -1.321943683     0.13508346     0.325223210
#> [119,] -0.799689891     0.35788680     0.154566010
#> [120,]  1.190500192     0.21838767     0.187670971
#> [121,] -0.951133755     0.61672730     0.174264703
#> [122,] -0.641871096     0.24142745     0.219690322
#> [123,] -2.023814368     0.50899085     0.257270860
#> [124,] -0.123562196     0.38939957     0.036156546
#> [125,]  0.371885879     0.36091038    -0.068830157
#> [126,] -0.169382666     0.23219359     0.512580708
#> [127,] -1.002198237     0.70314704     0.244409193
#> [128,] -0.087354573     0.41137580     0.093324594
#> [129,] -0.810714971     0.32984440     0.171215141
#> [130,]  1.064408638     0.48861760     0.006829402
#> [131,]  0.839843075     0.16567310     0.728072940
#> [132,]  0.992016980     0.18771880     0.137038127
#> [133,] -0.703462574     0.17921871     0.229330322
#> [134,] -1.576760697     0.33084155    -0.101760340
#> [135,]  1.875977449     0.73572042     0.102725902
#> [136,] -1.006466885     1.05334483     0.284474419
#> [137,]  1.708799378     0.51215669     0.384696197
#> [138,]  1.008138723     0.45237902     0.210699934
#> [139,] -0.939897914     0.67440291     0.174036062
#> [140,] -1.346852419     0.41414237     0.076803981
#> [141,] -1.085225077     0.34887907    -0.030010474
#> [142,] -0.810304223     1.05156573     0.456111692
#> [143,] -0.155809787     0.27248374    -0.058005087
#> [144,] -0.514450356     0.27959373     0.358574791
#> [145,] -0.102561445     0.55174695     0.212175485
#> [146,] -0.519232895     0.49548717    -0.020630967
#> [147,]  2.117142484     0.19084135     0.260579982
#> [148,] -0.754699527     0.42876519     0.344276863
#> [149,] -2.274583105     0.46931743    -0.028724490
#> [150,] -1.393937294     0.15221155     0.476261804
#> [151,]  1.151437211     0.48284320     0.041999360
#> [152,]  0.175364469     0.12966096     0.485942854
#> [153,] -0.637568158     0.54338957     0.395448873
#> [154,] -2.085136446     0.35300459     0.200576826
#> [155,]  1.637486963     0.86253742     0.158200988
#> [156,]  0.455371375     0.42180076     0.710077325
#> [157,] -0.809392581     0.44008792     0.092933589
#> [158,] -2.320157673     0.29409237     0.129241135
#> [159,] -0.240540128     0.29005971     0.221687349
#> [160,]  0.584170921     0.50949071     0.210559344
#> [161,]  0.474615984     0.19930008     0.228169225
#> [162,] -1.235163906     0.18276872     0.230798854
#> [163,] -0.291937508     0.22902187     0.162574276
#> [164,]  1.074492546     0.55861328     0.437695445
#> [165,]  0.783763419     0.52081881     0.270830973
#> [166,] -1.382655068     0.54256011    -0.022102951
#> [167,] -1.469769051     0.25393496     0.098680780
#> [168,] -1.399009061     0.56483977    -0.141917051
#> [169,] -1.149167531     0.35131821    -0.047069522
#> [170,]  0.170432325     0.17045278     0.569893658
#> [171,] -0.455689315     0.58492821     0.316805711
#> [172,]  0.402141009     0.19944161     0.483715981
#> [173,] -0.685416978     0.69651030    -0.016226248
#> [174,] -0.982928893     0.62045138     0.153281477
#> [175,]  0.030904308     0.20394433    -0.089276102
#> [176,] -0.112860233     0.65404118     0.082302072
#> [177,] -0.670963432     0.14184656     0.314270967
#> [178,]  1.237315014     0.12737854     0.261311720
#> [179,] -0.607920741     0.14258202     0.568143128
#> [180,]  0.018299903     0.21834898     0.357922764
#> [181,]  1.512005217     0.67092306     0.081235887
#> [182,]  1.193495261     0.25530692     0.428611013
#> [183,]  1.483720159     0.10710862     0.553603933
#> [184,] -1.230642757     0.48596399     0.353346919
#> [185,]  0.475303346     0.21137929     0.142493858
#> [186,]  2.571655254     0.61811963     0.371719759
#> [187,] -0.245824995     0.52046694     0.101655269
#> [188,]  0.334761348     0.23672115     0.315847354
#> [189,] -0.571679757     0.40878703     0.403963539
#> [190,] -3.129607691     0.41026781     0.183393576
#> [191,]  0.632571597     0.37673154     0.410636945
#> [192,] -0.648642459     0.72805471     0.032964876
#> [193,]  0.598692814     0.27037089     0.143150116
#> [194,] -0.669172665     0.12058406     0.202589210
#> [195,]  0.229852906     0.21213894     0.078113657
#> [196,]  0.239574064     0.25297142     0.045565540
#> [197,]  0.590884802     0.14668636     0.050430010
#> [198,]  2.238839870     0.31903782     0.014939928
#> [199,]  0.434537998     0.16838656     0.426706158
#> [200,]  1.798634610     0.67722346     0.079159317
#> [201,]  0.596982408     0.27947600     0.488707187
#> [202,]  0.484282378     0.06052637     0.671122763
#> [203,]  1.429122844     0.48182334     0.228611590
#> [204,]  1.571335574     0.05578615     0.706774510
#> [205,]  0.693070539     0.89685359     0.159158268
#> [206,]  0.060194710     0.19559123     0.294311065
#> [207,] -0.579562235     0.30292978     0.303286151
#> [208,]  0.423658454     0.35688438     0.148574657
#> [209,]  0.301720553     0.15353179     0.549836936
#> [210,] -0.333815600     0.38769766     0.249333704
#> [211,] -0.201413500     0.50376773    -0.058329907
#> [212,]  0.664640463     0.58424366    -0.128133237
#> [213,]  0.079130605     0.41393641     0.375584580
#> [214,] -1.139840077     0.58001610     0.096521026
#> [215,]  0.446564743     0.23411980     0.076120600
#> [216,]  0.617404926     0.27564064     0.559561443
#> [217,] -1.575399288     0.43907639     0.191893699
#> [218,]  0.169332888     0.37858510     0.557320186
#> [219,] -1.529668678     0.62909780     0.148136887
#> [220,]  1.577804564     0.47057984     0.135553904
#> [221,]  0.880164762     0.44391386     0.536224727
#> [222,] -1.138908239     0.67579189    -0.098141875
#> [223,] -0.607431331     0.51617912     0.073531885
#> [224,]  0.591999850     0.23826112     0.279374762
#> [225,] -0.631093144     0.47657373     0.169322013
#> [226,]  0.274787489     0.22884282     0.283811019
#> [227,]  1.005827582     0.90880128     0.474286321
#> [228,]  0.225613290     0.37504039     0.181926155
#> [229,] -0.892544375     0.34175318     0.339566391
#> [230,]  1.118204473     0.25037940     0.108209886
#> [231,]  1.185769023     0.35576887     0.212901589
#> [232,] -0.251354028     0.58105844     0.295433678
#> [233,] -1.637710770     0.40476871     0.560593919
#> [234,]  2.239005752     0.34230246     0.058209517
#> [235,]  1.453998991     0.43693347     0.205107522
#> [236,] -0.146925936     0.35173303     0.201024380
#> [237,]  0.623367317     0.40396288     0.065281717
#> [238,] -0.237590111     0.14869201     0.327074276
#> [239,] -0.149143639     0.34773735     0.555141144
#> [240,]  1.005339204     0.69734185     0.225972187
#> [241,] -0.070149342     0.65312070     0.187706741
#> [242,] -0.163170235     0.48671544     0.147514693
#> [243,]  0.387358066     0.27835934    -0.025887806
#> [244,] -0.475117686     0.27707251    -0.009245679
#> [245,] -0.058936477     0.26473717     0.230978750
#> [246,] -0.447864821     0.40438950     0.267502714
#> [247,] -1.282530693     0.06675126     0.390270407
#> [248,] -0.946771926     0.13866273     0.133807571
#> [249,] -0.593441243     0.74286432     0.246586401
#> [250,]  0.033814803     0.33493987     0.161331650
#> [251,] -0.245918860     0.43685200     0.311780474
#> [252,]  0.043701579     0.25500796     0.551877576
#> [253,] -1.697033916     0.23111943     0.381994070
#> [254,]  0.352207028     0.56831420     0.549540714
#> [255,] -0.411882671     0.30724636     0.402079180
#> [256,] -1.020375215     0.27229712     0.060233134
#> [257,]  0.751320823     0.19294665     0.046424918
#> [258,]  0.557649181     0.25074966     0.235563018
#> [259,] -0.138736839     0.16746668     0.511970133
#> [260,]  1.024161502     0.32300811     0.461873263
#> [261,] -1.216678958     0.18654614     0.318462986
#> [262,] -1.195366329     0.07964890     0.777815920
#> [263,]  0.253212091     0.47525373     0.470384943
#> [264,]  0.520849352     0.30109521    -0.179100975
#> [265,]  1.579470747     0.23120193     0.265128831
#> [266,] -1.141790230     0.70929859    -0.037191176
#> [267,] -0.254750855     0.23262909     0.075693617
#> [268,]  0.423420202     0.15895447     0.315370852
#> [269,] -0.207581367     0.54703190     0.273991735
#> [270,] -1.047732717     0.47303920     0.328069768
#> [271,] -0.262041737     0.20806194     0.410386759
#> [272,] -0.740291010     0.73346625     0.072640091
#> [273,]  0.194361627     0.23745594     0.125603176
#> [274,]  0.670366376     0.44160547     0.071259094
#> [275,] -1.168759226     0.23014233     0.205798892
#> [276,]  0.499987435     0.27762768     0.105312664
#> [277,]  1.133429059     0.34491991     0.159687386
#> [278,] -0.394984126     0.42100366     0.458747797
#> [279,]  0.577930686     0.46538165     0.403330877
#> [280,]  1.218166739     0.35088080     0.653258938
#> [281,]  0.119837354     0.37547313     0.309771654
#> [282,] -0.325061169     0.22669903     0.358274241
#> [283,]  0.132917648     0.56977631     0.158588783
#> [284,] -0.979375522     0.25185430     0.372458832
#> [285,] -0.656125415     0.18986831     0.100571486
#> [286,] -0.747643464     0.52543479     0.124492537
#> [287,] -1.353370047     0.43002348     0.516918435
#> [288,] -0.653004567     0.45803683     0.098866423
#> [289,] -0.185118754     0.48797813     0.174023676
#> [290,] -0.023129593     0.28139048     0.457055795
#> [291,]  0.403197117     0.54726308    -0.028568372
#> [292,]  1.228056644     0.29391725     0.033182422
#> [293,] -1.525474081     0.16418196     0.298867697
#> [294,] -1.143233600     0.26918308     0.296645706
#> [295,]  0.204848929     0.39981436     0.313162157
#> [296,]  0.738615822     0.28309330     0.370247676
#> [297,]  0.329825394     0.55817368    -0.047895865
#> [298,] -0.040928693     0.65682264    -0.093084053
#> [299,] -0.531981348     0.27569170     0.220902251
#> [300,] -0.844418991     0.26630267     0.226331205
#> 
#> $acov_order
#> [1] 0
#> 
#> $acor_order
#> [1] 1
#> 
#> $N
#> [1] 300
#> 
#> $S
#> [1] 50
#> 
#> $R
#> [1] 1000
#>