The HPJ bias-corrected estimation of the moments

The `hpjmoment()` function enables to implement the HPJ bias-corrected estimation of the moments of the heterogeneous mean, the heterogeneous autocovariance, and the heterogeneous autocorrelation. The method is developed by Okui and Yanagi (2019). For more details, see the package vignette with `vignette("panelhetero")`.

Usage

hpjmoment(data, acov_order = 0, acor_order = 1, R = 1000)

Arguments

data

A matrix of panel data. Each row corresponds to individual time series.

acov_order

A non-negative integer of the order of autocovariance. Default is 0.

acor_order

A positive integer of the order of autocorrelation Default is 1.

R

A positive integer of the number of bootstrap repetitions. Default is 1000.

Value

A list that contains the following elements.

estimate

A vector of the parameter estimates

se

A vector of the standard errors

ci

A matrix of the 95 percent confidence intervals

quantity

A matrix of the estimated heterogeneous quantities

acov_order

The order of autocovariance

acor_order

The order of autocovariance

N

The number of cross-sectional units

S

The length of time series

R

The number of bootstrap repetitions

References

Okui, R. and Yanagi, T., 2019. Panel data analysis with heterogeneous dynamics. Journal of Econometrics, 212(2), pp.451-475.

Examples

data <- panelhetero::simulation(N = 300, S = 50)
panelhetero::hpjmoment(data = data)
#> $estimate
#>         E(mean)         E(acov)         E(acor)       var(mean)       var(acov) 
#>     -0.03504845      0.37057006      0.30320216      0.99550729      0.04341729 
#>       var(acor) cor(mean, acov) cor(mean, acor) cor(acov, acor) 
#>      0.02487815     -0.16809632      0.02953499     -0.55595106 
#> 
#> $se
#>         E(mean)         E(acov)         E(acor)       var(mean)       var(acov) 
#>     0.055513176     0.012856854     0.012671287     0.091822674     0.005494192 
#>       var(acor) cor(mean, acov) cor(mean, acor) cor(acov, acor) 
#>     0.004040021     0.052422494     0.072285797     0.065204783 
#> 
#> $ci
#>                 95% CI lower 95% CI upper
#> E(mean)          -0.14078062   0.08142024
#> E(acov)           0.34609785   0.39667615
#> E(acor)           0.27752039   0.32733641
#> var(mean)         0.81534349   1.17476358
#> var(acov)         0.03326131   0.05460267
#> var(acor)         0.01712202   0.03281176
#> cor(mean, acov)  -0.26580923  -0.06043847
#> cor(mean, acor)  -0.11371098   0.16979717
#> cor(acov, acor)  -0.68081826  -0.43152909
#> 
#> $quantity
#>               mean autocovariance autocorrelation
#>   [1,]  3.04961986     0.29408795     0.363966447
#>   [2,] -0.64434820     0.28968536     0.461720311
#>   [3,]  0.42942217     0.20084377     0.629836171
#>   [4,] -0.16638936     0.27210115     0.256800274
#>   [5,] -0.81550504     0.17010867     0.233320328
#>   [6,] -1.29080059     0.58314716     0.008962518
#>   [7,] -1.28707255     0.38544229     0.357486022
#>   [8,] -0.85880114     0.40041141     0.164500349
#>   [9,]  1.30500088     0.19776878     0.281056958
#>  [10,]  0.23405446     0.14037010     0.306782110
#>  [11,] -0.90541219     0.29188139     0.430847978
#>  [12,] -0.17104596     0.30815667     0.492353666
#>  [13,] -0.85558929     0.31204654     0.123113342
#>  [14,] -1.85025273     0.14388306     0.238549890
#>  [15,] -0.40539822     0.35735876     0.394525947
#>  [16,]  3.52121652     0.27814619     0.186580779
#>  [17,] -0.10934803     0.31786758     0.103550780
#>  [18,] -0.34063468     0.40792645     0.062455271
#>  [19,] -0.51283991     0.12116064     0.683558417
#>  [20,] -0.50865821     0.73468592     0.058995111
#>  [21,] -0.07632436     0.47910859     0.595694404
#>  [22,]  1.13397080     0.07541483    -0.073782970
#>  [23,] -0.60973822     0.38953816     0.234620928
#>  [24,] -1.73392477     0.96530092     0.188195881
#>  [25,] -0.27682884     0.30512211     0.120331741
#>  [26,] -1.61964857     0.21635127     0.439600258
#>  [27,] -0.91323459     0.62294515     0.044465440
#>  [28,] -0.13130420     1.01780214     0.332357409
#>  [29,] -1.95000658     0.22429535     0.225426283
#>  [30,] -0.51161925     0.28933165     0.187066574
#>  [31,]  1.14932757     0.32775298     0.094571264
#>  [32,] -1.36974544     0.17815603     0.356120668
#>  [33,]  1.66882762     0.50211408    -0.044174768
#>  [34,]  1.14910581     0.15863709     0.343018727
#>  [35,] -2.12865687     0.17010424     0.434930160
#>  [36,]  0.06546449     0.04695349     0.358595237
#>  [37,]  0.19157777     0.65542798     0.299427460
#>  [38,] -1.20497956     0.41156403    -0.083301825
#>  [39,] -0.06187703     0.22626160     0.383731558
#>  [40,]  2.37900223     0.22589831     0.036253673
#>  [41,]  0.15074603     0.15415382     0.409745430
#>  [42,] -0.29156593     0.48934612     0.153134304
#>  [43,]  0.30404228     0.44171238     0.541388751
#>  [44,] -0.20494395     0.40810430     0.448871539
#>  [45,]  0.35871097     0.51892747     0.174454772
#>  [46,]  1.66663529     0.32714843     0.314226187
#>  [47,] -0.01199071     0.14138563     0.679093643
#>  [48,]  0.95262418     0.10033762     0.481704255
#>  [49,]  0.09382220     0.25418975     0.297159372
#>  [50,]  0.66215010     0.08905990     0.412533408
#>  [51,] -1.56764754     0.41256454    -0.043606870
#>  [52,] -0.40308001     1.36572913     0.291305244
#>  [53,]  0.08287375     0.80364485    -0.045287224
#>  [54,] -1.11526397     0.72282740     0.275484155
#>  [55,] -1.23510543     0.08599978     0.284761188
#>  [56,]  0.39131165     0.47744807     0.065666539
#>  [57,] -0.45324279     0.28907114    -0.002627504
#>  [58,]  1.94692949     0.28458741     0.370489342
#>  [59,] -0.11246807     0.60662326     0.357374695
#>  [60,] -0.89184805     0.18485065     0.343378916
#>  [61,]  0.49892342     1.07031778     0.198668192
#>  [62,] -1.38857084     0.13259724     0.483132537
#>  [63,]  0.28695698     0.36435925     0.508669453
#>  [64,] -1.14212561     0.93700097     0.047600359
#>  [65,]  0.02604875     0.25239617     0.282943033
#>  [66,] -1.20711161     0.37627940     0.245469450
#>  [67,] -0.35894154     0.16156280     0.326490195
#>  [68,] -1.81552814     0.49783611     0.181750067
#>  [69,] -1.18058086     0.16845951    -0.046816037
#>  [70,] -0.32726676     0.71599478     0.458162220
#>  [71,] -0.59508693     1.00988161     0.173981749
#>  [72,] -0.97447238     0.69857543     0.243294523
#>  [73,]  0.64938196     0.40055363     0.337675976
#>  [74,]  0.30004038     0.28258295     0.174641940
#>  [75,]  0.28852320     0.08146592     0.452697223
#>  [76,]  0.51542620     0.13178562     0.148669650
#>  [77,]  0.09485628     0.42012567     0.139969964
#>  [78,]  1.43166846     0.18733533     0.393463309
#>  [79,] -0.41315485     0.67296562     0.047520090
#>  [80,] -0.36203395     0.08711721     0.487338026
#>  [81,]  1.71483718     0.32395989     0.360560450
#>  [82,] -0.25789114     0.32926485     0.256456793
#>  [83,] -0.01138307     0.29953397     0.339812210
#>  [84,] -0.88387007     0.09098803     0.632705591
#>  [85,] -0.39693919     0.55302115     0.045671025
#>  [86,]  0.33217908     0.33190266     0.422959826
#>  [87,] -0.76594576     0.64175811     0.182226686
#>  [88,] -0.57067534     0.60821408    -0.030847434
#>  [89,] -1.21922747     0.32756626     0.149179625
#>  [90,] -0.93992998     0.29262283     0.379944462
#>  [91,]  0.58411864     0.36110871     0.181273134
#>  [92,] -1.13644624     0.27040597     0.244810816
#>  [93,]  0.12885508     0.36299424     0.114868689
#>  [94,] -0.86272001     0.41270568     0.262743405
#>  [95,]  0.96672094     0.10111139     0.515187160
#>  [96,] -0.16894282     0.33614182     0.156234445
#>  [97,]  1.63426664     0.25776734     0.035135657
#>  [98,]  0.09230101     0.53078540     0.221382474
#>  [99,] -2.34354545     0.60846857     0.130125065
#> [100,]  0.19510511     0.28171026     0.345672968
#> [101,]  0.20237613     0.37067683     0.155457310
#> [102,] -0.98371159     0.73593914     0.427915129
#> [103,]  0.47205965     0.26342548     0.143590281
#> [104,]  0.77177812     0.30867756     0.417815400
#> [105,]  0.92693642     0.33178492     0.163041619
#> [106,]  0.44495168     0.26425901     0.392426753
#> [107,]  0.63005651     0.18581970     0.110799783
#> [108,]  0.32305185     0.24383415     0.179514650
#> [109,] -0.08275834     0.36106130     0.068696911
#> [110,] -0.58408168     0.18312456     0.420970264
#> [111,] -0.43342173     0.55217021     0.331137983
#> [112,] -0.54769015     0.83831283     0.028813632
#> [113,] -0.99682446     0.33563015     0.288022280
#> [114,] -0.19897185     0.48014351     0.211350751
#> [115,]  0.83597015     0.08592484     0.381144148
#> [116,]  0.11914212     0.30246760     0.607058063
#> [117,] -0.14726207     0.77950681     0.112066285
#> [118,]  1.33582222     0.25229099     0.273883350
#> [119,] -1.18186805     0.19788742     0.289598448
#> [120,]  1.54714393     0.06755732     0.609976835
#> [121,]  0.64435798     0.31124188     0.010973119
#> [122,] -1.03460599     0.25826844    -0.137578931
#> [123,] -0.79039763     0.06712478     0.581394887
#> [124,]  1.05515060     0.22371681     0.272127824
#> [125,]  0.81800388     0.24045964     0.112472545
#> [126,]  0.60977930     0.30816281     0.132324868
#> [127,] -0.50775143     0.22203269     0.430570459
#> [128,]  0.93970693     0.13487105     0.248698858
#> [129,] -0.95781466     0.28505294     0.396785846
#> [130,]  0.06681203     0.50549559     0.279106404
#> [131,] -0.21447542     0.18471615     0.106598453
#> [132,]  1.22444234     0.32356260     0.007288011
#> [133,]  2.22768161     0.46785880     0.103524990
#> [134,]  1.27321811     0.42129265     0.030673697
#> [135,]  0.12515299     0.50546440    -0.044079282
#> [136,] -0.63239964     0.16832933     0.307869056
#> [137,] -0.61448005     0.07759439     0.207595715
#> [138,] -0.14482976     0.73416670     0.327344141
#> [139,] -0.35994075     0.16990836     0.396270696
#> [140,]  1.25770572     0.13910732     0.006226765
#> [141,] -0.04226786     0.34410698    -0.045878458
#> [142,]  0.73965427     0.55157684     0.283482408
#> [143,]  0.37420851     0.06112448     0.651732771
#> [144,]  0.40101722     0.86450553     0.384670806
#> [145,]  0.16208250     0.41859630     0.036487535
#> [146,]  0.26037304     0.80595588    -0.002701140
#> [147,]  0.04342312     0.28948682     0.287640308
#> [148,]  0.17263996     0.24937679     0.431799979
#> [149,] -0.39516642     0.87122595    -0.074440398
#> [150,]  0.63076489     0.28724944     0.335942641
#> [151,]  1.17138061     0.43533327     0.333699179
#> [152,]  0.29435947     0.21447784     0.339658897
#> [153,] -1.15419345     0.53471487     0.353579606
#> [154,] -0.59270630     0.58280527     0.032982742
#> [155,] -0.97861570     0.33479582     0.357175616
#> [156,] -0.61266166     0.35843830     0.065849411
#> [157,]  0.33220905     0.69959336     0.343317613
#> [158,] -0.10153223     0.23615097     0.391557224
#> [159,] -0.36153598     0.74956938     0.071648025
#> [160,] -1.23394356     0.53796203     0.045750535
#> [161,]  0.69707522     0.37539146     0.263364140
#> [162,]  0.89111684     0.13785149     0.410827217
#> [163,]  0.13939036     0.88089014    -0.042015242
#> [164,]  0.22587371     0.36345878     0.318194421
#> [165,]  1.77896922     0.14274017     0.459911126
#> [166,] -1.35686039     0.94911955    -0.360446671
#> [167,] -0.72871703     0.97170356     0.208836987
#> [168,] -0.92165413     0.34491219     0.181744187
#> [169,] -1.23080709     0.05294680     0.727293504
#> [170,]  2.06148294     0.33254997     0.208849884
#> [171,]  0.57491211     0.69040183     0.252888984
#> [172,]  1.24182473     0.24020849     0.323075231
#> [173,]  1.54993241     0.33548223     0.247895334
#> [174,] -1.55394231     0.56604762     0.388290959
#> [175,]  0.12131731     0.20911266     0.217808481
#> [176,] -0.76329046     0.37886648     0.317997994
#> [177,]  0.87478992     0.22717152     0.022510748
#> [178,] -0.60044413     0.17629632     0.158207592
#> [179,] -0.82449307     0.32953642    -0.052088564
#> [180,]  2.12727362     0.22843211     0.022885910
#> [181,]  1.56026592     0.20340610     0.243625778
#> [182,]  0.19231847     0.33039373     0.246295393
#> [183,]  1.20763693     0.40732355     0.252179764
#> [184,]  0.80197003     0.23054074     0.348799153
#> [185,] -1.18231836     0.18052379     0.413341852
#> [186,]  0.21971745     0.10937446     0.518069951
#> [187,] -0.84679475     0.32417933     0.369193311
#> [188,]  3.35789280     0.60253357     0.044409461
#> [189,]  0.94383212     0.18442829     0.312469888
#> [190,] -0.19892709     0.44663380     0.104191752
#> [191,]  0.69953298     0.36897412     0.163366581
#> [192,]  0.16384308     0.04931422     0.660089321
#> [193,] -0.46242973     0.13291722     0.280335925
#> [194,]  0.52238418     0.30252085     0.513930594
#> [195,]  0.73192640     0.14987604     0.631614284
#> [196,]  1.69272020     0.12505570     0.465756558
#> [197,]  0.93905145     0.41715460    -0.149186649
#> [198,] -0.12030685     0.11080287     0.559640033
#> [199,] -1.12589867     0.22715183     0.057508786
#> [200,]  0.47649127     0.23567799     0.339250014
#> [201,] -0.97188793     0.21561077     0.252095017
#> [202,]  1.31242189     0.50442927     0.378944527
#> [203,]  1.03373703     0.33562519     0.223118723
#> [204,] -0.31989787     1.04339182     0.285298716
#> [205,] -1.15021328     0.15296533     0.170931091
#> [206,] -0.02949243     0.18272581     0.544845440
#> [207,] -0.35184894     0.31311057     0.216579138
#> [208,]  1.52315979     0.03474522     0.495775158
#> [209,]  0.06928301     0.22859781     0.660550670
#> [210,] -2.16000290     0.21595251    -0.138362802
#> [211,] -1.00556430     0.53921082     0.453473000
#> [212,] -0.09124669     0.22373258     0.525594953
#> [213,]  0.38612638     0.49377002     0.364781158
#> [214,] -0.65405309     0.32192620    -0.169219606
#> [215,]  1.95957038     0.32857535     0.297604273
#> [216,] -0.26566828     0.06823033     0.777407281
#> [217,]  1.92066739     0.17905820     0.314288507
#> [218,] -1.59319875     0.11295825     0.383034569
#> [219,] -0.68058717     0.21932468     0.222075934
#> [220,] -0.56904531     0.39711394     0.111286075
#> [221,]  0.03859268     0.34622151    -0.028336212
#> [222,] -1.03326098     0.55112269     0.310596371
#> [223,] -2.48135305     0.55112350     0.458773485
#> [224,] -0.41178162     0.42975731     0.294922179
#> [225,]  0.44079003     0.36252227     0.109655953
#> [226,] -0.32707962     0.22363639     0.224724649
#> [227,]  0.10733780     0.36159166     0.057642050
#> [228,] -2.09299064     0.32911599     0.361164085
#> [229,] -0.63542194     0.38158786     0.162248066
#> [230,] -2.29632658     0.38981693     0.381840370
#> [231,] -1.52021416     0.83716059    -0.126900634
#> [232,] -0.61986881     0.13460688     0.414337255
#> [233,] -0.47374146     0.12857892     0.380202225
#> [234,] -0.24689856     0.30722984     0.208511359
#> [235,]  1.22217354     0.51581611     0.092605225
#> [236,]  0.25431229     0.18931042     0.660045411
#> [237,] -0.91152394     0.27145876     0.633669335
#> [238,]  1.01502629     0.25626579     0.481779479
#> [239,] -0.07823437     0.30366925     0.164549020
#> [240,]  1.33532660     0.25793483     0.391681163
#> [241,]  0.02443276     0.57369716     0.254025187
#> [242,]  0.24658010     0.44551374     0.404238653
#> [243,]  1.27087115     0.33342482    -0.005532753
#> [244,]  0.55243595     0.25278877     0.410737055
#> [245,]  0.15849087     0.59823691     0.345525917
#> [246,] -0.94745976     0.53371082     0.072226427
#> [247,] -0.34802893     0.19943555     0.660944212
#> [248,]  0.95913733     0.71170699    -0.142966682
#> [249,] -0.19346902     0.34916810     0.343572833
#> [250,] -1.15314362     0.02463764     0.616160399
#> [251,] -0.62591350     0.32562990     0.533678621
#> [252,]  0.54757848     0.10359417     0.526920283
#> [253,]  0.07729622     0.17298625     0.381204013
#> [254,]  1.03398024     0.10906636     0.426716338
#> [255,]  0.44390321     0.04714318     0.634880132
#> [256,]  0.69652334     0.34698024     0.214349995
#> [257,]  0.51301259     0.32845796     0.130593024
#> [258,] -0.42692855     0.52189188     0.243613214
#> [259,]  1.93898038     0.26894324     0.704431377
#> [260,] -2.07788877     0.50493102     0.050452236
#> [261,] -0.49621287     0.44745904     0.266472486
#> [262,] -0.03340608     0.32982194     0.304033371
#> [263,]  1.07511798     0.19079017    -0.105900120
#> [264,] -1.41967678     0.20771909     0.194816946
#> [265,]  0.89267518     0.19367024     0.474255919
#> [266,]  0.45294999     0.62453194    -0.015164938
#> [267,] -0.99971086     0.83948409     0.099220880
#> [268,]  0.12887181     0.43518101     0.447205456
#> [269,] -0.07301464     0.28921315     0.188429349
#> [270,] -0.47531945     0.19720041     0.401714705
#> [271,] -0.59372641     0.36419374     0.430389071
#> [272,] -1.50255260     0.20805844     0.549664866
#> [273,] -0.50650763     0.31165685     0.155337032
#> [274,]  0.93872836     0.63347433     0.049981933
#> [275,] -1.15910404     0.39804251    -0.275656181
#> [276,] -2.09644368     0.31788026     0.450613250
#> [277,]  0.42175290     0.28738765     0.170126697
#> [278,] -0.22202242     0.30935509     0.114341873
#> [279,] -0.12463293     0.45005259     0.330146025
#> [280,]  0.71231553     0.24536483     0.631489584
#> [281,] -0.11219356     0.43058941     0.192350421
#> [282,] -1.40441960     0.31248105     0.610662094
#> [283,]  0.94413752     0.07258836     0.233573715
#> [284,]  0.83384356     0.10360349     0.237516475
#> [285,] -0.52521862     0.26530273     0.201696271
#> [286,]  0.73293477     0.16642141     0.485408231
#> [287,]  0.45719117     0.94037842     0.532155644
#> [288,]  0.17237602     0.26044451     0.216502645
#> [289,] -1.15142979     0.37527319    -0.164214591
#> [290,] -1.32117934     0.24583342     0.130041640
#> [291,] -0.25744798     0.68184600    -0.065833628
#> [292,]  0.95462562     0.40491941     0.509842204
#> [293,]  0.48123135     0.16272501     0.530143534
#> [294,] -1.73545029     0.44171043    -0.076476900
#> [295,]  0.54634043     0.14790817     0.349647858
#> [296,] -0.39457062     0.64228983     0.003919434
#> [297,]  0.14613743     0.36904108     0.141313641
#> [298,]  0.81953776     0.31450203     0.151277881
#> [299,] -1.37824488     0.10200102     0.450189650
#> [300,]  0.28839399     0.94966306     0.256592423
#> 
#> $acov_order
#> [1] 0
#> 
#> $acor_order
#> [1] 1
#> 
#> $N
#> [1] 300
#> 
#> $S
#> [1] 50
#> 
#> $R
#> [1] 1000
#>