The HPJ bias-corrected estimation of the moments

The `hpjmoment()` function enables to implement the HPJ bias-corrected estimation of the moments of the heterogeneous mean, the heterogeneous autocovariance, and the heterogeneous autocorrelation. The method is developed by Okui and Yanagi (2019). For more details, see the package vignette with `vignette("panelhetero")`.

Usage

hpjmoment(data, acov_order = 0, acor_order = 1, R = 1000)

Arguments

data

A matrix of panel data. Each row corresponds to individual time series.

acov_order

A non-negative integer of the order of autocovariance. Default is 0.

acor_order

A positive integer of the order of autocorrelation Default is 1.

R

A positive integer of the number of bootstrap repetitions. Default is 1000.

Value

A list that contains the following elements.

estimate

A vector of the parameter estimates

se

A vector of the standard errors

ci

A matrix of the 95 percent confidence intervals

quantity

A matrix of the estimated heterogeneous quantities

acov_order

The order of autocovariance

acor_order

The order of autocovariance

N

The number of cross-sectional units

S

The length of time series

R

The number of bootstrap repetitions

References

Okui, R. and Yanagi, T., 2019. Panel data analysis with heterogeneous dynamics. Journal of Econometrics, 212(2), pp.451-475.

Examples

data <- panelhetero::simulation(N = 300, S = 50)
panelhetero::hpjmoment(data = data)
#> $estimate
#>         E(mean)         E(acov)         E(acor)       var(mean)       var(acov) 
#>     0.129611497     0.403443233     0.266378827     0.875222744     0.041888924 
#>       var(acor) cor(mean, acov) cor(mean, acor) cor(acov, acor) 
#>     0.023853650    -0.089299296    -0.004099997    -0.548742872 
#> 
#> $se
#>         E(mean)         E(acov)         E(acor)       var(mean)       var(acov) 
#>     0.053742998     0.012767877     0.012791660     0.073624840     0.004484584 
#>       var(acor) cor(mean, acov) cor(mean, acor) cor(acov, acor) 
#>     0.003972889     0.067270949     0.067329930     0.065508567 
#> 
#> $ci
#>                 95% CI lower 95% CI upper
#> E(mean)           0.03173726   0.23794030
#> E(acov)           0.37936701   0.42798691
#> E(acor)           0.24162313   0.29071802
#> var(mean)         0.73974831   1.02298414
#> var(acov)         0.03309569   0.05075610
#> var(acor)         0.01606805   0.03173719
#> cor(mean, acov)  -0.21939779   0.04774620
#> cor(mean, acor)  -0.13540990   0.12973618
#> cor(acov, acor)  -0.67544735  -0.42173436
#> 
#> $quantity
#>                mean autocovariance autocorrelation
#>   [1,]  1.993575122     0.26614860     0.149963788
#>   [2,] -0.580966275     0.34461750     0.471728513
#>   [3,]  1.258516947     0.62600574     0.239354009
#>   [4,] -1.146829930     0.59337692     0.283329594
#>   [5,]  1.287469539     0.28967338     0.255417449
#>   [6,]  0.566690609     0.35134922     0.435680273
#>   [7,] -0.061926689     0.47597894     0.054998826
#>   [8,]  1.221628609     0.50974032    -0.066218827
#>   [9,] -0.087096129     0.25999138     0.146672432
#>  [10,] -0.400984227     0.38504664     0.023936065
#>  [11,] -1.063542804     0.17174664     0.346413694
#>  [12,]  2.283491015     0.37748311     0.236253747
#>  [13,] -1.970887922     0.78835133     0.140526470
#>  [14,]  0.573091401     0.39177481     0.092360232
#>  [15,] -0.819196847     0.19838576     0.563173596
#>  [16,] -0.745131558     0.29351096     0.162296961
#>  [17,] -0.547859423     0.22566937     0.289675368
#>  [18,]  0.963748828     0.24867006     0.170062472
#>  [19,]  0.770727675     0.27744315     0.057187627
#>  [20,]  0.143113232     0.24655511     0.268360342
#>  [21,] -0.424156414     0.45989403     0.236456656
#>  [22,]  0.389838386     0.16673250     0.244017182
#>  [23,] -0.220347578     0.83035774    -0.121070466
#>  [24,]  0.645452051     0.24488912     0.282853753
#>  [25,]  0.205852703     0.35629770     0.071586537
#>  [26,] -0.277330754     0.17420567     0.404653596
#>  [27,] -0.327389675     0.27003058     0.118190766
#>  [28,] -0.016026414     0.51734529     0.430836418
#>  [29,]  0.204780252     0.27118099     0.185486883
#>  [30,] -0.563678494     0.43367739     0.242508954
#>  [31,] -0.802155952     0.20212886     0.389797606
#>  [32,] -0.894553311     0.29972408     0.407592003
#>  [33,] -1.152242683     0.21841501     0.458506822
#>  [34,]  2.116394252     0.14189361     0.403881239
#>  [35,] -0.335361342     0.15744418     0.464686015
#>  [36,]  0.043948289     0.28919369     0.090559737
#>  [37,]  0.356669218     0.26490301     0.712759884
#>  [38,]  0.466557430     0.40537390    -0.383577640
#>  [39,]  1.994690640     0.47025545     0.412376552
#>  [40,]  0.461535334     0.29883461     0.046831058
#>  [41,]  0.299621202     0.22274770     0.385891874
#>  [42,] -0.788293855     0.65262679     0.064754531
#>  [43,]  1.095678961     0.27138217     0.164642359
#>  [44,] -0.792491428     0.77765281     0.218756244
#>  [45,]  1.355157668     0.45385201     0.273210493
#>  [46,] -1.038187369     1.05018078     0.485970528
#>  [47,]  0.894584327     0.41961072     0.263217525
#>  [48,]  0.536258743     0.32137914     0.111702484
#>  [49,] -0.027246195     0.67250503     0.329700417
#>  [50,] -0.963444638     0.09830515     0.479730098
#>  [51,]  0.984514365     0.20104200     0.266223699
#>  [52,] -0.782649133     0.84065304     0.047485311
#>  [53,]  0.461152143     0.23688687     0.066897477
#>  [54,]  0.060136269     0.37248216     0.297274126
#>  [55,] -0.485291938     0.22028241     0.155354514
#>  [56,]  1.508019572     0.15748456     0.275561646
#>  [57,]  0.632763591     0.19240955     0.263994769
#>  [58,] -0.014877382     0.67885692    -0.091690772
#>  [59,]  0.170250964     0.41337338    -0.008120418
#>  [60,] -0.207555979     0.33151518     0.188252403
#>  [61,] -0.381313461     0.21144236     0.491971897
#>  [62,] -0.207248472     0.19612174     0.352652219
#>  [63,]  1.470781336     0.28067255     0.214600349
#>  [64,]  0.097350696     0.28131141     0.042287982
#>  [65,]  0.040852106     0.43719216     0.093355360
#>  [66,] -1.945484445     0.40785896     0.171115871
#>  [67,]  0.295822606     0.62041546     0.330168549
#>  [68,]  0.196516197     0.12766418     0.223585959
#>  [69,]  1.636531168     0.46096954     0.384518946
#>  [70,]  0.046449592     0.68409062     0.028642456
#>  [71,]  1.037334442     0.15903828     0.555245895
#>  [72,] -1.423594048     0.43355190    -0.038728596
#>  [73,] -0.203115339     0.27324316     0.326567644
#>  [74,]  0.064364868     0.50039573    -0.044190726
#>  [75,]  0.350397934     0.25760728     0.213056249
#>  [76,]  1.292827971     0.24896816     0.063986838
#>  [77,] -0.896696357     0.57549850     0.212255667
#>  [78,] -0.386660201     0.28518753     0.109720770
#>  [79,]  0.055199332     0.57858411     0.180869819
#>  [80,]  1.247686226     0.21720381     0.116736631
#>  [81,]  0.847017129     0.53237749     0.390719265
#>  [82,]  0.111373303     0.32952754     0.153864235
#>  [83,] -0.083382282     0.25460456     0.156011815
#>  [84,]  0.326073636     0.20188815     0.150363063
#>  [85,]  0.360163433     0.34001680     0.057390782
#>  [86,]  0.090479021     0.20517767     0.445021561
#>  [87,] -0.650880473     0.53422328     0.223215940
#>  [88,]  0.986345311     0.35404663     0.378656456
#>  [89,]  1.001992152     0.35959924     0.087518315
#>  [90,]  0.572651550     0.35611098     0.171408703
#>  [91,] -0.132870346     0.36514691     0.159172416
#>  [92,] -0.514519104     0.30384938     0.006318166
#>  [93,]  0.911657273     0.20167859     0.363763513
#>  [94,] -2.660190391     0.37068471     0.246993565
#>  [95,]  0.000110202     0.46211101     0.306908705
#>  [96,] -1.674714849     0.25555608     0.587863416
#>  [97,]  1.287196999     0.23477926     0.419930720
#>  [98,]  1.030386062     1.04545943     0.231191331
#>  [99,]  0.307692727     0.42460180     0.210661458
#> [100,]  0.457291479     0.15312099     0.275557521
#> [101,]  1.837410666     0.67169112     0.316454003
#> [102,]  1.148522858     0.42674289     0.043916952
#> [103,]  0.117574166     0.41182370    -0.091864081
#> [104,]  0.661708035     0.57666070    -0.020967463
#> [105,]  0.005082925     0.12716665     0.174894042
#> [106,]  1.130542633     0.22330770     0.067424741
#> [107,] -0.037044520     0.26316283     0.127504408
#> [108,] -0.767453923     1.03267662     0.252331749
#> [109,]  1.275623978     0.12022097     0.410523354
#> [110,] -0.712784549     0.28447090     0.035308112
#> [111,]  1.051507110     0.54297438     0.189865745
#> [112,]  1.484517956     0.47346323     0.151340365
#> [113,] -1.065640317     0.77646801     0.298290531
#> [114,]  0.640317406     0.41567788     0.260002330
#> [115,]  0.327647370     0.28679449     0.340607728
#> [116,] -1.439825687     1.07169971     0.161913863
#> [117,] -1.190147968     0.21178752     0.023291985
#> [118,] -0.596077936     0.41924580    -0.108810186
#> [119,]  1.412847030     0.27345326     0.271156526
#> [120,]  0.825510178     0.44809580     0.142123326
#> [121,] -0.392925378     0.20150792     0.210028680
#> [122,]  0.352886208     0.99204733     0.398888337
#> [123,]  1.109814778     0.45865164     0.086858843
#> [124,]  1.138946765     0.26101670    -0.028441551
#> [125,] -1.829725044     0.29086713     0.130211973
#> [126,] -0.507412115     0.45260261     0.338233360
#> [127,] -1.693857542     0.98181149    -0.229261594
#> [128,] -0.921089195     0.16145092     0.340763833
#> [129,] -0.348696918     0.23228387     0.007684107
#> [130,] -0.555724166     0.11064666     0.602995811
#> [131,]  1.623737768     0.91749555     0.175686160
#> [132,] -0.080250249     0.43179518     0.360459624
#> [133,]  1.123249071     0.37900630     0.419687561
#> [134,] -0.260758917     0.67145087    -0.192247128
#> [135,]  1.465908315     0.13965388     0.154661233
#> [136,]  1.588296548     0.47402899     0.151006367
#> [137,] -0.876832546     0.17406827     0.687502214
#> [138,] -0.747282629     0.45610452     0.284408118
#> [139,] -1.697166018     0.59670102     0.297764354
#> [140,] -0.264865087     0.26584522     0.380584730
#> [141,]  1.419074704     0.12186723     0.433183034
#> [142,] -0.789172506     0.50078296     0.163163104
#> [143,] -0.104431920     0.36556038     0.068668332
#> [144,]  0.067277017     0.08571884     0.575768287
#> [145,]  0.111812388     0.59996315     0.066430969
#> [146,]  0.842687785     0.42023243    -0.001384988
#> [147,] -2.315096154     0.23876981     0.185237332
#> [148,] -0.174079418     0.11702619     0.413831162
#> [149,]  2.164142917     0.19157508     0.181478011
#> [150,] -0.348217621     0.28414235     0.529091235
#> [151,]  0.271273502     0.21235784     0.227320871
#> [152,] -0.365817689     0.33449958     0.382171165
#> [153,]  1.137174336     0.27109565     0.167724614
#> [154,]  1.473884552     0.50805925     0.188324040
#> [155,]  0.952299811     0.11052328     0.258666317
#> [156,]  1.867640979     0.09093929     0.534887256
#> [157,]  1.610192731     0.70072384     0.285502755
#> [158,]  0.718764846     0.42306370     0.047347691
#> [159,]  1.505492168     0.15240136     0.593709434
#> [160,] -0.936668685     0.17434607     0.384597185
#> [161,]  0.769061582     0.29875581     0.459525270
#> [162,] -0.283199666     0.41075158     0.387609393
#> [163,]  1.794765198     0.10257163     0.135203773
#> [164,] -0.510661055     0.86048405    -0.049558660
#> [165,]  0.557267768     0.29655399     0.185004923
#> [166,] -0.066842417     0.15796164     0.324668139
#> [167,] -0.806955023     0.13574544     0.347479744
#> [168,]  1.047512102     0.73935809    -0.125450226
#> [169,] -1.047035993     0.57837189    -0.105529297
#> [170,] -2.471944607     0.62081115    -0.133272712
#> [171,]  0.613574683     0.31157260     0.290512381
#> [172,] -0.516285479     0.50050496     0.039553285
#> [173,] -0.338419853     0.29476844     0.259670462
#> [174,]  0.033142126     1.10238172     0.011811565
#> [175,]  0.483300474     0.16327843     0.667243234
#> [176,] -0.829368199     0.12887933     0.518698858
#> [177,]  1.382774737     0.70969632     0.234369980
#> [178,]  1.834531906     0.20048037     0.686053446
#> [179,]  0.252311531     0.65494611     0.656060030
#> [180,] -0.150701163     0.35810244     0.236659276
#> [181,] -0.769393940     0.24874536     0.169879476
#> [182,] -0.128206848     0.45691712     0.369126701
#> [183,] -0.938526320     0.26945894     0.351736682
#> [184,]  0.186307448     0.35004169    -0.031885360
#> [185,]  1.724171193     0.54924834     0.067366265
#> [186,]  0.567890730     0.16391057     0.076708540
#> [187,]  1.224924785     0.58743412     0.048796346
#> [188,] -0.167060700     0.45448302     0.201308075
#> [189,]  0.592238020     0.39893855     0.173741368
#> [190,]  1.551151979     0.42357698     0.082835588
#> [191,] -0.668390424     0.22067397     0.095523596
#> [192,] -2.354209731     0.55613973     0.190691163
#> [193,]  0.357165340     0.75474211    -0.181597485
#> [194,] -0.470430866     0.62199311     0.333470983
#> [195,]  1.000118854     0.48244464    -0.002231647
#> [196,] -0.220412924     0.38632331     0.184276926
#> [197,] -0.634785818     0.69616589     0.324453569
#> [198,] -0.601303805     0.60372689     0.052064554
#> [199,]  0.248210067     0.41954231     0.357405898
#> [200,]  1.088616608     0.32741683    -0.061612258
#> [201,] -0.609811413     0.13697778     0.644730135
#> [202,]  1.509365260     0.60937262     0.135070326
#> [203,] -0.659860698     0.29105332     0.420375990
#> [204,]  0.447736784     0.21063156     0.661648808
#> [205,] -0.568635515     0.45196667     0.126794285
#> [206,] -0.946564043     0.14044880     0.294818368
#> [207,]  0.293587542     0.15350888     0.465807791
#> [208,]  0.227183480     0.40104336     0.117529113
#> [209,]  0.008776047     0.91456807    -0.061657946
#> [210,] -0.131716571     0.52897667     0.009990940
#> [211,]  1.060114864     0.41334793     0.263500573
#> [212,] -1.524585636     0.67364288     0.300847115
#> [213,]  1.568890476     0.09758290     0.488414176
#> [214,]  0.638500488     0.48295713    -0.193582470
#> [215,] -0.721599070     0.12220171     0.545915893
#> [216,]  0.106601108     0.60017788    -0.135891795
#> [217,]  0.339852420     0.20792199     0.214807312
#> [218,]  0.258835956     0.53438464    -0.118151140
#> [219,] -0.016800676     0.21261804     0.393128200
#> [220,] -1.515453303     0.20620510     0.041425134
#> [221,]  0.642769182     0.43511261     0.278604725
#> [222,]  0.780286519     0.33580200     0.037396134
#> [223,] -0.560481275     0.17716920    -0.039295210
#> [224,] -0.307808436     0.38804591     0.503869986
#> [225,]  0.530734532     0.09383386     0.429215042
#> [226,]  1.558805297     0.26826551     0.270357500
#> [227,]  0.187167879     0.07747646     0.505381319
#> [228,]  0.121022798     0.25514426     0.459370929
#> [229,]  0.443920120     0.18323760     0.403499608
#> [230,]  0.117646837     0.19881624     0.257145085
#> [231,]  1.094569035     0.17502378     0.436530432
#> [232,]  1.025088988     0.52624747    -0.155696455
#> [233,]  0.264489099     0.43966964     0.278953726
#> [234,] -0.853838128     0.14975087     0.524237947
#> [235,]  0.544504274     0.38337213     0.260292340
#> [236,]  0.508714620     0.20763805     0.245717459
#> [237,] -0.611303271     0.29264397     0.664191580
#> [238,] -0.567453744     0.25599783     0.289637114
#> [239,] -0.728360306     0.44311608     0.195295283
#> [240,] -1.139150722     0.59346985     0.060951305
#> [241,] -0.529043841     0.79452960     0.002994837
#> [242,]  0.417262609     0.42700791    -0.014696466
#> [243,] -0.527823799     0.41585819     0.200199561
#> [244,]  0.797585996     0.31695842     0.127598389
#> [245,]  1.524616912     0.13644698     0.389873237
#> [246,]  1.785175663     0.27943366     0.264100144
#> [247,]  1.363921060     0.43295354    -0.070022007
#> [248,] -0.289278297     0.14322382     0.103153461
#> [249,]  0.928329529     0.36176463     0.019659147
#> [250,]  0.881761803     0.44140394     0.145999330
#> [251,]  0.034989750     0.52983970     0.153066058
#> [252,]  0.372981842     0.24397781     0.028880257
#> [253,]  0.001738448     0.47270212     0.115496685
#> [254,]  1.594097549     0.12706923     0.060162931
#> [255,] -0.943470300     0.26878157     0.303337382
#> [256,] -0.064150175     0.20257258     0.661379615
#> [257,] -2.406710137     0.26373075     0.154798688
#> [258,] -1.375427066     0.72520149     0.004296935
#> [259,] -0.444537618     0.22420260     0.442774801
#> [260,] -0.320302709     0.44232144     0.627672597
#> [261,]  0.577471274     0.53573339    -0.079694321
#> [262,]  0.875653685     0.13459978     0.619183004
#> [263,] -1.224229167     0.22834673     0.265700749
#> [264,]  0.179502445     0.39636901     0.273408485
#> [265,]  0.145749890     0.74394766     0.066439166
#> [266,] -0.123671673     0.42466676     0.226993188
#> [267,]  0.656359007     0.84648887     0.099905349
#> [268,] -0.666505026     0.11901684    -0.042691275
#> [269,] -0.234987585     0.30003168     0.109183695
#> [270,] -0.117432252     0.49926947     0.324959472
#> [271,] -0.750586962     0.45879643    -0.050908859
#> [272,] -0.545691182     0.71025296    -0.134115224
#> [273,] -0.719667640     0.62842614     0.485622168
#> [274,]  1.777730401     0.52839902     0.240443439
#> [275,]  0.699677502     0.33444820     0.251074046
#> [276,]  0.134244133     0.21024458     0.333995772
#> [277,]  0.927204360     1.05752541     0.171083605
#> [278,] -0.225148034     0.16689274     0.243376051
#> [279,]  0.375735681     0.55059759     0.398881358
#> [280,]  0.527849501     1.00774188    -0.050685969
#> [281,] -0.249953576     0.21787517     0.142869821
#> [282,]  0.042066813     0.31148695     0.450613394
#> [283,] -0.182098982     0.24773301     0.411645912
#> [284,] -0.260342730     0.37623098     0.124048963
#> [285,]  1.991502077     0.61868700     0.384144387
#> [286,] -0.117835224     0.59509656     0.502241659
#> [287,] -0.901649918     0.35163806     0.502728108
#> [288,]  1.708038292     0.63928209     0.108445418
#> [289,] -1.234392271     0.30645641     0.225962035
#> [290,]  0.662047556     0.25692856     0.163798150
#> [291,]  1.226646632     0.29173048     0.084265709
#> [292,]  0.637711335     0.32071012     0.575043065
#> [293,] -1.474672099     0.43881332     0.450339995
#> [294,]  0.359337807     0.28565664     0.161086098
#> [295,] -0.032743086     0.67641688     0.041821016
#> [296,] -0.240801794     0.14729151     0.509186558
#> [297,] -1.161160162     0.89969179     0.007400689
#> [298,]  1.526037336     0.97291833     0.049658384
#> [299,] -0.424742001     0.19928250     0.302873832
#> [300,] -1.025572218     0.40614159    -0.125540208
#> 
#> $acov_order
#> [1] 0
#> 
#> $acor_order
#> [1] 1
#> 
#> $N
#> [1] 300
#> 
#> $S
#> [1] 50
#> 
#> $R
#> [1] 1000
#>