The HPJ bias-corrected estimation of the moments

The `hpjmoment()` function enables to implement the HPJ bias-corrected estimation of the moments of the heterogeneous mean, the heterogeneous autocovariance, and the heterogeneous autocorrelation. The method is developed by Okui and Yanagi (2019). For more details, see the package vignette with `vignette("panelhetero")`.

Usage

hpjmoment(data, acov_order = 0, acor_order = 1, R = 1000)

Arguments

data

A matrix of panel data. Each row corresponds to individual time series.

acov_order

A non-negative integer of the order of autocovariance. Default is 0.

acor_order

A positive integer of the order of autocorrelation Default is 1.

R

A positive integer of the number of bootstrap repetitions. Default is 1000.

Value

A list that contains the following elements.

estimate

A vector of the parameter estimates

se

A vector of the standard errors

ci

A matrix of the 95 percent confidence intervals

quantity

A matrix of the estimated heterogeneous quantities

acov_order

The order of autocovariance

acor_order

The order of autocovariance

N

The number of cross-sectional units

S

The length of time series

R

The number of bootstrap repetitions

References

Okui, R. and Yanagi, T., 2019. Panel data analysis with heterogeneous dynamics. Journal of Econometrics, 212(2), pp.451-475.

Examples

data <- panelhetero::simulation(N = 300, S = 50)
panelhetero::hpjmoment(data = data)
#> $estimate
#>         E(mean)         E(acov)         E(acor)       var(mean)       var(acov) 
#>      0.00339141      0.35427880      0.29146545      1.01960599      0.03815814 
#>       var(acor) cor(mean, acov) cor(mean, acor) cor(acov, acor) 
#>      0.02411234     -0.05074737      0.03787705     -0.52491169 
#> 
#> $se
#>         E(mean)         E(acov)         E(acor)       var(mean)       var(acov) 
#>     0.056970099     0.012360734     0.012330468     0.082454735     0.005482713 
#>       var(acor) cor(mean, acov) cor(mean, acor) cor(acov, acor) 
#>     0.003824130     0.058874114     0.074397085     0.065618122 
#> 
#> $ci
#>                 95% CI lower 95% CI upper
#> E(mean)          -0.11209373   0.10741573
#> E(acov)           0.33258489   0.38043724
#> E(acor)           0.26710876   0.31526731
#> var(mean)         0.85592402   1.17061115
#> var(acov)         0.02815255   0.04977702
#> var(acor)         0.01686956   0.03189214
#> cor(mean, acov)  -0.17584832   0.05495548
#> cor(mean, acor)  -0.09783403   0.18841217
#> cor(acov, acor)  -0.65075715  -0.39613505
#> 
#> $quantity
#>                 mean autocovariance autocorrelation
#>   [1,] -1.6638107225     0.48676263     0.025502293
#>   [2,] -1.0055950398     0.30405463     0.269206005
#>   [3,]  0.0462899741     0.08495028     0.521149299
#>   [4,] -0.2458471447     0.49922547     0.108199735
#>   [5,] -0.9451270123     0.26511047     0.051544516
#>   [6,]  1.5405714119     0.14228549     0.259002357
#>   [7,] -1.4080304174     0.37842870     0.175659529
#>   [8,]  0.0706790308     0.16720554     0.475992094
#>   [9,] -0.7940863233     0.08308324     0.469623519
#>  [10,]  0.1662822513     0.20869488     0.093852797
#>  [11,] -0.2575395756     0.10006176     0.737707681
#>  [12,]  0.1436873824     0.47466890     0.452639936
#>  [13,] -2.1071105338     0.29017460     0.656453384
#>  [14,]  0.1367803418     0.28252693     0.051381669
#>  [15,] -0.0215962382     0.22781529     0.171395991
#>  [16,] -0.0041868104     0.24136529     0.175719796
#>  [17,]  0.2362218643     0.23047702    -0.140631474
#>  [18,] -0.8008250831     0.14879103     0.404745864
#>  [19,]  0.1508829894     0.82935128     0.253830778
#>  [20,]  1.2612623410     0.22605061     0.150712712
#>  [21,]  0.0581525549     0.28065550     0.334421349
#>  [22,]  0.4274887189     0.15807168     0.423381482
#>  [23,]  2.2290844930     0.58111206     0.169780593
#>  [24,]  1.4470919743     0.25847716     0.328364800
#>  [25,] -0.4218850384     0.18602824     0.305646918
#>  [26,]  1.9883150085     0.18736723     0.178865485
#>  [27,] -0.6345757963     0.41432964     0.236005424
#>  [28,]  1.6827623540     0.65856085     0.157087808
#>  [29,]  0.8263082291     0.45886819    -0.275471483
#>  [30,] -1.9444922738     0.57973064    -0.101386259
#>  [31,] -1.3077815900     0.28582064     0.038407163
#>  [32,]  0.5513037110     0.52085551    -0.101612364
#>  [33,] -0.3965586387     0.52201632     0.207800594
#>  [34,] -0.2117495525     0.14324315     0.441348778
#>  [35,]  0.6090089933     0.42216092     0.199281726
#>  [36,]  1.2233042219     0.39269841     0.134229782
#>  [37,] -1.5645622222     0.98204108    -0.057964091
#>  [38,]  0.1183195003     0.25534964     0.084773835
#>  [39,]  0.2720441700     0.26011987     0.493613806
#>  [40,]  1.6253228548     0.44313486     0.353045449
#>  [41,]  0.5981332469     0.28660584     0.248515213
#>  [42,] -1.2790298498     0.29977109     0.654522914
#>  [43,] -0.3101149098     0.29697906     0.412311424
#>  [44,] -0.9817489490     0.84392165     0.251514979
#>  [45,] -1.0970192843     0.27632154     0.164530339
#>  [46,]  1.2710429890     0.36700876     0.136228121
#>  [47,]  0.3005348464     0.25107051     0.089647305
#>  [48,]  0.8056207335     0.34402014     0.405309669
#>  [49,] -1.2776431644     0.22273697     0.224290373
#>  [50,]  0.0450937492     0.25678097     0.227372070
#>  [51,] -0.8090473118     0.14812763     0.283636756
#>  [52,] -1.4888467900     0.96361070     0.196302258
#>  [53,]  0.0662846842     0.23914272     0.376879102
#>  [54,]  0.5443663126     0.49347922     0.455179622
#>  [55,]  0.7153262231     0.52818262     0.067354803
#>  [56,] -0.1434511201     0.46431222     0.104846549
#>  [57,]  2.4299485844     0.32699816     0.481796397
#>  [58,] -1.3724882452     0.36780029     0.088856436
#>  [59,] -0.8851817436     0.41554653     0.091668412
#>  [60,] -2.7296892931     0.17422963     0.580350210
#>  [61,] -1.7955671713     0.28115052     0.345204761
#>  [62,]  0.1011585334     0.33134810     0.213092635
#>  [63,]  0.9515494178     0.83159126    -0.221767280
#>  [64,]  0.1205808487     0.41266546     0.426258756
#>  [65,]  0.7637743405     0.23834243     0.297860406
#>  [66,] -0.2783244615     0.15193800     0.278890104
#>  [67,] -1.0389912927     0.34831693     0.057121029
#>  [68,]  1.1017502003     0.93500803     0.280448794
#>  [69,]  1.3787066525     0.76170714     0.145204440
#>  [70,] -0.5456707787     0.71249497     0.152896920
#>  [71,] -2.2891112780     0.17344765     0.129735775
#>  [72,] -0.2747854625     0.15609809     0.452831149
#>  [73,]  0.7887137640     0.23536904     0.353074197
#>  [74,]  1.1282605968     0.48190704     0.110850322
#>  [75,]  0.0173379164     0.14296489     0.120815157
#>  [76,]  0.2982185428     0.10126258     0.395746980
#>  [77,] -0.8090611685     0.09126758     0.431955552
#>  [78,] -0.4322917144     0.09002672     0.453747272
#>  [79,] -0.6523180819     0.35384299     0.078777188
#>  [80,]  2.2214578475     0.34019424     0.221785135
#>  [81,]  0.4946312826     0.21945305     0.289872728
#>  [82,] -0.1923438777     0.14450328    -0.094553440
#>  [83,] -0.2820212248     0.10854380     0.415724342
#>  [84,]  0.3684846483     0.43470457    -0.016160288
#>  [85,]  1.6944744479     0.21759015     0.537168485
#>  [86,] -1.0670220494     0.49533235    -0.036009982
#>  [87,] -0.7957052063     0.36847153     0.225208099
#>  [88,]  0.2434007383     0.31403624    -0.079528370
#>  [89,]  0.2099072963     0.24601500     0.435222131
#>  [90,] -0.5387370421     0.34575547     0.096621413
#>  [91,]  0.5493626559     0.34736159     0.285389892
#>  [92,] -0.4611707509     0.21871366     0.429782704
#>  [93,]  0.6353565730     0.51493868     0.228808470
#>  [94,] -1.2518613613     0.13215689     0.380977744
#>  [95,] -0.8205438767     0.12773145     0.548280349
#>  [96,] -0.5453391553     0.16592424     0.542361380
#>  [97,] -0.3341136358     0.38403695     0.533735557
#>  [98,] -0.7805821920     1.08222194    -0.104582358
#>  [99,]  0.8593063306     0.16958893     0.290123663
#> [100,]  0.0043574779     0.61779939     0.256682550
#> [101,] -1.6637964225     0.08874371     0.353899312
#> [102,]  0.7227688233     0.33156086     0.199580922
#> [103,] -0.7167262736     0.36877588     0.177804352
#> [104,]  2.0972585179     0.24012265     0.677954310
#> [105,] -0.4490483743     0.64293521     0.442162735
#> [106,] -0.5057237824     0.12142480     0.433363029
#> [107,] -1.6434661302     0.47107711     0.046561014
#> [108,]  0.6423459233     0.30288322     0.596607249
#> [109,] -0.1987874366     0.72164699     0.049260208
#> [110,]  0.9912116946     0.45901661     0.546364534
#> [111,] -0.0457586208     0.26276818     0.289208801
#> [112,]  1.7457173007     0.50057672     0.119149110
#> [113,] -0.9379334501     0.36249405     0.362859970
#> [114,] -2.5968314724     0.24200483     0.394693112
#> [115,]  0.4155082683     0.42908453     0.328381717
#> [116,]  0.0003531599     0.32673165     0.635654387
#> [117,]  0.0598018829     0.37087135     0.045000414
#> [118,]  1.0300855594     0.84999995     0.086852586
#> [119,] -1.4419424220     0.41776330     0.175158552
#> [120,] -0.1619431498     0.28012907     0.434028787
#> [121,] -0.4035386787     0.56932201    -0.169442366
#> [122,]  1.1599973725     0.18737284     0.361786945
#> [123,]  0.2478259761     1.08032621    -0.166824634
#> [124,] -1.3507696171     0.45264211     0.405164997
#> [125,] -0.0081289732     0.31486341     0.383268303
#> [126,] -0.7416569610     0.54152424     0.071662483
#> [127,]  1.8662403691     0.22180593     0.538605639
#> [128,]  0.0566437423     0.24836397     0.113686296
#> [129,]  1.1429421262     0.34862327     0.513746512
#> [130,] -1.3127406302     0.08196184     0.389220089
#> [131,] -0.8722486434     0.22423975     0.466394306
#> [132,]  1.4865747950     0.19360970    -0.098972074
#> [133,] -0.2005760870     0.21065059     0.252961495
#> [134,]  0.0548550657     0.79705746     0.238969832
#> [135,] -0.4333642663     0.21118906     0.375810326
#> [136,]  0.3100080383     0.23764545     0.307173528
#> [137,]  0.0759695609     0.36226781     0.035453197
#> [138,]  0.0443698037     0.35438757     0.105230538
#> [139,] -0.0243921509     0.25969739     0.169029984
#> [140,]  0.5335964263     0.12027498     0.288097224
#> [141,] -0.6344521226     0.29533943     0.304883142
#> [142,] -0.2409895926     0.34385474     0.075685204
#> [143,]  0.1235055809     0.36986127     0.012753358
#> [144,]  0.5427179745     0.68037680     0.224171362
#> [145,]  0.3849794340     0.19702272     0.100386231
#> [146,] -0.6786618204     0.26279991     0.142931850
#> [147,] -0.8500215494     0.25062488     0.508600267
#> [148,]  0.8197302886     0.15841359     0.431710820
#> [149,] -0.7088002516     0.31872042     0.215215552
#> [150,]  2.1834257246     0.26407755     0.123688260
#> [151,] -0.7713472538     0.67828068     0.159090375
#> [152,]  0.5685373971     0.08779830     0.286217952
#> [153,]  0.7310657049     0.23839209     0.268525931
#> [154,]  0.4395127293     0.16278278     0.221445812
#> [155,]  1.8711201699     0.10016119     0.430948083
#> [156,]  0.9217182414     0.25633661     0.146869094
#> [157,]  0.6724415713     0.23799601     0.425105616
#> [158,] -0.5478389746     0.46896393     0.383682645
#> [159,]  0.1538362053     0.27539875     0.339309075
#> [160,]  0.5422574210     0.17975942     0.205463076
#> [161,] -0.1775242955     0.38891748    -0.063207541
#> [162,]  0.0346149038     0.23632201     0.158913960
#> [163,]  0.6137687795     0.47562076     0.230067977
#> [164,] -0.2511513663     0.16093271     0.354232842
#> [165,] -0.5751106264     0.56378366    -0.108747024
#> [166,]  0.2576210316     0.51380700     0.223049355
#> [167,] -0.9243426856     0.50152565     0.077759896
#> [168,]  0.0370077842     0.30868192     0.083908081
#> [169,] -1.0485920434     0.74010757     0.131135265
#> [170,]  0.8406406476     0.28792902     0.378899272
#> [171,] -1.8326958220     0.43424278     0.478920217
#> [172,]  1.2454436010     0.22700295     0.057141969
#> [173,] -0.2746725585     0.15132631     0.417618789
#> [174,] -1.9437093677     0.23275356     0.305025464
#> [175,] -1.3543891352     0.23104701     0.477987989
#> [176,] -0.5490213126     0.18912164     0.299258082
#> [177,] -1.5717431031     0.09690105     0.507714305
#> [178,]  0.7924073696     0.24618718     0.466035607
#> [179,]  0.4057097505     0.45399675     0.297625836
#> [180,] -0.5217577098     0.43189933     0.311060834
#> [181,]  0.5502993548     0.26784082     0.035767168
#> [182,] -1.4295823478     0.37416609     0.188914559
#> [183,]  0.8115960547     0.24440051     0.081660305
#> [184,] -2.6920998967     0.54879389     0.020393311
#> [185,]  0.5613011942     0.39251106     0.537303648
#> [186,]  0.1954358632     0.32995965     0.177704500
#> [187,]  2.5577639958     0.37245246     0.197112369
#> [188,]  1.9162102697     0.07213942     0.278707329
#> [189,]  0.9189269096     0.16326032     0.637297541
#> [190,] -0.7978503969     0.05823777     0.819713012
#> [191,]  1.2285742442     0.29034202     0.158446397
#> [192,]  0.2576755678     0.14076288     0.496811005
#> [193,] -1.3320131019     0.60720307     0.149569166
#> [194,]  0.9776169480     0.23936163     0.334121205
#> [195,]  0.4586216406     0.24968909     0.294575809
#> [196,]  0.6193914357     0.52442468    -0.119138707
#> [197,] -0.3104426289     1.40961758     0.297723368
#> [198,]  0.0887950159     0.32876619     0.377729507
#> [199,]  0.1271582953     0.38079775     0.465296713
#> [200,]  0.6096225232     0.41232808     0.071435371
#> [201,] -0.0144036507     0.10028022     0.264505511
#> [202,]  1.1399552752     0.11186915     0.577625696
#> [203,]  0.9981049421     0.44631152     0.297597864
#> [204,] -0.3435989426     0.27989024     0.353223760
#> [205,] -0.4658034384     0.60110380    -0.113609280
#> [206,]  0.6128339154     0.55934978     0.369382069
#> [207,] -0.0961168535     0.24973373     0.474927183
#> [208,]  1.7397626813     0.22338125     0.563179186
#> [209,] -0.1161043154     0.52053384    -0.171931793
#> [210,] -0.1714603559     0.37086263     0.409119107
#> [211,] -1.0541993536     0.25079615     0.256911689
#> [212,]  0.6357950065     0.27386763     0.357957350
#> [213,] -1.7147419564     0.23276594     0.619342145
#> [214,] -1.5947444232     0.40130014     0.344348705
#> [215,]  0.1048139363     0.17463786     0.387162358
#> [216,]  0.8200945426     0.18804848     0.528000111
#> [217,]  0.2347437543     0.31439407     0.112751017
#> [218,] -0.7374579337     0.45101644     0.098358630
#> [219,] -1.6665849966     0.21488350     0.279873263
#> [220,]  0.5328754419     0.51371699     0.117514767
#> [221,]  0.6796598630     0.33929381     0.380721926
#> [222,]  0.5713016162     0.40701437     0.042581723
#> [223,]  1.3508407031     0.15472614     0.051370475
#> [224,] -1.7082516130     0.38594983     0.048668978
#> [225,]  0.3894811250     0.12539809     0.285768376
#> [226,] -0.8436216710     0.27875041     0.498968307
#> [227,] -1.6326492793     0.39790720     0.145840968
#> [228,] -0.6255931218     0.61724332     0.178539939
#> [229,] -1.4943006052     0.53124647     0.494666067
#> [230,] -1.4325678625     0.39985210     0.065465759
#> [231,]  0.1046063355     0.34456641     0.226008443
#> [232,]  0.4422465438     0.37107303    -0.007497668
#> [233,]  2.5048321884     0.35668003     0.576800721
#> [234,] -0.5450300533     0.18827027     0.206672316
#> [235,]  0.6680784941     0.03729189     0.471930538
#> [236,] -1.3285810202     0.47106462     0.004895701
#> [237,]  0.6738113656     0.54106551     0.260746051
#> [238,] -0.4365563760     0.20012328     0.124678701
#> [239,] -0.0116536968     0.70017717     0.279027968
#> [240,]  0.5602698171     0.13482670     0.501195404
#> [241,] -0.5424345853     0.28650280     0.257130371
#> [242,] -0.1358457010     0.37085592     0.400325747
#> [243,]  0.1715052040     0.25270629     0.305900034
#> [244,]  0.3615003957     0.17189788     0.174973975
#> [245,] -0.2927466678     0.15206603    -0.075726251
#> [246,] -1.5131456880     0.36818381     0.281747633
#> [247,] -0.3802361869     0.13178866     0.044225713
#> [248,]  2.0592830898     0.19359928     0.140250456
#> [249,]  0.7797889576     0.23221885     0.540139754
#> [250,] -0.2897650129     0.32790992     0.029467336
#> [251,] -1.5059194574     0.49512232    -0.013625408
#> [252,] -0.1865358822     0.23126628     0.189879885
#> [253,]  1.5839870379     0.15669002     0.334128584
#> [254,] -0.6139380198     0.38261104     0.010570525
#> [255,]  1.3494046633     0.11485492     0.401744493
#> [256,] -0.5436367502     0.92540064     0.429079974
#> [257,]  0.3813646836     0.15938256     0.577062352
#> [258,]  0.2981513951     0.72759230    -0.017957311
#> [259,]  0.1408957687     0.60777084    -0.257334038
#> [260,]  0.6551422099     0.82282260     0.169970771
#> [261,] -0.1633345905     0.11199363     0.451752072
#> [262,]  1.7772738656     0.69356696     0.254704158
#> [263,]  0.2845905786     0.05190881     0.447907953
#> [264,] -0.2656798696     0.12439939     0.277883034
#> [265,]  0.6585659922     0.90758909     0.401680778
#> [266,] -0.3069447394     0.70024166     0.147648303
#> [267,]  0.4367607907     0.58743322     0.253792154
#> [268,]  0.4621132373     0.27820538     0.229954222
#> [269,] -0.7094518846     0.16349410     0.250112167
#> [270,] -1.8398877138     0.27822958     0.323264036
#> [271,]  0.0762233411     0.17109576    -0.026448618
#> [272,]  0.0394959777     0.16468888     0.002691012
#> [273,]  0.5416370605     0.17267045     0.461062204
#> [274,] -0.9847692043     0.33638527     0.249625320
#> [275,] -1.3460644331     0.08016307     0.264409011
#> [276,] -0.8180387175     0.15236523     0.477581549
#> [277,] -1.1033519522     0.33814113     0.009232393
#> [278,] -0.2199274714     0.26601366     0.393851177
#> [279,] -0.7366422604     0.17418419     0.264407801
#> [280,]  0.1212896923     0.69007750     0.091129966
#> [281,]  1.2647564721     0.18838612     0.366973113
#> [282,]  0.3736463043     0.16459113     0.042000124
#> [283,] -0.5622744763     0.55422962     0.263010364
#> [284,] -0.4254669051     0.38858068     0.368792133
#> [285,]  0.3385798125     0.32023647    -0.306354293
#> [286,]  1.3198689127     0.12766210     0.492489433
#> [287,] -0.3383369351     0.24280143     0.185197284
#> [288,]  0.5143099837     0.12170967     0.484738680
#> [289,] -0.8706164616     0.31744959     0.453512925
#> [290,]  0.3386194820     0.40034119    -0.025518900
#> [291,] -2.2007635170     0.13823346     0.472041617
#> [292,]  2.7429736765     0.10056333     0.306022832
#> [293,] -0.1727718050     0.20588193     0.319234958
#> [294,]  1.3569747374     0.34336874     0.426872299
#> [295,]  0.0802680742     0.62921211    -0.151910099
#> [296,] -0.1861503077     0.42421776     0.408086132
#> [297,]  1.2699658446     0.14490197     0.301359489
#> [298,]  1.2414673073     0.54206009     0.117523022
#> [299,]  2.4376446564     0.23924519     0.105155273
#> [300,]  0.8438989458     0.56557915     0.617554010
#> 
#> $acov_order
#> [1] 0
#> 
#> $acor_order
#> [1] 1
#> 
#> $N
#> [1] 300
#> 
#> $S
#> [1] 50
#> 
#> $R
#> [1] 1000
#>