The naive estimation of the moments

The `nemoment()` function enables to implement the naive estimation of the moments of the heterogeneous mean, the heterogeneous autocovariance, and the heterogeneous autocorrelation. The method is developed by Okui and Yanagi (2019). For more details, see the package vignette with `vignette("panelhetero")`.

Usage

nemoment(data, acov_order = 0, acor_order = 1, R = 1000)

Arguments

data

A matrix of panel data. Each row corresponds to individual time series.

acov_order

A non-negative integer of the order of autocovariance. Default is 0.

acor_order

A positive integer of the order of autocorrelation Default is 1.

R

A positive integer of the number of bootstrap repetitions. Default is 1000.

Value

A list that contains the following elements.

estimate

A vector of the parameter estimates

se

A vector of the standard errors

ci

A matrix of the 95 percent confidence intervals

quantity

A matrix of the estimated heterogeneous quantities

acov_order

The order of autocovariance

acor_order

The order of autocovariance

N

The number of cross-sectional units

S

The length of time series

R

The number of bootstrap repetitions

References

Okui, R. and Yanagi, T., 2019. Panel data analysis with heterogeneous dynamics. Journal of Econometrics, 212(2), pp.451-475.

Examples

data <- panelhetero::simulation(N = 300, S = 50)
panelhetero::nemoment(data = data)
#> $estimate
#>         E(mean)         E(acov)         E(acor)       var(mean)       var(acov) 
#>    -0.003855561     0.382177323     0.258490097     0.893982046     0.057133751 
#>       var(acor) cor(mean, acov) cor(mean, acor) cor(acov, acor) 
#>     0.041311950     0.027614914     0.035162825    -0.376468949 
#> 
#> $se
#>         E(mean)         E(acov)         E(acor)       var(mean)       var(acov) 
#>     0.054042831     0.014227635     0.011297178     0.067577896     0.008783777 
#>       var(acor) cor(mean, acov) cor(mean, acor) cor(acov, acor) 
#>     0.002935730     0.058785569     0.056213165     0.046683478 
#> 
#> $ci
#>                 95% CI lower 95% CI upper
#> E(mean)          -0.10914957   0.10159668
#> E(acov)           0.35579566   0.41137528
#> E(acor)           0.23520001   0.28103955
#> var(mean)         0.76203169   1.02454733
#> var(acov)         0.04267148   0.07576983
#> var(acor)         0.03574814   0.04695457
#> cor(mean, acov)  -0.08864266   0.15054433
#> cor(mean, acor)  -0.07328577   0.14787611
#> cor(acov, acor)  -0.46843059  -0.28411349
#> 
#> $quantity
#>                 mean autocovariance autocorrelation
#>   [1,]  1.558366e-01     0.62709569     0.029047906
#>   [2,] -8.072274e-01     0.71335224     0.138682107
#>   [3,] -2.617220e-02     0.38420154     0.135542393
#>   [4,] -2.194424e-01     0.15591783     0.493526658
#>   [5,]  1.130141e-01     0.10916517     0.646885452
#>   [6,]  8.302135e-01     0.30357438     0.395569591
#>   [7,] -6.817251e-01     0.28055773     0.236441589
#>   [8,] -2.918529e-01     0.37036245     0.008811425
#>   [9,]  1.261838e+00     0.21028133     0.497617997
#>  [10,] -3.659846e-01     0.29417704     0.062288035
#>  [11,] -1.619526e-01     0.73835887     0.197300170
#>  [12,] -1.674121e+00     0.13591781     0.057847336
#>  [13,] -1.503178e+00     0.57252238     0.028143059
#>  [14,] -1.703660e-01     0.12560746     0.271597128
#>  [15,] -3.280297e-01     0.22556512     0.098055230
#>  [16,] -1.089279e+00     0.08052159     0.515837922
#>  [17,] -7.531249e-01     0.04980539     0.482922584
#>  [18,]  8.903328e-01     0.30696423    -0.031379487
#>  [19,]  1.055559e+00     0.26662633     0.425683620
#>  [20,]  2.850905e-01     0.12990029     0.516698060
#>  [21,] -3.020065e-01     0.21965989     0.274340577
#>  [22,] -6.297730e-01     0.18747102    -0.112161548
#>  [23,] -3.075905e-02     0.56081920     0.369623556
#>  [24,]  1.673804e+00     0.15796717     0.631616894
#>  [25,]  1.381120e+00     0.61740954     0.080659085
#>  [26,]  5.811285e-01     0.11085435     0.077032357
#>  [27,] -1.327603e+00     0.20200855     0.340481724
#>  [28,]  8.674514e-01     0.61805181     0.211944377
#>  [29,]  2.552660e-03     0.31251890     0.124262322
#>  [30,] -1.665801e-01     0.44742725     0.155187829
#>  [31,] -1.369462e-02     0.19830747     0.217315090
#>  [32,] -9.412341e-01     0.24572414     0.549770661
#>  [33,] -1.530206e+00     0.38682703    -0.013412480
#>  [34,] -2.644105e-01     0.41778221     0.076481192
#>  [35,] -1.052344e+00     0.31756847     0.154616647
#>  [36,]  1.250844e+00     0.68066909     0.445113599
#>  [37,] -4.178985e-01     0.65918907     0.212929956
#>  [38,]  4.259961e-01     0.32785590     0.147926137
#>  [39,]  3.092511e-01     0.34602727     0.363368553
#>  [40,]  1.797458e+00     0.57222224     0.357775114
#>  [41,] -4.803069e-01     0.14109273     0.247949219
#>  [42,]  6.009569e-01     0.37483446     0.403736266
#>  [43,]  6.991602e-01     1.34425528     0.301755150
#>  [44,] -1.296492e+00     0.81596079     0.078641297
#>  [45,]  4.850332e-01     0.35503405     0.132868438
#>  [46,] -4.197529e-01     0.73241065    -0.001376381
#>  [47,]  5.784604e-01     0.37861561     0.352548847
#>  [48,]  2.922022e-01     0.23850935     0.325295946
#>  [49,] -4.180711e-01     0.28483840     0.317587104
#>  [50,] -1.325944e+00     0.79554012    -0.101778394
#>  [51,]  1.013332e+00     0.10980149     0.302826236
#>  [52,]  1.451832e+00     0.13926718     0.664587300
#>  [53,] -5.853587e-01     0.15001279     0.421130339
#>  [54,]  9.354759e-01     0.44602488     0.257870583
#>  [55,] -5.700189e-01     0.12936865     0.348709596
#>  [56,] -7.029753e-01     0.16987459     0.253100842
#>  [57,]  1.913747e+00     0.56384174     0.229896295
#>  [58,] -2.809645e-01     0.49277861     0.147037547
#>  [59,] -2.439048e-01     0.37145095     0.196689712
#>  [60,] -9.873373e-01     0.61920595    -0.208857127
#>  [61,]  6.698681e-01     0.46189181    -0.027774710
#>  [62,]  1.146083e+00     0.32379720     0.305152671
#>  [63,] -5.599235e-01     0.31428419    -0.063001843
#>  [64,] -1.675265e-01     0.23596305     0.180691395
#>  [65,]  6.574262e-02     0.32006620     0.009703070
#>  [66,] -6.538999e-01     0.39010167     0.328016967
#>  [67,]  6.933754e-01     0.27042254     0.290398427
#>  [68,] -1.910935e+00     0.16667905     0.045769550
#>  [69,]  5.769216e-01     0.18723144     0.159694826
#>  [70,]  1.306150e-01     0.79837459     0.275325236
#>  [71,] -4.298655e-01     0.60890055     0.009823536
#>  [72,]  5.699671e-01     0.31956622     0.706953752
#>  [73,] -1.309810e-01     0.07754788     0.523771998
#>  [74,]  5.674998e-01     0.51862674     0.309377367
#>  [75,] -5.754852e-01     0.23429719     0.583640499
#>  [76,]  1.357873e+00     0.34034377     0.727394864
#>  [77,] -6.394551e-01     0.42398896     0.676447841
#>  [78,] -2.343853e-01     0.20376622     0.406904588
#>  [79,] -1.277273e+00     0.27570837     0.280281724
#>  [80,] -6.713997e-01     0.28555013     0.615063662
#>  [81,] -1.724321e+00     0.23898378     0.576770031
#>  [82,] -2.272092e+00     0.18441548     0.514896768
#>  [83,]  9.010772e-01     0.07671849     0.161799010
#>  [84,] -4.308950e-01     0.62891549     0.343337265
#>  [85,]  8.417009e-05     0.23600809     0.266867660
#>  [86,]  9.630579e-01     0.33304913    -0.024563690
#>  [87,] -1.134028e+00     0.15603999     0.081077059
#>  [88,] -8.926693e-02     0.14578525     0.395004133
#>  [89,] -6.125062e-01     0.12688321     0.574148271
#>  [90,]  8.184750e-01     0.33336221     0.333678099
#>  [91,]  5.855587e-01     0.14814989     0.436937450
#>  [92,] -6.334033e-01     0.09603176     0.322390581
#>  [93,]  9.680517e-01     0.31990822     0.071605771
#>  [94,]  1.655740e-01     0.17283567     0.210278173
#>  [95,]  2.948485e-01     0.42446355     0.596273551
#>  [96,]  1.205516e+00     0.06485523     0.202891065
#>  [97,]  1.815498e+00     0.46208360     0.283887298
#>  [98,]  5.058863e-01     0.30005302     0.320508535
#>  [99,] -3.361032e-02     0.24280954     0.400195324
#> [100,] -6.101436e-01     0.74588127     0.281638282
#> [101,]  5.179075e-01     0.35613782    -0.073615105
#> [102,]  9.067495e-01     0.30793056     0.001637857
#> [103,] -8.376254e-01     0.41143137     0.486195823
#> [104,] -1.094050e+00     0.64918124     0.056462766
#> [105,] -3.452688e-01     0.48508403     0.466084254
#> [106,] -2.629923e-01     0.30500270     0.188936447
#> [107,]  8.439278e-01     0.42596645     0.462317825
#> [108,]  1.239821e+00     0.35000425     0.328194981
#> [109,] -9.993608e-01     0.30592195     0.621040758
#> [110,]  2.629099e-01     0.22256628     0.258475583
#> [111,] -1.673470e+00     0.14161952     0.346703162
#> [112,] -7.263826e-02     0.22570806     0.296777100
#> [113,] -4.539150e-01     0.29258950     0.119663223
#> [114,]  1.724756e+00     0.34335214     0.112291073
#> [115,] -2.131513e+00     0.51800106     0.174511214
#> [116,] -1.293816e-01     0.07549255     0.657728854
#> [117,]  9.007506e-01     0.88104408     0.261846711
#> [118,]  6.528337e-01     0.52140636     0.263133096
#> [119,]  1.470397e+00     0.29616020     0.311973714
#> [120,] -1.393154e+00     0.47505762     0.022467980
#> [121,]  7.949888e-01     0.50631198    -0.002400374
#> [122,] -4.184411e-01     0.13177719     0.666382899
#> [123,] -1.488110e+00     0.57485822     0.071120848
#> [124,]  1.042498e+00     0.18086796     0.096187488
#> [125,] -4.419382e-01     0.65870672     0.414064989
#> [126,] -4.938833e-01     0.36845303    -0.052849327
#> [127,]  1.195205e+00     0.47209868     0.464860737
#> [128,]  9.275874e-02     0.42164044     0.377721366
#> [129,] -6.906993e-01     0.34800373     0.220786339
#> [130,] -5.834773e-01     0.49627242     0.120798462
#> [131,]  1.313463e+00     0.25902677     0.327456594
#> [132,]  1.082059e+00     0.10015990     0.595573374
#> [133,]  1.835016e-01     0.43843023     0.124859422
#> [134,] -7.306101e-01     0.15282404     0.384783940
#> [135,] -1.029483e+00     0.15228861     0.139254613
#> [136,]  1.411173e+00     0.55580962    -0.151218876
#> [137,] -6.206564e-02     0.46159085     0.020435310
#> [138,]  4.860275e-01     0.57995348     0.108159961
#> [139,]  1.760634e+00     0.36376690     0.154399650
#> [140,]  2.455065e-01     1.13238646    -0.085366224
#> [141,] -6.209506e-01     0.12563346    -0.156679236
#> [142,] -6.545946e-01     1.10978730    -0.268539937
#> [143,]  1.244892e+00     0.32947352     0.078019489
#> [144,] -4.192010e-01     0.36022950     0.234595309
#> [145,]  8.924442e-02     0.23828613     0.210558065
#> [146,] -1.147875e+00     0.76960766     0.444230301
#> [147,]  2.093473e-01     0.21515314     0.448613814
#> [148,]  1.898216e+00     0.40573150     0.011436367
#> [149,] -4.776684e-01     0.41816687     0.465912524
#> [150,]  9.653499e-01     0.68877201     0.220685607
#> [151,] -1.269398e+00     0.36727091    -0.126105176
#> [152,] -1.185523e+00     0.11238411     0.388545574
#> [153,]  4.937618e-01     0.37725054     0.270763587
#> [154,]  7.792652e-01     0.24596075     0.637666643
#> [155,] -1.109112e+00     0.27969008     0.379391771
#> [156,] -1.552337e+00     0.40082236     0.273880568
#> [157,]  9.013971e-01     0.48624951     0.104549928
#> [158,] -1.327889e+00     0.44304278     0.234794568
#> [159,] -1.233007e-01     0.88492876     0.262388941
#> [160,]  9.782452e-01     0.24498745    -0.022708310
#> [161,]  1.692924e+00     0.26621319    -0.047711522
#> [162,] -1.112663e+00     0.05391517     0.562025603
#> [163,] -1.455375e-01     0.25901289     0.349313033
#> [164,] -1.503276e+00     0.28754867     0.231838211
#> [165,] -8.528601e-01     0.16292950     0.592787701
#> [166,]  1.505176e+00     0.25482316     0.354400658
#> [167,]  1.900375e-01     0.93512579     0.185647600
#> [168,] -7.276347e-01     0.35428393     0.392758120
#> [169,] -5.551836e-01     0.20102374     0.209873469
#> [170,]  7.667067e-01     0.68705599    -0.028433380
#> [171,]  7.878202e-01     0.40777781     0.542525293
#> [172,]  6.537810e-01     0.66022820    -0.056673084
#> [173,] -1.708555e-01     0.18998011     0.269993657
#> [174,]  9.417560e-01     0.67125491     0.452809842
#> [175,] -1.182025e+00     0.24499549     0.202682423
#> [176,] -1.599955e+00     0.27905051     0.275639803
#> [177,]  5.693537e-01     0.21428574     0.044728276
#> [178,] -6.809454e-01     0.45034326     0.367461615
#> [179,]  1.369494e-01     0.16118732     0.380426434
#> [180,] -2.582134e-01     0.26202577     0.112889391
#> [181,]  8.633576e-03     0.34254767     0.383920530
#> [182,]  6.421545e-01     0.30584059     0.378242606
#> [183,]  4.751024e-01     0.34919885     0.315029507
#> [184,]  3.638745e-01     0.10266982    -0.010666602
#> [185,] -9.223524e-01     0.23772052     0.231354106
#> [186,]  2.459479e-01     0.13850733     0.347413220
#> [187,] -6.319378e-01     0.48613608     0.448494016
#> [188,]  1.636160e+00     0.15888151     0.435066073
#> [189,] -1.014606e+00     0.30215684     0.250531624
#> [190,]  5.450858e-02     0.49036201     0.028594494
#> [191,] -1.168142e+00     0.29187142     0.034668831
#> [192,]  3.808651e-01     0.36740118     0.306266506
#> [193,]  2.338760e-01     0.54003471    -0.112984458
#> [194,] -5.246259e-01     0.14790747     0.445993872
#> [195,] -3.110543e-01     0.73388087    -0.263640145
#> [196,] -9.637146e-01     0.37346117     0.414193346
#> [197,] -1.257732e+00     0.96608785     0.029541751
#> [198,] -2.972553e-01     0.44799223    -0.026962336
#> [199,]  1.393030e+00     0.17614259     0.452041861
#> [200,] -7.717368e-01     0.20036882     0.439710976
#> [201,] -3.497683e-01     0.16018144     0.093318614
#> [202,] -1.234359e+00     0.48915803     0.224668643
#> [203,] -7.717769e-01     0.70554547     0.012913936
#> [204,] -7.231193e-01     0.13886957     0.395923348
#> [205,] -1.235490e+00     0.44729959     0.052997970
#> [206,] -1.413524e+00     0.29040124     0.137493217
#> [207,]  2.098626e+00     0.24776787     0.075862188
#> [208,] -1.463351e-01     0.10007721     0.647805928
#> [209,]  3.803692e-01     0.63916721    -0.077107176
#> [210,]  3.045783e-01     0.27497546     0.539885310
#> [211,]  3.145628e-01     0.26323543     0.488802961
#> [212,]  1.588342e+00     0.24011140     0.391979357
#> [213,]  3.740506e-01     0.14519304     0.225282258
#> [214,]  1.416261e+00     1.07362120     0.165538462
#> [215,]  2.834138e-01     0.30314639     0.250540892
#> [216,]  5.137352e-01     0.66131055     0.093237865
#> [217,] -1.184391e+00     0.38962874     0.365659501
#> [218,] -5.977806e-01     0.57505074     0.507473145
#> [219,] -5.760682e-01     0.86542562     0.013595359
#> [220,] -1.052580e+00     0.52867319     0.354210934
#> [221,] -1.052305e+00     0.74963161     0.219772347
#> [222,] -6.716569e-01     0.24180546     0.305081412
#> [223,]  5.198555e-01     0.04639773     0.131113139
#> [224,]  2.972193e-01     0.49272712     0.072275511
#> [225,] -1.884318e-01     0.68151837     0.383004688
#> [226,] -2.041057e-01     0.13824288     0.150108973
#> [227,] -6.700103e-01     0.17075113     0.298444707
#> [228,] -1.078705e-01     0.66883107     0.382567190
#> [229,] -4.088307e-01     0.17320653     0.607059596
#> [230,] -1.432954e-01     0.27417632     0.620716169
#> [231,] -1.312004e+00     0.26665035    -0.123386119
#> [232,] -3.830355e-01     0.34406942     0.386707458
#> [233,] -8.263262e-01     0.43526567     0.225848316
#> [234,] -1.298781e+00     0.68377350     0.221283883
#> [235,] -6.333267e-01     1.81069471    -0.038645761
#> [236,] -1.096572e-01     0.40637236     0.344917273
#> [237,]  1.444650e+00     0.76460179     0.379128627
#> [238,] -3.335861e-01     0.39570096     0.223694993
#> [239,]  1.011726e+00     0.56615026     0.023725817
#> [240,] -3.887394e-01     0.16585058     0.239296408
#> [241,] -2.768852e-01     0.56608523    -0.171280524
#> [242,]  5.835386e-01     0.38721155     0.215458860
#> [243,]  1.159905e+00     0.43141228     0.236940560
#> [244,] -2.778238e-01     0.15050464     0.348637898
#> [245,]  1.613866e+00     0.45935237     0.009230291
#> [246,]  8.883246e-01     0.32833882     0.285835463
#> [247,] -1.702211e-02     0.44106524     0.220660817
#> [248,]  1.351182e+00     0.36747290     0.159769224
#> [249,] -2.556307e+00     0.19715108     0.403522026
#> [250,] -7.532052e-01     0.14962648     0.159214337
#> [251,]  4.271920e-01     0.13222366     0.325539654
#> [252,] -5.545571e-01     0.15944044     0.241704744
#> [253,] -2.832947e-01     0.37633935     0.425739868
#> [254,]  1.004310e+00     0.08665446     0.713019212
#> [255,]  8.196869e-01     0.17556587     0.418438193
#> [256,]  1.211881e+00     0.43484294     0.244509986
#> [257,] -2.410236e-01     0.61037969     0.416494801
#> [258,]  1.064042e+00     0.16647389     0.133663838
#> [259,]  1.629273e+00     0.43601291     0.238327710
#> [260,]  1.253956e+00     0.67211541     0.176608247
#> [261,] -3.704263e-02     0.23717865     0.307832472
#> [262,]  4.508647e-01     0.34001618     0.531573302
#> [263,]  2.356827e+00     0.37454775    -0.008562516
#> [264,] -1.502915e+00     0.59676791    -0.003795531
#> [265,] -3.691460e-01     0.09920972     0.630826545
#> [266,]  1.868981e-01     0.87977149     0.249844216
#> [267,]  9.262888e-01     0.39859258     0.152118922
#> [268,]  9.931000e-01     0.59393190     0.427079373
#> [269,] -8.138163e-01     0.17276052     0.454638497
#> [270,]  7.447323e-02     0.42691369     0.304869277
#> [271,] -1.919183e-01     0.38327418     0.315565643
#> [272,] -3.552733e-01     0.82587044    -0.127004372
#> [273,]  2.656244e-01     0.15101685     0.599119102
#> [274,]  5.715141e-01     0.60225585     0.312600488
#> [275,]  1.176857e+00     0.46437148     0.297779857
#> [276,]  1.079628e+00     0.64430786     0.133806218
#> [277,] -1.814556e-01     0.35800207     0.614693082
#> [278,] -2.469149e+00     0.72148205     0.076197548
#> [279,]  3.047516e-01     0.28481410     0.335078410
#> [280,] -3.134211e-01     0.16892056     0.339438466
#> [281,] -5.161982e-01     0.32474431     0.260729216
#> [282,] -3.409274e-03     0.15765740     0.307729593
#> [283,]  1.186790e+00     0.13412358     0.513538264
#> [284,] -2.922525e-01     0.21158855     0.513305352
#> [285,]  4.721757e-01     0.15435729     0.292534293
#> [286,] -1.706365e-01     0.71197092    -0.044040131
#> [287,] -1.606088e+00     0.44684830     0.362467335
#> [288,] -8.444969e-01     0.11660834     0.527435192
#> [289,] -2.464233e-01     0.40719917     0.432799119
#> [290,] -1.496830e+00     0.35494139     0.053939860
#> [291,] -1.373226e+00     0.31707006     0.207462442
#> [292,]  2.150561e+00     1.09704159     0.083169960
#> [293,] -6.510208e-01     0.88949298     0.107864566
#> [294,]  1.809995e-01     0.44352804     0.161329540
#> [295,] -6.055975e-01     0.32995141     0.321871792
#> [296,]  1.264595e+00     0.25415378     0.240945168
#> [297,]  7.629439e-01     0.13396222     0.586465052
#> [298,]  3.003915e+00     0.36583868     0.331678890
#> [299,] -2.232297e-01     0.56249267     0.278264636
#> [300,]  6.250084e-01     0.64765571     0.241910219
#> 
#> $acov_order
#> [1] 0
#> 
#> $acor_order
#> [1] 1
#> 
#> $N
#> [1] 300
#> 
#> $S
#> [1] 50
#> 
#> $R
#> [1] 1000
#>