The naive estimation of the moments

The `nemoment()` function enables to implement the naive estimation of the moments of the heterogeneous mean, the heterogeneous autocovariance, and the heterogeneous autocorrelation. The method is developed by Okui and Yanagi (2019). For more details, see the package vignette with `vignette("panelhetero")`.

Usage

nemoment(data, acov_order = 0, acor_order = 1, R = 1000)

Arguments

data

A matrix of panel data. Each row corresponds to individual time series.

acov_order

A non-negative integer of the order of autocovariance. Default is 0.

acor_order

A positive integer of the order of autocorrelation Default is 1.

R

A positive integer of the number of bootstrap repetitions. Default is 1000.

Value

A list that contains the following elements.

estimate

A vector of the parameter estimates

se

A vector of the standard errors

ci

A matrix of the 95 percent confidence intervals

quantity

A matrix of the estimated heterogeneous quantities

acov_order

The order of autocovariance

acor_order

The order of autocovariance

N

The number of cross-sectional units

S

The length of time series

R

The number of bootstrap repetitions

References

Okui, R. and Yanagi, T., 2019. Panel data analysis with heterogeneous dynamics. Journal of Econometrics, 212(2), pp.451-475.

Examples

data <- panelhetero::simulation(N = 300, S = 50)
panelhetero::nemoment(data = data)
#> $estimate
#>         E(mean)         E(acov)         E(acor)       var(mean)       var(acov) 
#>    -0.005912684     0.358984830     0.254335797     0.999944800     0.049853894 
#>       var(acor) cor(mean, acov) cor(mean, acor) cor(acov, acor) 
#>     0.040663733    -0.066204267     0.032247944    -0.205950778 
#> 
#> $se
#>         E(mean)         E(acov)         E(acor)       var(mean)       var(acov) 
#>     0.058980637     0.012864439     0.011369719     0.085973992     0.007569823 
#>       var(acor) cor(mean, acov) cor(mean, acor) cor(acov, acor) 
#>     0.003033109     0.052136316     0.056530161     0.049593406 
#> 
#> $ci
#>                 95% CI lower 95% CI upper
#> E(mean)          -0.12599365   0.10881685
#> E(acov)           0.33609562   0.38437939
#> E(acor)           0.23184749   0.27623289
#> var(mean)         0.83797360   1.16309514
#> var(acov)         0.03623923   0.06563880
#> var(acor)         0.03465946   0.04650035
#> cor(mean, acov)  -0.16176577   0.04500703
#> cor(mean, acor)  -0.08131518   0.14230559
#> cor(acov, acor)  -0.29617890  -0.10078409
#> 
#> $quantity
#>                mean autocovariance autocorrelation
#>   [1,] -1.827372357     0.11776094     0.047138693
#>   [2,] -0.362969751     0.27522083     0.074181305
#>   [3,] -0.393917584     0.41793766    -0.041411189
#>   [4,] -0.345936798     0.62065275    -0.138743592
#>   [5,] -0.810353244     0.28839790     0.278693144
#>   [6,] -1.137927185     0.32505889     0.089651694
#>   [7,]  0.744660400     0.14091882     0.186314980
#>   [8,]  0.180359921     0.80320694     0.561570934
#>   [9,]  0.768392977     0.35315864     0.051248228
#>  [10,] -0.417818871     0.09433883     0.493488054
#>  [11,] -0.560214305     0.21163219     0.265286243
#>  [12,] -0.885235899     0.77390538     0.510137284
#>  [13,]  0.258918470     0.39525097    -0.059308595
#>  [14,] -0.386197589     0.25419774     0.616486798
#>  [15,] -0.704356905     0.29334656     0.386009767
#>  [16,]  1.019156830     0.40959405     0.414775702
#>  [17,] -0.210553865     0.19535913     0.289606021
#>  [18,]  1.388723733     0.30086495     0.270483148
#>  [19,]  0.240354967     0.33915102     0.482258780
#>  [20,] -0.132286987     0.70251947     0.332022739
#>  [21,]  0.335919743     0.31312306     0.255607839
#>  [22,]  0.874657215     0.12930398     0.473168998
#>  [23,] -0.418450193     0.25312009     0.228383106
#>  [24,] -0.219344634     0.12219053     0.423844541
#>  [25,] -1.751133203     0.18063755     0.404402553
#>  [26,] -0.535790874     0.27045759     0.500749843
#>  [27,] -1.080453998     0.17288850     0.473686983
#>  [28,]  2.302932440     0.72072060     0.166829384
#>  [29,] -0.905568292     0.56642944     0.229176673
#>  [30,]  0.383017549     0.21858773     0.262990479
#>  [31,] -0.198214480     0.43754810     0.609243675
#>  [32,] -0.365431979     0.44276945     0.268543749
#>  [33,] -0.532491867     0.18078459     0.202952859
#>  [34,] -0.408417203     0.16574520     0.292422513
#>  [35,] -0.821419242     0.24823185     0.160541621
#>  [36,] -0.406287956     0.65032243     0.395460922
#>  [37,] -1.528866542     0.26814529     0.026868466
#>  [38,] -0.848092895     0.28069625     0.742978157
#>  [39,] -1.359132791     0.46092287     0.250083210
#>  [40,]  0.132983244     0.05204204     0.437889241
#>  [41,] -0.531043114     0.32505351     0.333549360
#>  [42,] -0.091531164     0.08194349     0.678579216
#>  [43,]  1.173150341     0.29285535     0.181997513
#>  [44,] -0.921921376     0.25160872     0.236401891
#>  [45,]  0.119296132     0.51398152     0.152806303
#>  [46,] -0.090639417     0.40317081     0.331966939
#>  [47,] -0.592046795     0.33039034     0.305820100
#>  [48,] -0.190851616     0.20897278     0.296219380
#>  [49,]  0.008973148     0.28699902     0.047916728
#>  [50,] -1.148305458     0.61396659    -0.023709751
#>  [51,] -1.761696217     0.13907214     0.108947086
#>  [52,] -0.957206827     0.56697016     0.101955716
#>  [53,]  1.169457458     0.37698346     0.562756298
#>  [54,]  0.991687136     0.14116500     0.095240035
#>  [55,] -1.081572885     0.28090212     0.229594403
#>  [56,] -0.372708185     0.42396470     0.485551783
#>  [57,]  0.148178491     0.68733285     0.133012707
#>  [58,] -0.866368788     0.42798465     0.282731701
#>  [59,]  0.721650330     0.76124036    -0.082566322
#>  [60,] -0.049409399     0.22246921     0.545011057
#>  [61,] -1.357178158     0.72900245     0.127161883
#>  [62,] -0.117835905     0.21663417     0.517029138
#>  [63,] -0.054730139     0.28196776     0.232469275
#>  [64,] -2.345751402     0.15357778     0.077363580
#>  [65,] -1.510638266     0.34647842     0.483040837
#>  [66,]  2.257180284     0.23265752     0.072001464
#>  [67,] -1.198306717     1.02916713     0.359469790
#>  [68,] -0.670641183     0.19863708     0.196717589
#>  [69,] -1.918074566     0.28506422    -0.028383277
#>  [70,] -1.445479568     0.25884854     0.320610173
#>  [71,]  0.561236386     0.34237112    -0.096028607
#>  [72,]  0.671560458     0.16978975     0.348276316
#>  [73,]  0.920229271     0.50140804     0.138334663
#>  [74,] -1.320949454     0.15751248     0.635518726
#>  [75,] -0.568285207     0.32985339     0.449462128
#>  [76,]  0.712950688     0.55777010    -0.146502182
#>  [77,]  1.702856350     0.07989781     0.494308832
#>  [78,]  0.156191583     0.22076358    -0.236876669
#>  [79,] -0.924917543     0.43643662     0.286900202
#>  [80,]  0.424982253     0.74372079     0.156235350
#>  [81,] -0.129235116     1.26078442     0.182175859
#>  [82,] -1.028646347     0.11764076     0.107991674
#>  [83,] -0.622639836     0.55330986     0.251317461
#>  [84,] -0.313915015     0.28479292     0.041597640
#>  [85,]  1.323422391     0.18455336     0.030747776
#>  [86,] -0.268394580     0.05113608     0.580303590
#>  [87,]  1.041419117     0.39636281     0.543902302
#>  [88,] -0.017425536     0.57928106     0.495797995
#>  [89,] -0.780131897     0.25245045     0.409763837
#>  [90,] -0.311795239     1.00956300    -0.022079406
#>  [91,] -0.114859505     0.23234111     0.391697138
#>  [92,] -0.295545815     0.12812212     0.343371532
#>  [93,]  0.119403262     0.66921777     0.079351574
#>  [94,] -1.404728044     0.43354778    -0.270527010
#>  [95,] -0.304515867     0.56833644     0.304464683
#>  [96,] -0.665171961     0.54255580     0.360949388
#>  [97,] -2.617927394     0.35348208     0.028383458
#>  [98,]  1.452577767     0.21940857     0.130052846
#>  [99,] -0.569532889     0.45399180    -0.118566595
#> [100,] -0.591305287     1.02058562     0.217115583
#> [101,] -0.380112853     0.27265200     0.437532077
#> [102,] -0.190368553     0.20755703     0.497749295
#> [103,]  0.425629033     0.14743186     0.216267536
#> [104,] -0.158450532     0.65370156     0.064148733
#> [105,] -0.057710012     0.34736496     0.145747922
#> [106,] -0.610102042     0.33244379     0.155989827
#> [107,]  1.169520117     0.49024349     0.018017628
#> [108,]  0.898086904     0.22296263     0.607467562
#> [109,] -1.216896645     0.53588515     0.275390597
#> [110,] -0.365251444     0.16405806     0.400403733
#> [111,] -0.187412799     0.15227175     0.343108772
#> [112,]  2.166575394     0.13175373     0.370586470
#> [113,] -0.950522739     0.37085677     0.575044920
#> [114,] -0.605099002     0.21198908     0.399896623
#> [115,] -1.680602978     0.32069738     0.090385024
#> [116,] -0.407211705     0.46156313     0.157265280
#> [117,] -0.378047679     0.83387712    -0.018793509
#> [118,] -0.629862545     0.25197037     0.124823027
#> [119,] -1.661391331     0.25334576     0.624779591
#> [120,] -0.161128196     0.22967769     0.616643453
#> [121,]  1.349071181     0.19824307     0.235183942
#> [122,] -0.989254867     0.25365383     0.393508462
#> [123,] -0.200046663     0.09352849     0.231781932
#> [124,]  0.795500472     0.29251669     0.198426931
#> [125,] -1.216401386     1.02458005     0.222149719
#> [126,]  0.344130244     0.57940724     0.045732598
#> [127,]  1.369123506     0.17567643     0.307163284
#> [128,] -0.786735511     0.33674391     0.228056277
#> [129,] -0.117410043     0.27962957     0.574402895
#> [130,] -1.502203960     0.91626377     0.307463213
#> [131,]  0.165593331     0.15807908     0.069402823
#> [132,]  0.131296972     0.39351163     0.068917670
#> [133,]  0.396492403     0.46376984    -0.050463556
#> [134,]  0.028228068     1.16920989     0.054252715
#> [135,]  0.195559197     0.30324173     0.339993503
#> [136,]  1.645369979     0.39807381     0.366012553
#> [137,]  0.330939488     0.33897470     0.414848994
#> [138,]  0.199008545     0.24141349     0.078366541
#> [139,] -1.103990287     0.25529937     0.334945192
#> [140,]  1.461715144     0.20041129     0.369127971
#> [141,]  0.512827046     0.30334177     0.239987361
#> [142,]  1.743681006     0.44243493     0.247895984
#> [143,] -1.360408331     1.01195394    -0.035160790
#> [144,]  0.194557768     0.19097401     0.389314684
#> [145,] -0.561053447     0.21067551     0.217543466
#> [146,] -0.763232108     0.58894149     0.250653050
#> [147,] -0.572091034     0.43119137    -0.084274741
#> [148,]  2.484213707     0.49211645     0.070567051
#> [149,]  0.896570572     1.50092103     0.356644826
#> [150,]  0.951207986     0.23999381     0.303469313
#> [151,]  0.368908381     0.41626035     0.216496392
#> [152,]  1.338349654     0.13714253     0.383023109
#> [153,]  1.068369004     0.18001019     0.391826095
#> [154,]  1.036060865     0.32078610     0.700310643
#> [155,] -0.322116199     0.27036025    -0.128307730
#> [156,]  1.713194837     0.52684173     0.539385162
#> [157,] -1.474766064     0.69853482     0.393147210
#> [158,] -0.350420724     0.49649501     0.628088046
#> [159,] -1.078172881     1.40225766     0.079095809
#> [160,] -0.138987166     0.26774179     0.015677080
#> [161,] -1.302848456     0.20270811     0.400971791
#> [162,]  0.236936579     0.56942304     0.423369870
#> [163,]  0.479327431     0.29173323     0.429781283
#> [164,] -1.063741006     0.23543530     0.152507642
#> [165,]  1.009215864     0.13835937     0.071410860
#> [166,] -0.021536434     0.64045788     0.008687075
#> [167,]  0.837004785     0.42100296     0.204830564
#> [168,]  1.911972975     0.42776695     0.172032184
#> [169,]  1.060414518     0.26701067     0.232902315
#> [170,]  1.490042188     0.36520484     0.065750356
#> [171,] -0.041815083     0.31637712     0.071142959
#> [172,]  0.265555401     0.31419131    -0.031142210
#> [173,]  0.918356762     0.23744296     0.444814898
#> [174,]  0.367062835     0.45384020     0.193039007
#> [175,]  0.152247117     0.34510517     0.130503493
#> [176,]  0.604650824     0.23078983     0.348893514
#> [177,] -0.927588252     0.22335138     0.271594518
#> [178,]  0.994317305     0.33168568    -0.175604727
#> [179,] -0.710736770     0.56370287     0.479541351
#> [180,]  0.067697292     0.21854743     0.504274162
#> [181,]  0.088319915     0.09319122     0.419503974
#> [182,]  0.422112244     0.36761618     0.609247856
#> [183,]  0.365736240     0.14210520     0.535343763
#> [184,]  0.226561392     0.22094787     0.361978824
#> [185,] -1.264931388     0.40912012     0.215273416
#> [186,] -0.656147304     0.15143656     0.317477674
#> [187,]  0.245335131     0.34443244     0.417952530
#> [188,]  1.664220389     0.16182220     0.219851751
#> [189,]  0.218478248     0.46930825     0.608318787
#> [190,] -0.762372311     0.36318095     0.491324752
#> [191,]  0.477415209     0.38225149     0.190670740
#> [192,]  0.950422759     0.54315890     0.196442667
#> [193,] -0.839518613     0.25386742     0.274104357
#> [194,] -0.390553787     0.20610168     0.019904680
#> [195,]  0.393927809     0.28389319     0.213292383
#> [196,] -0.336984120     0.27391824     0.214955312
#> [197,] -1.856214176     0.59862413     0.384395804
#> [198,] -0.825678259     0.48243327     0.148392970
#> [199,]  1.314259366     0.82469567     0.311252363
#> [200,]  0.658284800     0.21614439     0.117162657
#> [201,] -0.911524963     0.21259144     0.384175120
#> [202,]  1.664605313     0.71598235    -0.107276210
#> [203,] -0.434503597     0.21016591     0.098773377
#> [204,] -0.045146263     0.41397630     0.254467525
#> [205,]  1.107880701     0.52674399     0.261203864
#> [206,]  1.105532494     0.43276987     0.063045754
#> [207,]  0.982646245     0.16000625     0.589243143
#> [208,]  0.267968053     0.54449544     0.187183323
#> [209,]  1.528006250     0.26024168     0.338780462
#> [210,]  1.536589238     0.22255021     0.304470147
#> [211,]  1.225077309     0.24397058     0.231713728
#> [212,] -0.585073628     0.82531388     0.348950159
#> [213,]  2.328732779     0.08727410     0.587298647
#> [214,]  1.168162900     0.12762917     0.264546774
#> [215,]  0.147293189     0.26300222     0.209127508
#> [216,] -0.723459644     0.38071262    -0.118024792
#> [217,] -0.378446524     0.35028822     0.069864630
#> [218,]  1.156320351     0.26697629     0.057333134
#> [219,] -1.248658580     0.14907496     0.262705237
#> [220,]  0.097258581     0.16464107    -0.250445601
#> [221,] -0.606243852     0.04991500     0.486668508
#> [222,]  1.793971173     0.44958678    -0.062668020
#> [223,]  0.501606952     0.56126449     0.391428445
#> [224,]  1.928860613     0.44453427    -0.173335258
#> [225,] -0.069583401     0.19808074     0.445590926
#> [226,] -0.278421037     0.41535707     0.040575704
#> [227,]  0.633482265     0.23917310     0.152097542
#> [228,]  0.589913536     0.16481413     0.328106894
#> [229,] -0.582948167     0.31647035    -0.031814853
#> [230,]  0.232735151     0.67548652     0.192091206
#> [231,] -0.702164129     0.82850340     0.249212386
#> [232,] -0.289463008     0.20028499     0.310986205
#> [233,]  1.476624113     0.27945903     0.137618577
#> [234,] -0.439228812     0.55848483     0.061093294
#> [235,] -0.742121446     0.47089885     0.340481845
#> [236,]  1.162316954     0.18719839     0.196337589
#> [237,] -1.360167594     0.49902826     0.406006896
#> [238,]  0.062101349     0.18688477     0.339904022
#> [239,] -0.509566672     0.24863927     0.265050716
#> [240,]  2.975686567     0.26020915     0.379138424
#> [241,] -0.388585877     0.25936641     0.131358238
#> [242,]  0.369645370     0.39226431     0.194622857
#> [243,] -0.382896774     0.11364991     0.050234831
#> [244,] -0.866886739     0.28587576     0.015148474
#> [245,] -0.509066606     0.42534869    -0.096547280
#> [246,] -0.122448335     0.18177748     0.700301095
#> [247,] -1.082868689     0.36128175    -0.095818266
#> [248,] -1.187531184     0.28020819     0.053332819
#> [249,] -0.037208118     0.30225313     0.164894915
#> [250,]  0.663594641     0.37423994     0.638368435
#> [251,] -0.489680098     0.48958885     0.243271423
#> [252,] -0.248558547     0.43929475     0.064757068
#> [253,]  3.433531186     0.39919838     0.236384994
#> [254,]  0.698382877     0.17283756     0.201626636
#> [255,]  1.855844479     0.14311718     0.454045267
#> [256,] -0.236565359     0.33630750    -0.069445438
#> [257,]  0.822071680     0.30271064     0.196638088
#> [258,] -2.160047306     0.23168335     0.275942915
#> [259,]  1.101821001     0.27401275     0.419635694
#> [260,]  0.825248885     0.28687642     0.361861621
#> [261,]  0.792722327     0.43038199     0.356946914
#> [262,] -0.007966809     0.23304512     0.148257578
#> [263,]  0.744106285     0.41619921     0.377574874
#> [264,] -0.350676952     0.11741625     0.741960831
#> [265,] -1.080745505     0.25557791     0.021462482
#> [266,]  0.617317664     0.46982659     0.223140038
#> [267,]  0.426764055     0.27491366     0.176983166
#> [268,] -0.052774038     0.26771653     0.190069962
#> [269,] -0.403533461     0.45324718     0.258777627
#> [270,] -1.500673539     0.22106897     0.181481739
#> [271,] -0.428083759     0.44616473     0.378876244
#> [272,] -0.551039798     0.14367236     0.623511599
#> [273,] -0.809945581     0.30447929     0.320226879
#> [274,] -0.063691722     0.06742362     0.465530112
#> [275,] -0.218838749     0.40480221     0.559857458
#> [276,] -0.840388861     0.31560675     0.197474097
#> [277,]  0.850892207     0.20988626     0.205534498
#> [278,]  1.196380094     0.30859351     0.426378434
#> [279,] -1.802718871     0.35878505     0.359405347
#> [280,]  0.807539414     0.31514163     0.026812658
#> [281,]  0.498570248     0.07054769     0.085592689
#> [282,]  1.468455790     0.10542578     0.325544110
#> [283,]  0.585856919     0.73319535    -0.007543652
#> [284,] -0.046032483     0.36364468     0.365701011
#> [285,]  1.122922753     0.49000918    -0.004718015
#> [286,] -0.376417579     0.68929063     0.148246725
#> [287,]  1.598896391     0.47248934     0.401680415
#> [288,] -0.112820902     0.15784670     0.513228412
#> [289,] -0.357593017     0.32057578     0.105898401
#> [290,] -1.081600137     0.27501590     0.351610203
#> [291,] -1.416712983     0.22727150    -0.067575888
#> [292,] -1.385807722     0.12581781     0.336805849
#> [293,] -0.593184486     0.18089334     0.352940220
#> [294,]  2.691878069     0.27122376     0.626300641
#> [295,]  0.420064533     0.42069627     0.112361064
#> [296,] -1.958132257     0.37731269     0.134141365
#> [297,] -1.575781464     0.21761135     0.445382308
#> [298,]  1.679105171     0.32711963     0.245071066
#> [299,]  0.330452943     0.16945610     0.346390573
#> [300,] -0.664928207     0.09840204     0.483421760
#> 
#> $acov_order
#> [1] 0
#> 
#> $acor_order
#> [1] 1
#> 
#> $N
#> [1] 300
#> 
#> $S
#> [1] 50
#> 
#> $R
#> [1] 1000
#>