The naive estimation of the moments

The `nemoment()` function enables to implement the naive estimation of the moments of the heterogeneous mean, the heterogeneous autocovariance, and the heterogeneous autocorrelation. The method is developed by Okui and Yanagi (2019). For more details, see the package vignette with `vignette("panelhetero")`.

Usage

nemoment(data, acov_order = 0, acor_order = 1, R = 1000)

Arguments

data

A matrix of panel data. Each row corresponds to individual time series.

acov_order

A non-negative integer of the order of autocovariance. Default is 0.

acor_order

A positive integer of the order of autocorrelation Default is 1.

R

A positive integer of the number of bootstrap repetitions. Default is 1000.

Value

A list that contains the following elements.

estimate

A vector of the parameter estimates

se

A vector of the standard errors

ci

A matrix of the 95 percent confidence intervals

quantity

A matrix of the estimated heterogeneous quantities

acov_order

The order of autocovariance

acor_order

The order of autocovariance

N

The number of cross-sectional units

S

The length of time series

R

The number of bootstrap repetitions

References

Okui, R. and Yanagi, T., 2019. Panel data analysis with heterogeneous dynamics. Journal of Econometrics, 212(2), pp.451-475.

Examples

data <- panelhetero::simulation(N = 300, S = 50)
panelhetero::nemoment(data = data)
#> $estimate
#>         E(mean)         E(acov)         E(acor)       var(mean)       var(acov) 
#>     -0.06498559      0.37436366      0.22264075      0.92027650      0.05367739 
#>       var(acor) cor(mean, acov) cor(mean, acor) cor(acov, acor) 
#>      0.03992576      0.07061541     -0.00922347     -0.27078170 
#> 
#> $se
#>         E(mean)         E(acov)         E(acor)       var(mean)       var(acov) 
#>     0.054655006     0.013561027     0.011731090     0.078774394     0.006101294 
#>       var(acor) cor(mean, acov) cor(mean, acor) cor(acov, acor) 
#>     0.003365896     0.057775441     0.059625233     0.043642061 
#> 
#> $ci
#>                 95% CI lower 95% CI upper
#> E(mean)          -0.16805954   0.03988915
#> E(acov)           0.34781463   0.40011882
#> E(acor)           0.20053272   0.24633435
#> var(mean)         0.76328395   1.07113949
#> var(acov)         0.04222642   0.06546380
#> var(acor)         0.03381989   0.04680631
#> cor(mean, acov)  -0.04105165   0.17971662
#> cor(mean, acor)  -0.13438762   0.10024627
#> cor(acov, acor)  -0.35955735  -0.18869228
#> 
#> $quantity
#>                mean autocovariance autocorrelation
#>   [1,] -0.996526721     0.24322532    0.0004888641
#>   [2,]  0.594936656     0.28468083    0.1073518183
#>   [3,] -0.128136268     0.10515777    0.1715133781
#>   [4,] -1.863903002     0.13288558    0.0651573029
#>   [5,]  0.173199317     0.17130407    0.5115081848
#>   [6,]  0.441632314     0.29581194    0.4306723197
#>   [7,]  0.270657972     0.24387159    0.5396690766
#>   [8,]  0.459681677     0.44472716    0.1940842907
#>   [9,] -0.288780325     0.24541729   -0.1858592729
#>  [10,] -1.641718730     0.53437327    0.0136636586
#>  [11,] -0.675976113     0.23930337    0.2917848041
#>  [12,] -0.283753450     0.60206843    0.1839641812
#>  [13,] -0.474972039     0.30460776   -0.0953306627
#>  [14,]  0.019312667     0.20079971   -0.2192191576
#>  [15,]  1.287719753     0.85555001    0.0839694432
#>  [16,] -0.831035442     0.19749549    0.5449652586
#>  [17,]  0.739348169     0.60668404    0.1171075985
#>  [18,]  0.728648631     0.50822978    0.3826063211
#>  [19,]  1.413106314     0.61208987    0.2202802438
#>  [20,] -1.750137987     0.19165970    0.0540399569
#>  [21,] -0.005522551     0.40353686    0.1027812777
#>  [22,] -0.356741657     0.42228394    0.1055018711
#>  [23,] -0.528192278     0.22259186    0.3209355606
#>  [24,]  0.449037545     0.09601357    0.8835621956
#>  [25,]  0.299494178     0.35552963    0.3549826172
#>  [26,]  1.206931467     0.85274387    0.2951938211
#>  [27,] -0.538339641     0.13393626    0.3519855854
#>  [28,] -0.036538271     0.49725014    0.1995224364
#>  [29,]  0.860792072     0.27152778    0.6451535265
#>  [30,]  0.858360572     0.29409632    0.2704066857
#>  [31,] -0.081646875     0.40675944    0.3595630432
#>  [32,] -1.664498516     0.21682472    0.3262257480
#>  [33,] -0.117610075     0.29759371    0.3475235627
#>  [34,] -0.043453753     0.23144602   -0.1383647928
#>  [35,] -0.298893155     0.14272447    0.4315708233
#>  [36,] -0.323245090     0.65176943   -0.0153514525
#>  [37,]  0.529452430     0.34262615    0.1493812339
#>  [38,]  0.908443362     0.34112208    0.1084859358
#>  [39,] -1.225250063     0.10434041    0.4889030860
#>  [40,] -0.878676232     0.43361923    0.1558226247
#>  [41,]  0.688393591     0.25638676    0.0485527208
#>  [42,]  0.126174006     0.21503295    0.2841040417
#>  [43,] -1.214226765     0.28680306    0.0470627949
#>  [44,] -0.084138919     0.71948362    0.1817914205
#>  [45,]  0.782827136     0.21035020    0.4736750912
#>  [46,]  1.091255588     0.16560024    0.3381602966
#>  [47,] -1.994622139     0.20951392    0.3365541239
#>  [48,] -1.458735660     0.51664825    0.4827412785
#>  [49,]  2.246201022     0.67963871    0.1157445129
#>  [50,]  0.531482817     0.20964414    0.2828411994
#>  [51,] -0.714514751     0.31090217    0.1909129878
#>  [52,] -0.766969684     0.23277310    0.5796842085
#>  [53,]  0.752371914     0.12086131    0.0091190060
#>  [54,]  0.207928581     0.06807477    0.5278978862
#>  [55,]  0.197082314     0.25950201    0.1677213945
#>  [56,] -1.976558155     0.12979317    0.0446235455
#>  [57,] -0.966432310     0.12572694    0.3427077580
#>  [58,] -0.874828902     0.29728492    0.1377659600
#>  [59,] -1.138073125     0.12170184    0.3388095632
#>  [60,] -0.449605521     0.25408835   -0.0541987065
#>  [61,]  0.986309669     0.64195905   -0.0781964144
#>  [62,]  1.195852722     0.07155153    0.6553832287
#>  [63,]  0.514826688     0.12088018    0.3231158917
#>  [64,]  2.291940403     0.81968721   -0.2221473430
#>  [65,]  0.418443197     0.68834501    0.0276142964
#>  [66,]  0.860390123     0.34249446    0.1384207732
#>  [67,]  0.468066399     0.26155453    0.1539770163
#>  [68,] -1.116878026     0.20708542    0.0710700445
#>  [69,] -0.202327186     0.41457073   -0.0172272750
#>  [70,]  0.063286043     0.53684906    0.0136362874
#>  [71,] -0.741453351     0.36789343    0.0875479293
#>  [72,]  0.581678392     0.43950832    0.1967912025
#>  [73,] -0.625596306     0.35358814    0.1743864691
#>  [74,]  0.355647915     0.29997290    0.1262186230
#>  [75,]  0.727554728     0.19036536    0.5596330088
#>  [76,] -1.269701592     1.04310088    0.2498559745
#>  [77,]  0.127746018     0.50191380    0.2332111094
#>  [78,]  1.015706047     0.44419008    0.1141038929
#>  [79,]  0.515057623     0.15640818    0.5042087955
#>  [80,] -0.257886049     0.04826158    0.3624439884
#>  [81,]  1.044771404     0.45666809    0.5395202648
#>  [82,] -1.226390679     0.23403242    0.3302362775
#>  [83,]  0.075565392     0.50846365    0.3717190498
#>  [84,] -1.399268148     0.28445500    0.1805339106
#>  [85,] -2.525615624     0.20046718    0.0447513498
#>  [86,] -1.746022319     0.58394614    0.2211825339
#>  [87,]  0.616170513     0.26299815    0.5090847090
#>  [88,] -2.651476798     0.23738974    0.0218927812
#>  [89,] -0.131404199     1.06647228    0.2920038882
#>  [90,] -0.555629599     0.71074880    0.3784673967
#>  [91,] -0.251810675     0.41174599    0.5067868448
#>  [92,] -1.410104548     0.15928756    0.2495030301
#>  [93,]  1.628299628     0.18418767    0.3570865026
#>  [94,] -0.250898567     0.49396103    0.0361703584
#>  [95,]  1.510763285     0.34227831    0.1819982255
#>  [96,] -1.255992381     0.47000843    0.1609816632
#>  [97,] -0.398175638     0.20895278    0.2778809064
#>  [98,]  0.562397045     0.85669667    0.3449719502
#>  [99,]  0.135883871     0.21169386    0.1788238020
#> [100,] -1.087364586     0.65278349   -0.0645906363
#> [101,]  1.668893208     0.41846903    0.4404233528
#> [102,]  0.120804549     0.14041357    0.3895195876
#> [103,]  0.291689960     0.25002468    0.2456949449
#> [104,]  0.179390571     0.97130811    0.2891054637
#> [105,]  0.751657879     0.19303871    0.0237450706
#> [106,] -0.488485696     0.82952555    0.2309061575
#> [107,]  0.903829901     0.40831968    0.4907411734
#> [108,]  1.669879719     0.28219524    0.2888611776
#> [109,] -1.600558384     0.16405787    0.2755971262
#> [110,]  0.203966138     0.28000851    0.1001068974
#> [111,] -1.789591582     0.32388701    0.3027749775
#> [112,] -0.477450471     0.21543243    0.4910525953
#> [113,]  0.190320661     0.79113107    0.1110263000
#> [114,] -0.427966378     1.20029208    0.1054091988
#> [115,]  0.476879018     0.24817390    0.0520878667
#> [116,] -1.249297133     0.15380945    0.2945102753
#> [117,] -0.544697069     0.26252600   -0.0046443082
#> [118,] -0.138573738     0.52857882    0.0998871766
#> [119,] -0.414294458     0.37255324    0.1432660961
#> [120,] -0.197175435     0.37723919    0.0609224412
#> [121,]  0.395044121     0.60333139   -0.0925680605
#> [122,] -1.266558233     0.10330897    0.4747673534
#> [123,] -0.016983027     0.20815271    0.4182493346
#> [124,]  0.310979095     0.37493250    0.1287785967
#> [125,] -0.285273186     0.35200015   -0.0585196474
#> [126,] -0.503029825     0.56435151    0.2066766267
#> [127,] -0.599444567     0.29759002    0.2950329830
#> [128,]  0.615158515     0.24763732    0.2011955793
#> [129,]  1.540992479     0.38286296    0.2572837877
#> [130,] -0.016289389     0.41853340   -0.1550200662
#> [131,] -2.162946036     0.59604240   -0.1161386244
#> [132,] -1.620188277     0.18098135    0.3146180553
#> [133,]  0.752531291     0.33619083    0.2180194989
#> [134,]  0.679722420     0.22389858   -0.0709755117
#> [135,] -0.436312014     0.35134108    0.2625420982
#> [136,] -0.370290518     0.17972321    0.3182841874
#> [137,]  0.690389278     0.33215738   -0.0030303261
#> [138,] -1.419507710     0.19022969    0.6795730495
#> [139,] -0.647087161     0.28404487    0.5447284947
#> [140,] -0.526791352     0.39181092    0.3259914514
#> [141,] -2.393713076     0.19641328    0.1440211845
#> [142,]  1.247222645     0.56250908    0.0949467978
#> [143,]  0.558961153     0.50796855    0.2393216336
#> [144,] -0.361175581     0.32307641    0.4696805658
#> [145,]  1.211436196     0.20246345    0.3009357409
#> [146,]  0.363778056     0.29882411    0.2429697823
#> [147,] -0.533820712     0.13194800    0.7563909706
#> [148,]  0.527654477     0.25499300    0.4013055612
#> [149,]  0.220492310     0.51486546    0.2107009002
#> [150,] -0.060692836     0.56109009    0.0468716533
#> [151,] -0.347540202     0.18730416   -0.0231430112
#> [152,] -0.158702954     0.83330182    0.0964843938
#> [153,] -1.463218628     0.28983971    0.3497468634
#> [154,]  0.131331455     0.47222969    0.0014350501
#> [155,]  0.170442464     0.60878955   -0.0705799148
#> [156,] -0.381288593     0.34253707    0.2118631374
#> [157,] -0.969843401     0.44579463    0.3646682955
#> [158,]  0.388914817     0.25879959    0.5458554216
#> [159,] -1.736050735     0.51726120   -0.0079977331
#> [160,] -0.912450068     1.21393131    0.1665092073
#> [161,] -0.770847990     0.16495311    0.3988436243
#> [162,] -0.111110772     0.46631257   -0.0438739120
#> [163,]  0.363150839     0.14777106    0.3391025762
#> [164,] -0.348413200     0.22605075    0.3170694220
#> [165,] -2.260642710     0.61135515    0.5100691739
#> [166,]  0.156791560     0.12979191    0.7502344639
#> [167,]  0.092366419     0.60731450    0.2557813053
#> [168,]  1.371690283     0.37978666    0.1024360210
#> [169,] -1.056102278     0.17644768    0.0070122282
#> [170,]  0.302450647     0.49048583    0.0323245116
#> [171,] -0.803727860     0.23962886    0.4425483022
#> [172,] -1.381416913     0.15615412    0.5608581381
#> [173,] -0.916328153     0.99372759    0.2929142260
#> [174,] -0.178184057     0.66026199   -0.0408417031
#> [175,] -0.520734379     0.24600498   -0.0525695294
#> [176,]  0.217984587     0.74691372    0.1307312405
#> [177,] -0.884412117     0.26548294    0.1906759688
#> [178,]  0.769261921     0.29718901    0.3891439617
#> [179,]  0.074752331     0.71693627    0.2476885794
#> [180,]  0.386634165     0.22782507    0.1966995201
#> [181,] -0.004117706     0.29560985    0.3060409656
#> [182,] -0.687684762     0.45220519    0.0446683257
#> [183,] -0.431517974     0.34545746    0.3441995049
#> [184,]  0.177388945     0.37017063    0.4799497993
#> [185,]  0.100673619     0.30847681    0.6512146703
#> [186,] -0.088725684     0.37630239    0.2171013928
#> [187,]  0.658609235     0.50027567    0.4296020248
#> [188,] -1.406337863     0.26302363    0.2396588366
#> [189,] -0.543072636     1.14709653    0.1906191264
#> [190,]  0.246640430     0.29055213    0.2074454067
#> [191,]  0.722156438     0.52416412    0.1436160175
#> [192,]  1.173295744     0.28222623   -0.0104233080
#> [193,]  0.384212357     0.26439997    0.0617592359
#> [194,]  0.667349490     0.82383840   -0.0031727939
#> [195,] -0.472657889     0.14630663    0.2333071610
#> [196,]  0.288240457     0.47069382    0.2495852696
#> [197,] -0.116948038     0.20333649   -0.0200552818
#> [198,] -1.316722556     0.34457892    0.1850673289
#> [199,] -0.340309974     0.29156136    0.2959106093
#> [200,]  0.826503106     0.46508478    0.1806698401
#> [201,]  0.122428281     0.57777036    0.0752581098
#> [202,]  1.496884911     0.32578658    0.2166479289
#> [203,] -1.167929640     0.35417834    0.3848659427
#> [204,] -1.518021740     0.19416911   -0.0699951034
#> [205,] -0.402665325     0.45246184    0.0232554202
#> [206,]  0.049656017     0.57060951    0.1311125896
#> [207,]  0.945011745     0.23323520    0.3842476903
#> [208,]  0.535194027     0.35755319   -0.2151817502
#> [209,]  0.847733634     0.16707915    0.1837515244
#> [210,] -0.049803268     0.33304017    0.2515720944
#> [211,]  0.357133102     0.37908099    0.1173750645
#> [212,]  0.026939573     0.27287172    0.0027966817
#> [213,]  0.959572040     0.54657778    0.4193450133
#> [214,]  0.193207477     0.19487559    0.3941813502
#> [215,]  1.129207015     0.26903870    0.2582455076
#> [216,] -3.543168949     0.53807424    0.0044247268
#> [217,] -0.783494039     0.23962408    0.2824051686
#> [218,]  1.181916824     0.20301810    0.0934581924
#> [219,] -0.647891420     0.07389473    0.3716258306
#> [220,] -0.427156575     0.31373522    0.6183800915
#> [221,] -0.730371924     0.19769468    0.2189816759
#> [222,]  0.281280803     0.19828568    0.0488524092
#> [223,] -0.009537231     0.23439507    0.4652608480
#> [224,]  1.403673157     0.36487262    0.0861512844
#> [225,] -0.696707902     0.56935616    0.1244652319
#> [226,]  0.854744614     0.86841033    0.2499434059
#> [227,]  1.944473950     0.29982644    0.1071979186
#> [228,]  2.215987962     0.52798754    0.1799883360
#> [229,]  2.106381119     0.46691789    0.2050954527
#> [230,] -0.640157218     0.09528780    0.5312092248
#> [231,]  0.120662524     0.38424785    0.2074580782
#> [232,]  0.265105920     0.36332876    0.1477620898
#> [233,]  2.137251647     0.40991451    0.2979058427
#> [234,] -1.185636813     0.04194094    0.4437647134
#> [235,] -0.169430924     0.19504624    0.1247256887
#> [236,] -0.052572228     0.45291822    0.0985181786
#> [237,]  0.108887802     0.31564076    0.2081711168
#> [238,]  0.128685223     0.17027326   -0.0159111993
#> [239,]  0.790045055     0.25466648    0.0662061540
#> [240,]  0.891664343     0.82483169   -0.0747465688
#> [241,] -1.743982360     0.76187203    0.2897333993
#> [242,]  0.255288822     0.24910994    0.2519298110
#> [243,]  1.691742945     0.06940678    0.2175984414
#> [244,] -1.360887983     0.13084145    0.1642271654
#> [245,]  1.386344137     0.23661906    0.0799802622
#> [246,]  0.491940913     0.40653247    0.2106059112
#> [247,]  0.858261980     0.18262341    0.5028814137
#> [248,] -0.719841685     0.08174480    0.1713376199
#> [249,] -0.358939741     0.32733930    0.2451659871
#> [250,] -0.135816364     0.23672795    0.1034159930
#> [251,]  0.457825336     0.71025821    0.1365637173
#> [252,]  0.608331098     0.19244039    0.5461241589
#> [253,] -0.225849308     0.28636577    0.8391874225
#> [254,] -0.503212691     0.17431550    0.1132443894
#> [255,] -0.277291437     0.40655456    0.3461762468
#> [256,] -0.140792294     0.54820624    0.0566248098
#> [257,] -1.222098415     0.77871460    0.2566922925
#> [258,]  2.168819775     0.17248373   -0.2341344373
#> [259,] -0.987522756     0.29292047    0.3467693135
#> [260,] -1.192962771     0.53456516    0.1932745483
#> [261,]  0.299884872     0.23640333    0.3796780308
#> [262,]  0.262726168     0.57763317    0.1012701334
#> [263,] -0.079551277     0.12540163    0.8009486859
#> [264,] -0.415125753     0.58526330    0.2884166866
#> [265,] -0.995685670     0.33638701    0.1937970879
#> [266,]  0.090627963     0.26311337    0.2075441663
#> [267,]  0.796103814     0.29747024    0.5872928186
#> [268,]  1.284111336     0.06310519    0.3673986529
#> [269,] -0.344543147     0.38784936   -0.0244976632
#> [270,] -0.172314933     1.06295754    0.0193555941
#> [271,]  0.458891689     0.13318189    0.3736471311
#> [272,] -0.243158764     0.11473630    0.2823000525
#> [273,]  0.896301821     0.69170326    0.1068868816
#> [274,] -0.199603872     0.85912559    0.2155815191
#> [275,] -1.080198012     0.09486859    0.4927449670
#> [276,]  0.468511469     0.37427742    0.0205992363
#> [277,]  0.269088037     0.40784333    0.3577730199
#> [278,]  0.974399643     0.20903980   -0.1184086933
#> [279,]  1.550090340     0.18557966    0.3694351046
#> [280,] -1.254296663     0.66937521    0.0845642961
#> [281,] -2.427735337     0.70754493   -0.0451343067
#> [282,]  0.442926966     0.26416333    0.2646521019
#> [283,]  1.551415233     1.14712615    0.1278140564
#> [284,]  0.213134374     1.22243236    0.1501295665
#> [285,]  1.029248662     0.61769762    0.0616713253
#> [286,] -1.707402339     0.25644506   -0.1290169346
#> [287,]  1.171922632     0.61542785    0.2450299752
#> [288,] -0.174864221     0.11622204    0.3185574526
#> [289,]  1.153398153     0.09000472    0.4130626712
#> [290,] -1.150271172     0.23812219   -0.0321013039
#> [291,] -0.283654365     0.36616104    0.1682571896
#> [292,]  0.376458738     0.30211480    0.2618478772
#> [293,]  0.909904977     0.45890523   -0.1348829141
#> [294,]  0.335062323     0.38543930    0.0680314361
#> [295,] -0.540728572     0.17309592    0.6278392933
#> [296,] -0.922924036     0.58656253    0.3265047215
#> [297,] -1.130617922     0.23462207    0.2712991657
#> [298,] -0.383173391     0.12688810    0.4048090802
#> [299,] -0.871349505     0.25611442    0.3948776138
#> [300,]  0.358982833     0.37122711    0.0348705793
#> 
#> $acov_order
#> [1] 0
#> 
#> $acor_order
#> [1] 1
#> 
#> $N
#> [1] 300
#> 
#> $S
#> [1] 50
#> 
#> $R
#> [1] 1000
#>