The TOJ bias-corrected estimation of the moments

The `tojmoment()` function enables to implement the TOJ bias-corrected estimation of the moments of the heterogeneous mean, the heterogeneous autocovariance, and the heterogeneous autocorrelation. The method is developed by Okui and Yanagi (2019). For more details, see the package vignette with `vignette("panelhetero")`.

Usage

tojmoment(data, acov_order = 0, acor_order = 1, R = 1000)

Arguments

data

A matrix of panel data. Each row corresponds to individual time series.

acov_order

A non-negative integer of the order of autocovariance. Default is 0.

acor_order

A positive integer of the order of autocorrelation Default is 1.

R

A positive integer of the number of bootstrap repetitions. Default is 1000.

Value

A list that contains the following elements.

estimate

A vector of the parameter estimates

se

A vector of the standard errors

ci

A matrix of the 95 percent confidence intervals

quantity

A matrix of the estimated heterogeneous quantities

acov_order

The order of autocovariance

acor_order

The order of autocovariance

N

The number of cross-sectional units

S

The length of time series

R

The number of bootstrap repetitions

References

Okui, R. and Yanagi, T., 2019. Panel data analysis with heterogeneous dynamics. Journal of Econometrics, 212(2), pp.451-475.

Examples

data <- panelhetero::simulation(N = 300, S = 50)
panelhetero::tojmoment(data = data)
#> $estimate
#>         E(mean)         E(acov)         E(acor)       var(mean)       var(acov) 
#>     0.044040991     0.388329623     0.294987261     1.044563813     0.036557242 
#>       var(acor) cor(mean, acov) cor(mean, acor) cor(acov, acor) 
#>     0.003929695    -0.063083602    -0.002207621    -0.665942947 
#> 
#> $se
#>         E(mean)         E(acov)         E(acor)       var(mean)       var(acov) 
#>     0.060377613     0.013691971     0.016664008     0.093864745     0.010118357 
#>       var(acor) cor(mean, acov) cor(mean, acor) cor(acov, acor) 
#>     0.007820768     0.066113861     0.090447920     0.099966799 
#> 
#> $ci
#>                 95% CI lower 95% CI upper
#> E(mean)          -0.07157737   0.16231487
#> E(acov)           0.36283179   0.41652916
#> E(acor)           0.26330289   0.32766366
#> var(mean)         0.87979255   1.22845055
#> var(acov)         0.01845589   0.05855364
#> var(acor)        -0.01163408   0.01914029
#> cor(mean, acov)  -0.19547112   0.06691406
#> cor(mean, acor)  -0.17241924   0.17546916
#> cor(acov, acor)  -0.87976473  -0.48005964
#> 
#> $quantity
#>                mean autocovariance autocorrelation
#>   [1,] -0.001501665     0.47403254    0.2332542814
#>   [2,]  0.567838086     0.41398818    0.0270804796
#>   [3,] -1.890238065     0.14162362    0.2610873867
#>   [4,]  0.045281450     0.49555509    0.3659023726
#>   [5,] -2.376538046     0.74803495   -0.1116901714
#>   [6,]  0.167799347     0.41587853    0.3150319941
#>   [7,]  0.451803104     0.36333408    0.1332813319
#>   [8,] -0.328018059     0.26575600    0.3929099447
#>   [9,]  0.146625655     0.48710876    0.3039339085
#>  [10,]  0.063524583     0.24522163    0.6266179144
#>  [11,] -0.603999300     0.17573856    0.2609112902
#>  [12,] -0.302427509     0.60391591    0.4701467733
#>  [13,]  1.230283887     0.21113734    0.3107888287
#>  [14,]  1.141719990     0.26756236    0.0279428467
#>  [15,] -0.305959196     0.14741354    0.6122713767
#>  [16,]  0.005195534     0.72523256   -0.1986907492
#>  [17,]  1.245404570     0.16770469    0.4840062512
#>  [18,]  1.568268400     0.39140703    0.1109659746
#>  [19,] -0.279698487     0.53243588    0.1648118312
#>  [20,]  1.135357736     0.71753031    0.1554041118
#>  [21,] -0.044891821     0.23236383    0.1644674476
#>  [22,] -0.195740354     0.34745971    0.0948644896
#>  [23,] -0.726733016     0.56553161   -0.0057973595
#>  [24,] -0.246175992     0.41494934    0.2979267827
#>  [25,]  3.484225620     0.22709249    0.2620799539
#>  [26,]  0.138598484     0.24305789   -0.0518610798
#>  [27,]  0.110509693     0.31474508    0.4929748597
#>  [28,] -1.269981506     0.22161587    0.3517693258
#>  [29,]  0.718887099     0.40885264    0.1771353850
#>  [30,]  1.792986398     0.18599914    0.4227076239
#>  [31,]  2.034097276     0.33751623    0.2025511747
#>  [32,] -0.432781760     0.45343259    0.1483030124
#>  [33,] -0.100620837     0.30790625   -0.0229401947
#>  [34,] -0.790476442     0.23169655    0.2505197992
#>  [35,] -0.347966219     0.30395754    0.3069098495
#>  [36,]  1.382720858     0.33346731    0.2101792015
#>  [37,] -0.524309221     0.45399073   -0.0492568798
#>  [38,]  0.869442895     0.37380589    0.1518226364
#>  [39,]  0.756608489     0.20714434    0.2486130538
#>  [40,] -1.957783937     0.71783407    0.1531410258
#>  [41,] -0.515609722     0.29774493    0.1873962568
#>  [42,] -1.479909502     0.26247330    0.1059276092
#>  [43,] -1.403764118     0.33646166    0.2581037021
#>  [44,] -2.644216422     0.19319766    0.3612894479
#>  [45,] -1.348514928     0.21830342   -0.0858033513
#>  [46,] -1.859417377     0.20246386    0.6212578447
#>  [47,]  0.388851672     0.28797413    0.0836080006
#>  [48,] -0.488529869     0.29296536    0.3508106971
#>  [49,]  0.115460771     0.18674864    0.2924031835
#>  [50,] -0.166071743     0.44628686   -0.0791441888
#>  [51,]  1.514691878     0.26764769    0.3563225936
#>  [52,] -0.426271139     0.23996511    0.5709825781
#>  [53,]  1.970123533     0.23933479    0.3714288250
#>  [54,]  0.020066340     0.72455338    0.1856364426
#>  [55,] -0.316670099     0.20188080    0.0556326219
#>  [56,] -0.838775525     0.98206347    0.0280845657
#>  [57,] -1.157477218     0.42503087    0.0421567275
#>  [58,]  0.843680712     0.34084037    0.0463098939
#>  [59,]  1.160624964     0.24731665    0.1681670418
#>  [60,] -0.060694527     0.10904998    0.1694057421
#>  [61,] -1.669629776     0.25527256    0.1198009221
#>  [62,]  0.319843119     0.24078243    0.7623974674
#>  [63,]  1.317939980     0.44696519    0.2848775230
#>  [64,] -0.117826572     0.61818011   -0.1758249748
#>  [65,] -0.885513160     0.26147331    0.2355714712
#>  [66,] -0.511209885     0.61754332    0.0396363668
#>  [67,] -0.711479574     0.39924887    0.2818367570
#>  [68,] -0.618950826     0.31454249    0.2708497586
#>  [69,] -0.150495687     0.22252867    0.1809339250
#>  [70,]  0.330415881     0.20106393    0.1235362865
#>  [71,]  0.610099190     0.12987070    0.7294500155
#>  [72,]  1.941867719     0.35222448    0.5029798728
#>  [73,]  0.219220689     0.08983702    0.4201014158
#>  [74,]  2.323104301     0.49878452    0.2343902634
#>  [75,] -0.154603628     0.30100526    0.0951480154
#>  [76,] -2.992479495     0.20844272    0.5007785368
#>  [77,] -1.423032186     0.11135408    0.4919719736
#>  [78,]  1.197912674     0.41549407    0.0551493995
#>  [79,]  0.073328778     0.50416794    0.0624229667
#>  [80,]  1.149690687     0.26433982    0.2340817079
#>  [81,]  0.507111645     0.71549015    0.1436839189
#>  [82,] -1.638298993     0.40760057    0.5027824626
#>  [83,]  1.644814233     0.37033220    0.2312245809
#>  [84,]  0.515395885     0.08082237    0.6050213719
#>  [85,]  0.881532032     0.29367784   -0.0008865025
#>  [86,] -1.070935997     0.50994520    0.0296972752
#>  [87,] -0.863711872     0.10798378    0.5343002544
#>  [88,]  1.144887066     0.36649250    0.3476765728
#>  [89,]  0.182421645     0.25702803    0.3772497113
#>  [90,]  0.689582992     0.42630524    0.1817752421
#>  [91,] -0.721737398     0.64900482    0.2268568343
#>  [92,] -0.687194991     0.28440029    0.1363718299
#>  [93,] -0.704332217     0.46817450    0.4854468185
#>  [94,] -0.401072959     0.13807479    0.6880834795
#>  [95,]  0.462997203     0.20851710    0.2218851259
#>  [96,] -1.190153019     0.23244661    0.1783114177
#>  [97,]  1.228326912     0.78788635    0.1090563609
#>  [98,] -0.274708701     0.90444376    0.0132723719
#>  [99,]  0.152567139     0.47662746    0.3217202562
#> [100,]  0.062642351     0.24188422    0.6864646809
#> [101,]  1.216969566     0.37104823    0.3043410749
#> [102,]  0.195658302     0.48244509    0.4955555575
#> [103,]  0.581541934     0.29678170    0.5716959485
#> [104,] -1.241106258     0.30581093    0.3197507795
#> [105,] -0.017307821     0.40777111   -0.1217690722
#> [106,]  0.033261263     0.41421982    0.5134053131
#> [107,]  0.036194501     0.44283115    0.2988877358
#> [108,]  0.223271864     0.29233113    0.0356030686
#> [109,] -0.449085644     0.38676703    0.3234594768
#> [110,]  0.855717383     0.36928904    0.2870718453
#> [111,] -1.326961298     0.14020653    0.3712393633
#> [112,]  0.700858111     0.23521092    0.0894275733
#> [113,]  0.185098022     0.18667437    0.1554678510
#> [114,]  2.501191959     0.39423918    0.3874588168
#> [115,]  0.323628487     0.29612323    0.0195599355
#> [116,] -0.079003977     0.25036366    0.2356311883
#> [117,]  0.692699754     0.42966901    0.2304261727
#> [118,] -1.362910358     1.29466270    0.2064881851
#> [119,]  0.006180809     0.25536048    0.0443241086
#> [120,]  0.857791872     0.68862209    0.1903587220
#> [121,] -0.849313516     0.64173388    0.3230200851
#> [122,]  1.923652377     0.33855395    0.2329341184
#> [123,] -1.041110193     0.23901913    0.4345285396
#> [124,] -0.005606970     0.26082829    0.1574121786
#> [125,]  0.002640370     0.55303647    0.6404808349
#> [126,] -0.856678181     0.31591732    0.3948383686
#> [127,]  0.542172302     0.19220747    0.3929996523
#> [128,] -0.961065867     0.44813402    0.1061933729
#> [129,] -0.968342580     0.32187752    0.2332209940
#> [130,] -0.355846213     0.29096351   -0.0104768620
#> [131,] -0.209558896     0.56176652    0.1664837856
#> [132,]  0.805285208     0.59064423    0.2028342981
#> [133,]  1.023156133     0.22084557    0.2684276779
#> [134,] -0.474516948     0.09513083    0.3922765562
#> [135,]  0.360224084     0.24389556    0.4159673306
#> [136,]  0.011388097     0.81732486    0.2580515100
#> [137,] -1.655057362     0.35044394    0.2619870402
#> [138,]  1.862577872     0.29408386   -0.0312657629
#> [139,] -0.700932772     0.64353076    0.0810386021
#> [140,] -1.420250444     0.77645044    0.1531575543
#> [141,]  1.307633530     0.29266537    0.3970446470
#> [142,] -0.332998477     0.32324462    0.2800680947
#> [143,] -0.379463702     0.50700668    0.0719697420
#> [144,]  0.601423988     0.18051915    0.1600041003
#> [145,]  1.207850389     0.48903145    0.3468075574
#> [146,]  0.438269916     0.50558161    0.4590392560
#> [147,] -0.614489137     0.16088424    0.4601137558
#> [148,]  0.248674023     0.51829224    0.0019046520
#> [149,]  0.357019510     1.32521421    0.3093759047
#> [150,] -0.030890305     0.21433049    0.3128374105
#> [151,]  0.203026425     0.34813234    0.1244316531
#> [152,]  0.565839300     0.33011779    0.0563461602
#> [153,]  0.151824191     0.43009303    0.5640270562
#> [154,]  1.255150067     0.19467208    0.5911904639
#> [155,] -0.307634384     0.23003703    0.0955245468
#> [156,] -0.118521817     0.43152729    0.2552274991
#> [157,]  0.374412764     1.51419903    0.2950717000
#> [158,] -1.241156308     0.29679612    0.5076895347
#> [159,] -0.352584911     0.08664676    0.4957846563
#> [160,]  1.450027291     0.62372207    0.0270381850
#> [161,]  0.312495602     0.49698579    0.1167969500
#> [162,]  0.460245199     0.87674400    0.1654429394
#> [163,]  0.448554463     0.44928031    0.2216441935
#> [164,] -0.235768627     0.28612276    0.2318764290
#> [165,]  0.193551825     0.35997192    0.2954435596
#> [166,]  0.639778672     0.52674278    0.1898598492
#> [167,]  0.456755498     0.29391342    0.2273372662
#> [168,]  0.387189086     0.15691728    0.1906578650
#> [169,]  0.526903044     0.30020355    0.1559599973
#> [170,]  1.025470184     0.35067442    0.2286756230
#> [171,] -1.095501521     0.50006668    0.2237031727
#> [172,] -1.181687409     0.03344736    0.6367662845
#> [173,]  0.463061415     0.79485925    0.0487447665
#> [174,] -0.913374718     0.58628273    0.3354479075
#> [175,]  0.901016818     0.42547147    0.0214060824
#> [176,] -0.990396276     0.31138468   -0.1031519222
#> [177,]  0.261259368     0.68812859    0.3305679964
#> [178,] -1.535550776     0.43375834    0.4815794808
#> [179,]  1.769363927     0.32551256    0.1982242902
#> [180,]  0.166121795     0.30332948    0.2258141446
#> [181,]  1.028220420     0.52260490    0.0893866341
#> [182,] -1.106166297     0.21845732    0.3418321638
#> [183,] -0.912087528     0.43510547    0.1580495094
#> [184,]  1.965192172     0.32867935    0.2292087582
#> [185,] -1.134364685     0.29292862    0.4323880401
#> [186,]  0.457740782     0.25071638    0.4572706646
#> [187,]  0.911386199     0.30999781   -0.0129298467
#> [188,]  0.188637619     0.51525488    0.2813046851
#> [189,]  1.136548431     0.31172482    0.2015309143
#> [190,] -1.224088394     0.18115324    0.7566782004
#> [191,] -0.215351444     0.35930333    0.3305069379
#> [192,]  0.332836355     0.34066236    0.1489241737
#> [193,]  0.343134206     0.21579446    0.0453023702
#> [194,]  1.463705939     0.44744625    0.2399549101
#> [195,] -0.222909217     0.75290428    0.1203432321
#> [196,] -2.991508522     0.80470980    0.1322977785
#> [197,]  1.154700317     0.57958413    0.3909346421
#> [198,]  0.706633755     0.16582083   -0.0300008536
#> [199,]  1.275766491     0.17950509    0.1388030633
#> [200,]  0.853436118     0.78004040    0.1591909066
#> [201,] -0.027317922     0.23300852    0.1410855500
#> [202,] -0.899958615     0.71704617    0.2248955717
#> [203,]  1.214569444     0.34199553    0.0431847939
#> [204,]  0.308689642     0.31104428    0.0841519884
#> [205,]  0.438130327     0.58447946    0.1252569185
#> [206,] -1.269066969     0.52756922    0.3332638315
#> [207,]  1.789108225     0.23415084    0.4051794420
#> [208,]  1.183197679     0.18567347    0.3444880919
#> [209,]  1.428338855     0.14676859    0.4299300471
#> [210,]  0.443874613     0.37669070    0.0238705990
#> [211,] -1.065692552     0.12509731    0.4527037880
#> [212,]  0.438332095     0.10339931    0.4861039985
#> [213,] -0.785399575     0.18195694    0.0679607010
#> [214,] -0.129388094     0.29431736    0.0605932869
#> [215,]  0.941511742     0.34289535    0.3332967297
#> [216,]  0.200205660     0.45411444    0.2265543908
#> [217,]  0.681476750     0.18801110    0.2385263479
#> [218,] -0.672584642     0.23622299    0.1692431232
#> [219,]  1.721680706     0.29831959    0.2221455856
#> [220,]  1.872896746     0.62518179    0.0603506794
#> [221,]  0.564241359     0.36416206    0.1361465799
#> [222,] -0.732541673     0.19290093    0.3999298326
#> [223,]  0.719386045     0.12214644    0.2746197704
#> [224,] -0.213605712     0.64269007    0.3167060868
#> [225,] -0.627122937     0.55843104   -0.0796299339
#> [226,] -0.225217716     0.38558904   -0.0920415361
#> [227,] -0.719388608     0.31642684    0.1056102914
#> [228,] -0.114857321     0.39168069    0.3673918439
#> [229,] -1.746634205     0.21245915    0.0268765347
#> [230,] -1.408902957     0.46068614    0.0916757799
#> [231,]  0.256144574     0.63543771    0.1123718401
#> [232,] -0.885568308     0.38145341    0.3733070569
#> [233,] -2.406139505     0.58460259    0.0245698062
#> [234,]  0.828978020     0.45967977    0.5657249177
#> [235,] -0.865095277     0.32119087   -0.1213823331
#> [236,] -0.570822620     0.26772108   -0.0461708194
#> [237,] -0.581449728     0.52604119   -0.0294420267
#> [238,] -1.507105693     0.23698325    0.2097881584
#> [239,]  1.144998675     0.11314841    0.1401266354
#> [240,] -0.205641754     0.18080395    0.3557216355
#> [241,] -2.387452783     0.25300275    0.0079578913
#> [242,] -1.126575125     0.18425744    0.0926656846
#> [243,]  0.661834573     0.78605189    0.0821756363
#> [244,] -0.021863147     0.21409670    0.4104352388
#> [245,] -2.495748477     0.26695536    0.1725653917
#> [246,] -0.158458563     0.56930686    0.4735019910
#> [247,]  0.842484030     0.13944676    0.4725959695
#> [248,]  0.528024979     0.19702994    0.1836834595
#> [249,]  0.501923592     0.40570639    0.2474038976
#> [250,]  0.947638648     0.25109389    0.4833547829
#> [251,] -0.582817147     0.54711115    0.1369467510
#> [252,] -1.467669037     0.45791078    0.2630642289
#> [253,]  0.115052947     0.32436314    0.3505380338
#> [254,] -0.957737168     0.10085806    0.2001031406
#> [255,]  1.147452716     0.80822322    0.0166086736
#> [256,]  0.009445767     0.15277506    0.2171038792
#> [257,] -1.277136221     0.47546985    0.2360689033
#> [258,]  1.449143893     0.21703414    0.2626242654
#> [259,] -0.807951666     0.26111292    0.1621802311
#> [260,]  0.053901133     0.18506152    0.3686789803
#> [261,] -0.962775405     1.01382096   -0.2977482377
#> [262,]  1.076430491     0.42460607    0.1762855547
#> [263,]  0.893764630     0.45798990   -0.0842455091
#> [264,] -0.242118849     0.42369864   -0.0890278726
#> [265,] -0.767515000     0.24105549    0.3826059728
#> [266,]  0.207639111     0.22808342    0.1063902815
#> [267,] -0.516534130     0.16014893    0.1185823460
#> [268,]  1.204563577     0.46741120    0.1669653497
#> [269,]  0.875252116     0.27138389    0.3881981145
#> [270,]  1.052375301     0.51705521   -0.0443912977
#> [271,]  0.119842149     0.14445426    0.5446025182
#> [272,] -0.591473928     0.24708460    0.1634224336
#> [273,]  1.102317481     0.20341989    0.3920812416
#> [274,]  1.830794993     0.42852231    0.1486004889
#> [275,] -0.793364899     0.30873126    0.3764955255
#> [276,] -0.132473909     0.25397394    0.0870372665
#> [277,]  1.237665708     0.19183075    0.1499154409
#> [278,]  1.377078453     0.30142813    0.2515439743
#> [279,]  0.986889910     0.43017329   -0.0324888360
#> [280,] -1.506396279     0.34612556    0.1337107030
#> [281,] -0.030031789     0.22483778   -0.0760724303
#> [282,]  0.453823236     0.24221465    0.0926880938
#> [283,] -1.590293502     0.17636731    0.0395568625
#> [284,] -0.738049398     0.37223685   -0.0236197081
#> [285,] -0.865787627     0.25823399    0.4078584189
#> [286,]  0.358388360     0.31341712    0.2329593706
#> [287,]  0.144858005     0.29316195    0.4616203732
#> [288,]  1.966841697     0.27312863    0.0271878164
#> [289,]  0.923013467     0.46788841    0.2295964378
#> [290,] -0.213757055     0.26073625    0.2592098414
#> [291,] -0.893039869     0.23633736    0.2129744136
#> [292,]  0.626832064     0.08405582    0.6743703765
#> [293,]  0.040058521     0.40633478    0.1825239590
#> [294,]  2.563053630     0.20144246    0.6477398059
#> [295,] -0.507909287     0.59558317    0.1701090932
#> [296,]  1.219807201     0.26497658   -0.0071247286
#> [297,] -1.872991405     0.58282341    0.1811228699
#> [298,] -0.215758656     1.38529496    0.1129840513
#> [299,]  0.108932245     0.08081627    0.6061543866
#> [300,]  0.168922603     0.39649065    0.2837466575
#> 
#> $acov_order
#> [1] 0
#> 
#> $acor_order
#> [1] 1
#> 
#> $N
#> [1] 300
#> 
#> $S
#> [1] 50
#> 
#> $R
#> [1] 1000
#>