The TOJ bias-corrected estimation of the moments

The `tojmoment()` function enables to implement the TOJ bias-corrected estimation of the moments of the heterogeneous mean, the heterogeneous autocovariance, and the heterogeneous autocorrelation. The method is developed by Okui and Yanagi (2019). For more details, see the package vignette with `vignette("panelhetero")`.

Usage

tojmoment(data, acov_order = 0, acor_order = 1, R = 1000)

Arguments

data

A matrix of panel data. Each row corresponds to individual time series.

acov_order

A non-negative integer of the order of autocovariance. Default is 0.

acor_order

A positive integer of the order of autocorrelation Default is 1.

R

A positive integer of the number of bootstrap repetitions. Default is 1000.

Value

A list that contains the following elements.

estimate

A vector of the parameter estimates

se

A vector of the standard errors

ci

A matrix of the 95 percent confidence intervals

quantity

A matrix of the estimated heterogeneous quantities

acov_order

The order of autocovariance

acor_order

The order of autocovariance

N

The number of cross-sectional units

S

The length of time series

R

The number of bootstrap repetitions

References

Okui, R. and Yanagi, T., 2019. Panel data analysis with heterogeneous dynamics. Journal of Econometrics, 212(2), pp.451-475.

Examples

data <- panelhetero::simulation(N = 300, S = 50)
panelhetero::tojmoment(data = data)
#> $estimate
#>         E(mean)         E(acov)         E(acor)       var(mean)       var(acov) 
#>    -0.049192284     0.380922436     0.290531050     1.044624624     0.036145990 
#>       var(acor) cor(mean, acov) cor(mean, acor) cor(acov, acor) 
#>     0.014202369    -0.062391441    -0.009502196    -0.628268583 
#> 
#> $se
#>         E(mean)         E(acov)         E(acor)       var(mean)       var(acov) 
#>     0.057033428     0.013206558     0.016863231     0.090053800     0.005232293 
#>       var(acor) cor(mean, acov) cor(mean, acor) cor(acov, acor) 
#>     0.007163084     0.061599611     0.089168205     0.088070461 
#> 
#> $ci
#>                 95% CI lower 95% CI upper
#> E(mean)         -0.160132502   0.06261817
#> E(acov)          0.356210944   0.40698579
#> E(acor)          0.256536322   0.32335523
#> var(mean)        0.872320535   1.22945946
#> var(acov)        0.026271260   0.04718517
#> var(acor)        0.000170456   0.02823125
#> cor(mean, acov) -0.184663719   0.05585371
#> cor(mean, acor) -0.185812635   0.16909143
#> cor(acov, acor) -0.795892203  -0.44989153
#> 
#> $quantity
#>                mean autocovariance autocorrelation
#>   [1,] -0.213608501     0.12425250     0.198774036
#>   [2,]  1.436796065     0.30501662     0.358474097
#>   [3,] -1.954641270     0.36150380     0.462529756
#>   [4,] -1.375384291     0.18622672     0.277332150
#>   [5,] -2.488045524     0.43304986     0.243255104
#>   [6,] -0.991085904     0.36644335     0.235078669
#>   [7,]  2.265653545     0.16174393     0.001357980
#>   [8,] -1.225472469     0.33065011     0.222290442
#>   [9,] -1.466507082     0.22169403     0.222794021
#>  [10,]  1.224424684     0.74044610     0.376320059
#>  [11,] -0.135796161     0.47593738     0.089183023
#>  [12,] -0.056068362     0.84731860     0.294657794
#>  [13,]  0.137534656     0.19646731     0.634520435
#>  [14,]  0.715094755     1.09471023     0.172538500
#>  [15,]  1.053044175     0.38464716     0.069372417
#>  [16,] -0.921697892     0.52056278     0.283035286
#>  [17,]  0.904927589     0.32744356     0.165700559
#>  [18,]  0.022839816     0.46566246     0.263724672
#>  [19,] -0.326114967     0.51103470     0.227939994
#>  [20,] -0.779801348     0.63766400     0.536281916
#>  [21,]  1.803304480     0.76555572     0.198884642
#>  [22,] -1.331082936     0.18644465     0.145090617
#>  [23,] -0.738546788     0.74811895     0.105937655
#>  [24,] -0.143052595     0.06211099     0.404488978
#>  [25,]  0.905642761     0.20128323     0.523702418
#>  [26,] -0.985048860     0.45845330     0.249500371
#>  [27,] -1.251393162     0.19420308     0.290391164
#>  [28,]  1.403399852     0.22452536     0.254677542
#>  [29,] -0.080667804     0.13177745     0.360063132
#>  [30,]  0.619239431     0.40832854     0.069820625
#>  [31,] -0.724126591     0.23494873     0.161252547
#>  [32,] -0.713763520     0.28141933     0.277819190
#>  [33,]  0.576587144     0.15002007     0.534850209
#>  [34,] -1.137436350     0.41376270     0.023100928
#>  [35,] -0.539948315     0.32833322     0.199049392
#>  [36,]  0.563224344     0.49305892     0.006473744
#>  [37,]  0.032329469     0.44689308     0.364156429
#>  [38,] -1.324501085     0.38787915     0.651641326
#>  [39,] -0.229753292     0.16572302     0.196602430
#>  [40,]  0.059805650     0.22922628     0.430721601
#>  [41,]  0.017000867     0.31626659    -0.047316718
#>  [42,]  1.982281269     0.58583448     0.201181706
#>  [43,] -0.464558219     0.22207451     0.270720834
#>  [44,] -1.845952866     0.35052185     0.510398184
#>  [45,]  0.590249330     0.51083572     0.281504708
#>  [46,] -1.926381583     0.33327119     0.281649488
#>  [47,] -1.146701112     0.19668268     0.384013437
#>  [48,] -0.211208468     0.56654732     0.492234954
#>  [49,] -1.297983040     0.20413779     0.391688389
#>  [50,]  0.497053995     0.33971394     0.182697215
#>  [51,]  0.382675726     0.25439568     0.314828648
#>  [52,] -0.167608133     0.65163144     0.332221318
#>  [53,] -0.110914683     0.18901808     0.568984942
#>  [54,]  0.190421428     0.09688350     0.668047975
#>  [55,] -0.954695377     0.50771964     0.197084219
#>  [56,]  0.185822832     0.30812463     0.490122730
#>  [57,]  0.030700885     0.23114685     0.630156554
#>  [58,]  0.477049812     0.11288559     0.406904204
#>  [59,] -1.695441010     0.53739515     0.131163045
#>  [60,]  0.356849549     0.18497006     0.343017187
#>  [61,] -0.916094675     0.28672305     0.324185765
#>  [62,]  0.337861145     0.18952750     0.379914794
#>  [63,] -0.626132328     0.22795670     0.472538351
#>  [64,]  0.542592809     0.14475955     0.126015546
#>  [65,] -0.072980289     0.31001965    -0.134943681
#>  [66,] -0.174997606     0.60914089     0.094506752
#>  [67,]  1.581463466     0.60879994     0.410086632
#>  [68,] -1.251457168     0.31069629     0.167730687
#>  [69,] -0.879366845     0.24532670     0.007581016
#>  [70,]  0.200837837     1.04214919     0.302963759
#>  [71,] -0.086152763     0.74960338    -0.040295211
#>  [72,]  0.025077309     0.12437869     0.339369963
#>  [73,] -1.561113498     0.25172928     0.153840519
#>  [74,] -0.081254891     0.06889043     0.478808322
#>  [75,]  0.603693468     0.52782646     0.343024802
#>  [76,]  0.607606925     0.35898846     0.159195414
#>  [77,]  1.870724360     0.22859328     0.140134894
#>  [78,] -1.270942070     0.33473866     0.168612127
#>  [79,]  0.516144584     0.57156950    -0.091643750
#>  [80,]  0.773516847     0.46517788     0.406526912
#>  [81,]  0.506389541     0.39976621    -0.074618165
#>  [82,]  0.599686509     0.09753647     0.485239772
#>  [83,] -0.668208857     0.88356778    -0.355733285
#>  [84,] -1.000022401     0.09881832     0.668749424
#>  [85,]  0.275897993     0.14942970     0.611075479
#>  [86,]  1.052143150     0.17154764     0.077978544
#>  [87,] -2.257458569     0.58444272     0.159413092
#>  [88,] -0.273878813     0.60608431     0.098667974
#>  [89,]  0.775974098     0.58957002     0.266054708
#>  [90,] -0.945906588     0.14972051     0.678646603
#>  [91,] -0.056814605     0.25786996    -0.069186514
#>  [92,]  1.992039158     0.42009345     0.344601092
#>  [93,] -0.373320105     0.40095781     0.323298453
#>  [94,] -0.789115926     0.76015035     0.164777273
#>  [95,] -0.564988615     0.14599312     0.211911363
#>  [96,]  0.225030843     0.19157263     0.202749284
#>  [97,] -0.673215027     0.76478291     0.173186069
#>  [98,]  0.207537169     0.51972272     0.224720154
#>  [99,] -0.346096875     0.08838454     0.249524364
#> [100,] -0.081055657     0.17347548     0.687870324
#> [101,] -0.624372481     0.15825896     0.794355468
#> [102,]  0.089902869     0.95763015    -0.153401322
#> [103,]  1.027288307     0.74360853     0.220488075
#> [104,] -1.508516429     0.16188846     0.232777289
#> [105,] -0.495097298     0.98495206    -0.040817560
#> [106,] -1.577471977     0.20103738     0.282792229
#> [107,]  0.602801706     0.82260899     0.345451089
#> [108,]  1.030650545     0.19042911     0.002578848
#> [109,] -0.851815408     0.70425740     0.234092495
#> [110,] -0.310877180     1.08106386     0.009439485
#> [111,] -1.514075704     0.36663373    -0.027840159
#> [112,] -0.795246461     0.13059292     0.464149530
#> [113,] -1.279199411     0.56905814    -0.016803855
#> [114,]  1.561253984     0.16560760     0.333160777
#> [115,]  0.415123260     0.64614569     0.056122472
#> [116,] -1.272347029     0.19177879     0.180403935
#> [117,]  1.069232570     0.19115610     0.297268808
#> [118,] -1.016959955     0.53141881     0.213588314
#> [119,]  0.244118779     0.24888215     0.456053392
#> [120,]  3.859080062     0.34084221     0.241086340
#> [121,] -0.053645271     0.73399633     0.086635498
#> [122,]  0.607766176     0.12262438     0.239925914
#> [123,] -1.443360286     0.63729889     0.202846699
#> [124,]  0.632207738     1.11755172    -0.076899158
#> [125,]  0.349027911     0.20335649     0.484430701
#> [126,]  0.213272886     0.33747510     0.091054907
#> [127,]  1.251681590     0.15253257    -0.164632779
#> [128,] -0.303222749     0.29130108     0.033432298
#> [129,]  2.427514702     0.38274339     0.552747628
#> [130,] -0.766292372     0.30199242     0.046067627
#> [131,]  0.023290691     0.23678821     0.067640988
#> [132,] -0.664566479     0.68122310     0.374009039
#> [133,]  0.590319469     0.12079593     0.356371639
#> [134,]  0.254782352     0.80201322     0.400699722
#> [135,] -0.002124489     0.37422379     0.177144472
#> [136,] -1.157021359     0.17712479     0.285548703
#> [137,]  0.767792962     0.18673813     0.205872358
#> [138,]  0.566022041     0.32209365     0.218838658
#> [139,] -0.240767059     0.38811109    -0.020427112
#> [140,]  1.238401307     0.20367323     0.370065777
#> [141,] -0.391539665     0.30713071     0.006390556
#> [142,] -0.370088484     0.15811333     0.197428949
#> [143,] -0.199212119     0.43329469     0.301450152
#> [144,]  0.292491155     0.47506751     0.359906368
#> [145,]  1.142555138     0.08273131     0.416250827
#> [146,]  0.757605446     0.49874509     0.105897074
#> [147,] -1.448857343     0.44890019     0.438055337
#> [148,] -0.754833103     0.32687847     0.244856569
#> [149,] -1.764480100     0.38859318     0.185971130
#> [150,] -0.560362188     0.88453645    -0.122155393
#> [151,]  2.084735484     0.19166553     0.575504619
#> [152,]  0.177252569     0.50489452     0.143823616
#> [153,] -0.267501297     0.28308295     0.132591070
#> [154,]  1.194903572     0.13302122     0.269915436
#> [155,] -0.848709885     0.17401572     0.234265008
#> [156,] -1.664761084     0.97509369     0.247392609
#> [157,] -0.561565075     0.32133287     0.294757277
#> [158,]  0.710980588     0.55468949     0.140037682
#> [159,]  0.841754794     0.18136885     0.181702145
#> [160,]  1.408765150     0.46037027     0.363481406
#> [161,]  0.342151061     0.18108982     0.097924567
#> [162,]  0.053434212     0.48810318     0.226976965
#> [163,] -0.682500740     0.57253996     0.051351856
#> [164,]  0.533656103     0.21245504     0.005928281
#> [165,]  0.923249163     0.47472594     0.416109074
#> [166,]  0.849787430     0.61567337     0.435990740
#> [167,]  0.792961805     0.13485238     0.557242547
#> [168,]  0.197509068     0.44017563     0.200384228
#> [169,]  1.961833291     0.49401208     0.543501446
#> [170,]  0.248500553     0.29087181     0.276501516
#> [171,] -0.516804218     0.27243611     0.086177832
#> [172,] -0.313884751     0.26916384     0.341434352
#> [173,] -0.940627040     0.49330351     0.311698491
#> [174,]  1.091695000     0.63575643     0.024084030
#> [175,] -1.632301345     0.41547739     0.274845184
#> [176,]  0.290282826     0.21100348     0.250955714
#> [177,] -0.875494495     0.28560222    -0.014972039
#> [178,] -0.149048343     0.87036765     0.059136687
#> [179,]  1.657610438     0.09456416     0.475440019
#> [180,]  0.355687181     0.11442459     0.643961350
#> [181,] -1.184461580     0.30901217     0.160519534
#> [182,]  0.055604259     0.15876472     0.487690187
#> [183,]  1.026142580     0.30894139     0.116440594
#> [184,]  1.647700739     0.52275897     0.168744225
#> [185,] -0.402013318     0.42798749     0.126111000
#> [186,]  0.195131752     0.39599775     0.152578769
#> [187,] -0.778825102     0.53303454     0.351271082
#> [188,] -0.269272472     0.16015388     0.525112071
#> [189,] -1.082697213     0.45109206     0.122725294
#> [190,]  1.367199504     0.17877936     0.590591389
#> [191,]  1.131446325     0.41987218     0.323736355
#> [192,]  0.149156099     0.39970709     0.020601898
#> [193,]  0.529159725     0.44289883     0.196790961
#> [194,] -0.040768134     0.44197647     0.065957636
#> [195,] -0.390192872     0.37267392     0.365742843
#> [196,]  1.333558719     0.13228707     0.619590211
#> [197,] -0.782652610     0.27924727    -0.110960483
#> [198,] -0.859345141     0.38471365     0.462781155
#> [199,]  1.736415276     0.25018242     0.299371333
#> [200,]  0.330707561     0.13875223     0.474856158
#> [201,] -0.577451964     0.41178337     0.408446980
#> [202,] -0.022688798     0.95747367     0.260524329
#> [203,] -1.960979630     0.86453300    -0.185387398
#> [204,] -1.794191160     0.66328806     0.248056342
#> [205,]  0.188920574     0.88944032     0.172790145
#> [206,] -2.375478136     0.15272703     0.231174231
#> [207,]  0.331780983     0.28156174     0.183572245
#> [208,]  1.532001986     0.29043073    -0.040289219
#> [209,] -0.078525116     0.13791736     0.256088574
#> [210,] -0.042812534     0.25200209     0.123353553
#> [211,]  0.598525913     0.11385187     0.449401358
#> [212,] -0.599218393     0.20449928     0.380453840
#> [213,] -1.333651884     0.36365018     0.076571798
#> [214,] -0.516557022     0.18966218     0.208762476
#> [215,] -1.647181289     0.55866737     0.515593931
#> [216,] -0.658507080     0.22524773     0.258948398
#> [217,]  0.558753938     0.30325507    -0.006210116
#> [218,] -0.530675159     0.25615492     0.352766530
#> [219,] -1.547376929     0.19577427     0.290876657
#> [220,] -1.618753583     0.50624532     0.542378372
#> [221,]  1.652066347     0.70763369    -0.113684075
#> [222,]  0.229780783     0.11404423     0.770543926
#> [223,] -0.976998633     0.33117014     0.413884015
#> [224,]  0.199620949     0.23522260    -0.185298070
#> [225,] -0.028602806     0.55933703    -0.063080089
#> [226,] -0.668026969     0.32342404     0.076533466
#> [227,] -1.075435188     0.26765442     0.602857446
#> [228,]  0.593110597     0.34997458     0.211517651
#> [229,]  1.057252709     0.13209799     0.262648421
#> [230,] -0.648754898     0.08449352     0.335251249
#> [231,] -0.035286557     0.38692800     0.286090544
#> [232,]  0.123561190     0.28842612     0.316305719
#> [233,]  1.280296454     0.12888855     0.526365197
#> [234,] -0.732970674     0.37077902     0.441458240
#> [235,] -0.518432509     0.21713778     0.241421746
#> [236,] -0.054565582     0.44428083     0.374884174
#> [237,]  1.025094676     0.14044131     0.233862651
#> [238,] -0.805879414     0.58960716     0.462067321
#> [239,]  2.074504663     0.10524818     0.404611420
#> [240,] -0.875002094     0.39106815     0.101670076
#> [241,] -1.266782620     0.63948876     0.081367673
#> [242,] -0.260155489     0.34533840     0.249172398
#> [243,]  0.038034833     0.44879752     0.335911125
#> [244,] -0.231705386     0.42028807     0.002271494
#> [245,]  0.419365532     0.34524091     0.720711661
#> [246,]  1.015185653     0.40009161    -0.056161437
#> [247,]  1.341930970     0.23292642     0.299155642
#> [248,] -1.143742166     0.16416153     0.005707448
#> [249,] -2.487659921     0.38391307     0.214685192
#> [250,]  1.061081682     0.66308028     0.165308276
#> [251,] -1.611011829     0.49317562     0.437574544
#> [252,]  1.759256053     0.46720877    -0.005523462
#> [253,] -2.142420070     0.21513487     0.168326218
#> [254,] -0.681221460     0.35809420     0.246762496
#> [255,] -0.969323690     0.31633419     0.024290389
#> [256,]  0.747291635     0.21831710    -0.072052440
#> [257,]  2.286075721     0.56226382     0.436433411
#> [258,] -0.209305975     0.40809332     0.254304133
#> [259,] -0.467449626     0.37591323    -0.155927526
#> [260,] -0.709511834     0.15594628     0.353373635
#> [261,] -1.633951704     0.85536425     0.051912685
#> [262,] -0.159453157     0.33005236     0.133663055
#> [263,] -0.589173191     0.46976381     0.307622249
#> [264,] -2.229117250     0.40301200     0.531863307
#> [265,]  0.149127647     0.20848267     0.211275280
#> [266,]  2.678860453     0.62579966     0.063730581
#> [267,] -0.616148539     0.19140780     0.721708863
#> [268,] -0.576182496     0.10928557     0.462787329
#> [269,]  0.201677716     0.56344701     0.034574983
#> [270,] -1.418394871     0.26139907     0.550793644
#> [271,] -0.747506216     0.08344598     0.551543042
#> [272,]  1.301977650     0.39304636     0.114175718
#> [273,] -0.022718621     0.17637147    -0.139247685
#> [274,] -0.042943981     0.33945682     0.463060361
#> [275,]  0.086120738     0.22567807     0.383097307
#> [276,]  2.122679753     0.34848126     0.261697060
#> [277,]  0.213831475     0.35254659     0.154755463
#> [278,] -0.071937038     0.53100479     0.152522619
#> [279,]  0.526000175     0.23742428     0.031503018
#> [280,]  1.227055414     0.56841082     0.387413208
#> [281,]  0.676268292     0.56954795     0.154448950
#> [282,] -0.269656266     0.43448678     0.302777067
#> [283,] -1.504221752     0.09844186     0.258158277
#> [284,]  0.370985097     0.52826014     0.137926239
#> [285,]  0.508810190     0.56672151     0.119358231
#> [286,] -1.176381436     0.36096677     0.104305351
#> [287,] -0.423072260     0.23050482     0.484201045
#> [288,]  1.097346573     0.14402093     0.245494295
#> [289,]  0.175698446     0.45193791    -0.052613907
#> [290,]  1.119665430     0.14391526     0.647382102
#> [291,]  0.701372797     0.26509049     0.100830725
#> [292,]  0.618090185     0.29442852     0.479863023
#> [293,] -1.295403111     0.22071879     0.034998363
#> [294,]  0.066576995     0.35410953    -0.159708601
#> [295,]  1.012707570     0.29694148     0.358829087
#> [296,] -0.926596793     0.07931723     0.256243447
#> [297,]  0.740981448     0.17797805     0.280022973
#> [298,]  0.324627610     0.29501512     0.520477763
#> [299,]  0.533896465     0.18407882     0.434900503
#> [300,]  0.441001437     0.64930083     0.325124390
#> 
#> $acov_order
#> [1] 0
#> 
#> $acor_order
#> [1] 1
#> 
#> $N
#> [1] 300
#> 
#> $S
#> [1] 50
#> 
#> $R
#> [1] 1000
#>