The TOJ bias-corrected estimation of the moments

The `tojmoment()` function enables to implement the TOJ bias-corrected estimation of the moments of the heterogeneous mean, the heterogeneous autocovariance, and the heterogeneous autocorrelation. The method is developed by Okui and Yanagi (2019). For more details, see the package vignette with `vignette("panelhetero")`.

Usage

tojmoment(data, acov_order = 0, acor_order = 1, R = 1000)

Arguments

data

A matrix of panel data. Each row corresponds to individual time series.

acov_order

A non-negative integer of the order of autocovariance. Default is 0.

acor_order

A positive integer of the order of autocorrelation Default is 1.

R

A positive integer of the number of bootstrap repetitions. Default is 1000.

Value

A list that contains the following elements.

estimate

A vector of the parameter estimates

se

A vector of the standard errors

ci

A matrix of the 95 percent confidence intervals

quantity

A matrix of the estimated heterogeneous quantities

acov_order

The order of autocovariance

acor_order

The order of autocovariance

N

The number of cross-sectional units

S

The length of time series

R

The number of bootstrap repetitions

References

Okui, R. and Yanagi, T., 2019. Panel data analysis with heterogeneous dynamics. Journal of Econometrics, 212(2), pp.451-475.

Examples

data <- panelhetero::simulation(N = 300, S = 50)
panelhetero::tojmoment(data = data)
#> $estimate
#>         E(mean)         E(acov)         E(acor)       var(mean)       var(acov) 
#>     -0.08700719      0.36182451      0.28495515      1.04907084      0.04023036 
#>       var(acor) cor(mean, acov) cor(mean, acor) cor(acov, acor) 
#>      0.02552158     -0.03496879      0.02051687     -0.57794775 
#> 
#> $se
#>         E(mean)         E(acov)         E(acor)       var(mean)       var(acov) 
#>     0.058242085     0.012990547     0.016198992     0.091578631     0.008034246 
#>       var(acor) cor(mean, acov) cor(mean, acor) cor(acov, acor) 
#>     0.007741548     0.059334044     0.084992141     0.088033459 
#> 
#> $ci
#>                 95% CI lower 95% CI upper
#> E(mean)         -0.195112405   0.03554435
#> E(acov)          0.336619250   0.38735852
#> E(acor)          0.254579575   0.31664192
#> var(mean)        0.872311572   1.22382965
#> var(acov)        0.026760427   0.05737743
#> var(acor)        0.009714733   0.03989370
#> cor(mean, acov) -0.151934270   0.08714370
#> cor(mean, acor) -0.139199909   0.18902727
#> cor(acov, acor) -0.755785257  -0.39548891
#> 
#> $quantity
#>                mean autocovariance autocorrelation
#>   [1,]  0.219332190     0.99289073     0.079566574
#>   [2,] -0.248865456     0.20046756     0.498685485
#>   [3,]  2.011858854     0.41643777     0.309906996
#>   [4,]  1.366532151     0.45012870     0.015491766
#>   [5,] -2.209998481     0.14250720     0.489092737
#>   [6,]  1.926656355     0.33473297     0.281859178
#>   [7,]  0.151014954     0.34247503     0.412841905
#>   [8,]  0.932420940     0.33061914     0.488545747
#>   [9,]  0.981551558     0.50764532     0.195425658
#>  [10,] -0.279679757     0.31486451     0.188247200
#>  [11,]  1.050250193     0.39710147     0.449230319
#>  [12,]  1.478725106     0.21007320     0.025725773
#>  [13,]  0.661047883     0.52438435     0.374665145
#>  [14,] -0.681299200     0.19897937    -0.184988593
#>  [15,] -0.568324984     0.69988069     0.228897834
#>  [16,]  0.499250703     0.32644109    -0.033460779
#>  [17,]  0.088731084     0.18107860     0.709086062
#>  [18,] -1.301848082     0.10685522     0.581727556
#>  [19,] -0.445585268     0.36391211     0.258581323
#>  [20,]  0.238722231     0.29120628     0.358094478
#>  [21,]  0.560754771     0.24286522     0.454915546
#>  [22,] -1.224561356     0.93278235     0.122739380
#>  [23,] -1.070792903     0.22794797     0.421798185
#>  [24,] -0.829443706     0.23613387     0.362035785
#>  [25,] -0.188137704     0.16316132     0.636541899
#>  [26,] -0.436924548     0.29663397     0.352352001
#>  [27,] -0.902663735     0.33430443     0.336777115
#>  [28,]  0.890365868     0.30918029     0.266471880
#>  [29,]  0.665289797     0.30144398     0.166078134
#>  [30,] -0.791217247     0.45550259     0.046073682
#>  [31,]  1.540207494     0.32487508     0.340135700
#>  [32,]  0.585363905     0.52117845     0.500480893
#>  [33,]  0.480876122     0.32068974    -0.085830411
#>  [34,] -1.749992196     0.43394005     0.153531774
#>  [35,] -1.375864426     0.61614544     0.085749816
#>  [36,] -0.990292829     0.24639595     0.037480721
#>  [37,]  0.081908777     0.42054569     0.198982230
#>  [38,] -1.487242577     0.40887571     0.625683340
#>  [39,] -0.681748900     0.54129932     0.256717072
#>  [40,]  0.320491675     0.39108335     0.714539395
#>  [41,]  0.842752714     0.81511857     0.033964407
#>  [42,] -1.190270168     0.14161420     0.473295092
#>  [43,] -2.714667784     0.33861880     0.411521680
#>  [44,]  0.052846327     0.18016859     0.435956443
#>  [45,]  1.778721787     0.39810618     0.424189295
#>  [46,] -2.093511474     0.18505880     0.398398829
#>  [47,]  0.018498668     0.15791142     0.495877566
#>  [48,] -2.164281852     0.08489475     0.315151468
#>  [49,]  0.851496270     0.73923761     0.274435029
#>  [50,] -0.179031003     0.88009705     0.105277142
#>  [51,] -1.344846598     0.31031723     0.001475992
#>  [52,] -0.926317746     0.43553959     0.083382726
#>  [53,] -0.974677977     0.44521451     0.129130513
#>  [54,]  0.098009360     0.20071776     0.461372755
#>  [55,]  0.615362820     0.15614964     0.603902692
#>  [56,] -1.177363184     0.27483929     0.490527533
#>  [57,]  2.119216435     0.12148705     0.760054981
#>  [58,] -0.925089551     0.20235463     0.311906787
#>  [59,]  0.779879473     0.06518019     0.320759715
#>  [60,] -1.124827812     0.43537212     0.495710260
#>  [61,] -0.037194570     0.25051146     0.155658719
#>  [62,] -0.597277805     0.67738250    -0.043409545
#>  [63,] -0.504053004     0.47959270     0.286910965
#>  [64,] -0.654942365     0.09744588    -0.016161075
#>  [65,] -1.399039395     0.23678321     0.638313491
#>  [66,] -0.257108939     0.34619971     0.402555067
#>  [67,]  2.793119557     0.64921064    -0.063499547
#>  [68,]  0.390971951     0.10934582     0.435379503
#>  [69,] -0.888129520     0.24637520     0.185288806
#>  [70,] -0.439451506     0.78778212     0.342261395
#>  [71,]  0.512407874     0.26127910     0.293401566
#>  [72,] -0.332995885     0.68030439     0.352141478
#>  [73,]  0.295038411     0.16567571     0.515512431
#>  [74,]  1.603125410     0.11883556     0.133312654
#>  [75,]  0.109097580     0.18147942     0.400389794
#>  [76,]  0.799414415     0.07741576     0.405840653
#>  [77,] -0.151918591     0.22512645     0.031696485
#>  [78,] -0.458812482     0.14922388     0.340498703
#>  [79,] -0.125388183     0.57758907     0.247601896
#>  [80,]  0.538389418     0.23695323     0.215343830
#>  [81,]  0.755192714     0.21984659     0.273027145
#>  [82,]  0.874225959     0.45288728    -0.295732118
#>  [83,] -0.392413828     0.33216454    -0.019397078
#>  [84,] -0.225561539     0.07600991     0.552635002
#>  [85,]  0.052210719     0.32632513     0.330490490
#>  [86,]  1.095748074     0.48101608     0.256547453
#>  [87,] -2.222477644     0.25297717     0.248971296
#>  [88,]  1.032375053     0.18944758     0.178785617
#>  [89,]  1.219702291     0.37401614     0.262637445
#>  [90,]  0.548201495     0.23417206     0.513241446
#>  [91,] -0.332754148     0.38223731     0.244055771
#>  [92,] -0.009049570     0.27935006     0.274281259
#>  [93,]  0.263804132     0.41507267     0.074065176
#>  [94,] -0.386161620     1.02104731    -0.107297270
#>  [95,] -0.286866941     0.43479390     0.204076341
#>  [96,] -1.362048759     0.91503591    -0.158839758
#>  [97,]  0.043271293     0.19392576     0.393450460
#>  [98,]  0.326396469     0.46558172     0.255592678
#>  [99,] -0.528684691     0.22115577     0.655281921
#> [100,] -1.701001752     0.25094932     0.046172696
#> [101,]  1.040828398     0.18226721     0.436969313
#> [102,] -0.179572535     0.22845679     0.147310313
#> [103,]  0.528321493     0.62609533     0.395736662
#> [104,] -1.297392300     0.31018779     0.342357266
#> [105,]  0.314347940     0.51201928     0.639333458
#> [106,] -0.184248684     0.31233303     0.053138952
#> [107,] -1.782820203     0.03826448     0.555724290
#> [108,] -0.053830459     0.49143572     0.218001515
#> [109,]  0.383297705     0.17508623     0.528240866
#> [110,]  0.634137843     0.11882638     0.401401029
#> [111,] -0.073695760     0.57523758     0.208412285
#> [112,] -1.724603581     0.38333511     0.128590746
#> [113,]  0.175768329     0.19865714    -0.054602897
#> [114,]  0.777298664     0.15723943     0.286929892
#> [115,]  0.721884062     0.39456436    -0.004543616
#> [116,]  1.068921403     0.46396425     0.016420184
#> [117,]  0.108586216     0.21291944     0.239264192
#> [118,] -1.186898889     0.35239684    -0.142583760
#> [119,]  0.546851999     0.19688820     0.250615299
#> [120,] -0.063600232     0.50320535    -0.183427923
#> [121,] -1.514317541     0.47388202    -0.121548365
#> [122,]  0.630052805     0.12924453     0.161379090
#> [123,] -2.362915832     0.70917804     0.038978459
#> [124,]  0.191104136     0.06457047     0.530600190
#> [125,] -0.737057050     0.32519296     0.168179901
#> [126,]  0.187497704     0.40623863     0.053872961
#> [127,]  0.859342970     0.09698865     0.650218588
#> [128,] -0.528342424     0.28293469    -0.034278662
#> [129,] -1.679853591     0.16107794     0.248876005
#> [130,] -2.402665932     0.33544969     0.565231500
#> [131,]  0.784445488     0.19753331     0.823643646
#> [132,] -2.210235975     0.49086705    -0.014632923
#> [133,] -0.754328490     0.63596965    -0.074543060
#> [134,]  1.282393130     0.12301173     0.530766562
#> [135,] -0.221564915     0.39702950     0.297636010
#> [136,] -1.063583402     0.19575965     0.173502163
#> [137,] -0.034379764     0.10089274     0.542507578
#> [138,]  1.721492327     0.18857911     0.396549309
#> [139,]  0.832895439     0.48364304     0.580264966
#> [140,]  0.883929374     0.36527365    -0.114833246
#> [141,] -0.086455534     0.53655923     0.113751724
#> [142,] -0.763621743     0.15009640     0.592164820
#> [143,]  0.417082648     0.12541092     0.381629252
#> [144,]  0.169440506     0.19133624     0.279043310
#> [145,] -0.875761225     0.73932560     0.123212161
#> [146,] -0.993533111     0.23735411     0.113193518
#> [147,]  0.305700712     0.11500019     0.249636018
#> [148,] -0.541248629     0.15543035     0.350397207
#> [149,] -0.815370589     0.40912092     0.170383514
#> [150,] -2.010273922     0.35263452     0.341905094
#> [151,]  0.295005656     0.40959918     0.109975371
#> [152,]  2.246270294     0.24251631     0.316027422
#> [153,] -0.975224646     0.30050475     0.188996209
#> [154,]  0.886079167     0.19300008     0.509790395
#> [155,] -1.169841472     0.21657645     0.342870197
#> [156,] -0.101520763     0.49009978    -0.004523505
#> [157,] -0.586720942     0.28485786     0.402913941
#> [158,]  0.778959665     0.23001921     0.518019449
#> [159,]  0.160468187     0.47896728     0.032264667
#> [160,] -1.004981941     0.08783161     0.403333912
#> [161,] -0.343494409     0.18119029     0.351338012
#> [162,] -2.349681089     0.16354633     0.215266197
#> [163,] -0.116036347     0.29968475     0.179906107
#> [164,]  0.722870089     0.18834035     0.397407739
#> [165,] -0.506255135     0.70765307     0.207035612
#> [166,]  1.554149620     0.26108786     0.091800136
#> [167,]  0.566125214     0.55342291     0.260130815
#> [168,]  0.342685136     0.57275117     0.171112742
#> [169,]  0.963009077     0.45415652     0.143338443
#> [170,] -1.043706624     0.64090959     0.136972267
#> [171,] -1.161534193     0.21977422     0.312123585
#> [172,]  0.391727454     0.28371097     0.079494501
#> [173,] -1.079946519     0.28635515     0.215247272
#> [174,] -1.425216511     0.16193171     0.200620254
#> [175,] -0.490134784     0.23280745     0.374086743
#> [176,] -0.837270767     0.31434576     0.323027159
#> [177,]  0.051659715     0.32063226     0.237845955
#> [178,] -1.227902256     0.10625974     0.404624403
#> [179,] -0.043001000     0.18231387     0.093526604
#> [180,]  0.893163873     0.06694330     0.302795704
#> [181,] -0.850865983     0.51752190    -0.158434698
#> [182,] -1.242645034     0.50461553     0.018754235
#> [183,] -0.563871783     0.29125978     0.398573411
#> [184,]  0.038567968     0.25709258     0.541983410
#> [185,]  0.292552416     0.21619690     0.233297888
#> [186,]  0.927669510     0.16241096     0.640734691
#> [187,] -0.394574423     0.24408695     0.276819082
#> [188,] -0.073400421     0.11526775     0.563959190
#> [189,] -0.108145539     0.41839150    -0.003198343
#> [190,]  0.315023233     0.53729639     0.421368944
#> [191,] -0.112324108     0.21565610     0.899957608
#> [192,]  0.922601566     0.37404778     0.365789939
#> [193,]  3.491501425     0.17439675     0.245951705
#> [194,] -0.109917420     1.27320106     0.316432091
#> [195,] -1.151206307     0.22785399     0.226225313
#> [196,] -0.411169656     0.43853259     0.161071327
#> [197,]  0.727502827     0.18395906     0.101441727
#> [198,]  0.462816766     0.52939044     0.241442373
#> [199,] -0.553116970     0.47136697    -0.147359582
#> [200,]  0.819579729     0.39752493     0.456881106
#> [201,]  0.557657598     0.33635939     0.173932063
#> [202,] -0.656671531     0.22196072     0.610409114
#> [203,] -0.934642823     0.26396593     0.397337616
#> [204,] -0.611573208     0.95555728     0.179376926
#> [205,]  0.284139028     0.64487225     0.029299328
#> [206,]  1.017629994     0.36553495     0.107306593
#> [207,] -0.243315312     0.32436847     0.362604708
#> [208,] -0.465120175     0.17359828     0.401923222
#> [209,] -0.865005075     0.29671804     0.253361071
#> [210,]  2.288558766     0.19990356     0.321181823
#> [211,]  1.047014751     0.78056147     0.021740767
#> [212,] -0.449347140     0.28891895     0.081014291
#> [213,]  0.462279973     0.27087705     0.639244124
#> [214,] -0.088222884     0.30650776     0.089045784
#> [215,] -1.787152072     0.40581769     0.052102588
#> [216,]  0.555463322     0.23482543    -0.105422341
#> [217,] -2.111656927     0.13076891     0.115356239
#> [218,] -1.758366045     0.11200710     0.476257109
#> [219,]  1.998266278     0.52028256     0.064722795
#> [220,] -0.908579001     0.48398312    -0.027267564
#> [221,]  0.451958663     0.23728525     0.428245211
#> [222,] -0.346691972     0.41317790     0.220481490
#> [223,]  0.471825906     0.56222939     0.083435128
#> [224,]  0.332638894     0.21459957     0.315784541
#> [225,]  0.879096384     0.26166697     0.159201546
#> [226,] -0.998755232     0.16390514     0.469024977
#> [227,]  0.326136682     0.76906757     0.281670241
#> [228,]  0.192258879     0.30375875     0.096331994
#> [229,]  0.414104781     0.38594636    -0.061649752
#> [230,]  1.176692121     0.39732629     0.167978779
#> [231,] -2.316525567     0.25855842     0.665876752
#> [232,] -1.130290621     0.28952720     0.305697706
#> [233,]  0.745170940     0.20816115     0.172885813
#> [234,] -1.198845701     0.25961638     0.132521806
#> [235,]  1.490330738     0.39205259     0.079597788
#> [236,]  0.228218370     0.25933699     0.303852285
#> [237,] -0.118451825     0.21833486     0.643458944
#> [238,] -0.214661149     0.11941939     0.663683025
#> [239,] -0.746471916     0.09508894     0.728475603
#> [240,] -0.106514455     0.40510267     0.295228220
#> [241,]  0.065398087     0.45666019    -0.021591480
#> [242,]  0.977987124     0.09724505     0.385094276
#> [243,] -0.560426136     0.85033598     0.365208042
#> [244,] -0.683226128     0.23210525     0.052072202
#> [245,]  0.225133765     0.14030635     0.598891525
#> [246,] -0.709087150     0.56679526     0.021260948
#> [247,] -0.356494384     0.74445407     0.261493098
#> [248,] -0.177511188     0.52242468     0.178798696
#> [249,] -1.000178894     0.48758211     0.160306003
#> [250,]  2.511615882     0.44983784     0.172596332
#> [251,] -1.299778015     0.33864412     0.141829155
#> [252,] -0.257809236     0.30514311     0.281421787
#> [253,] -0.959359889     0.24483072     0.201944695
#> [254,] -0.112835157     0.57251584     0.177251473
#> [255,]  0.493562049     0.20550071     0.052526342
#> [256,] -0.965639474     0.11870158     0.475965239
#> [257,] -1.681109931     0.30468023    -0.116348854
#> [258,]  0.171487349     0.15881948     0.187964957
#> [259,] -0.404192528     0.24797537     0.368233938
#> [260,]  0.982356717     0.19143347     0.337823613
#> [261,]  1.301819921     0.48122207     0.360454802
#> [262,] -1.160276756     0.59279433    -0.032096974
#> [263,]  1.557564674     0.38888862    -0.178807484
#> [264,]  0.796280005     0.65315570    -0.116272944
#> [265,] -0.701818410     0.80680322     0.108913378
#> [266,] -0.940187679     0.46750356     0.154131004
#> [267,] -1.357403798     0.81476303     0.218066445
#> [268,]  2.735312496     0.59094098     0.256973553
#> [269,] -0.419233273     0.34425999    -0.275388007
#> [270,]  0.739160574     0.28877277     0.098988517
#> [271,]  1.675004985     0.18298032     0.413148682
#> [272,] -1.207671789     0.51454971     0.122527154
#> [273,]  0.007166133     0.10624051     0.277802398
#> [274,] -0.780330447     1.32353755     0.220096311
#> [275,]  1.712185464     0.08809252     0.608200920
#> [276,] -0.001835135     0.27881721     0.485288761
#> [277,] -2.052177362     0.29415961     0.157161474
#> [278,]  0.539104847     0.49251106     0.123111329
#> [279,]  0.283962498     0.23990220     0.145572236
#> [280,]  0.464453363     0.64490291     0.739043638
#> [281,]  0.875405413     0.08861441     0.506546341
#> [282,] -0.977338454     0.62741986     0.234547223
#> [283,]  0.177483675     0.14479916     0.170141185
#> [284,]  0.813553476     0.54550901     0.071665778
#> [285,] -1.804564278     0.20090144     0.520263378
#> [286,] -0.049813238     0.36653388     0.243956796
#> [287,]  0.931284390     0.75221703     0.333977252
#> [288,]  0.319993199     0.42323198     0.178293410
#> [289,]  1.472530145     0.50655709     0.246479353
#> [290,] -0.130467925     0.12247499     0.638513954
#> [291,] -0.363825494     0.62675589     0.351608576
#> [292,] -0.831891654     0.22879349     0.595645458
#> [293,] -1.579175921     0.23189956     0.131593391
#> [294,]  0.637510498     0.19352894     0.070529731
#> [295,]  1.496447377     0.51035843     0.283224443
#> [296,] -0.450955296     0.43125946     0.423665446
#> [297,] -0.934790585     0.38071812     0.096658621
#> [298,]  0.667908741     0.51245713     0.189684597
#> [299,]  0.094554089     0.11386013     0.420114440
#> [300,] -0.583344728     0.45445311     0.318993894
#> 
#> $acov_order
#> [1] 0
#> 
#> $acor_order
#> [1] 1
#> 
#> $N
#> [1] 300
#> 
#> $S
#> [1] 50
#> 
#> $R
#> [1] 1000
#>