The TOJ bias-corrected estimation of the moments

The `tojmoment()` function enables to implement the TOJ bias-corrected estimation of the moments of the heterogeneous mean, the heterogeneous autocovariance, and the heterogeneous autocorrelation. The method is developed by Okui and Yanagi (2019). For more details, see the package vignette with `vignette("panelhetero")`.

Usage

tojmoment(data, acov_order = 0, acor_order = 1, R = 1000)

Arguments

data

A matrix of panel data. Each row corresponds to individual time series.

acov_order

A non-negative integer of the order of autocovariance. Default is 0.

acor_order

A positive integer of the order of autocorrelation Default is 1.

R

A positive integer of the number of bootstrap repetitions. Default is 1000.

Value

A list that contains the following elements.

estimate

A vector of the parameter estimates

se

A vector of the standard errors

ci

A matrix of the 95 percent confidence intervals

quantity

A matrix of the estimated heterogeneous quantities

acov_order

The order of autocovariance

acor_order

The order of autocovariance

N

The number of cross-sectional units

S

The length of time series

R

The number of bootstrap repetitions

References

Okui, R. and Yanagi, T., 2019. Panel data analysis with heterogeneous dynamics. Journal of Econometrics, 212(2), pp.451-475.

Examples

data <- panelhetero::simulation(N = 300, S = 50)
panelhetero::tojmoment(data = data)
#> $estimate
#>         E(mean)         E(acov)         E(acor)       var(mean)       var(acov) 
#>      0.01419760      0.35053197      0.26629650      0.88534533      0.03278017 
#>       var(acor) cor(mean, acov) cor(mean, acor) cor(acov, acor) 
#>      0.01558230      0.05481255     -0.04340559     -0.46481522 
#> 
#> $se
#>         E(mean)         E(acov)         E(acor)       var(mean)       var(acov) 
#>     0.053104833     0.012522822     0.017271266     0.067946555     0.006461312 
#>       var(acor) cor(mean, acov) cor(mean, acor) cor(acov, acor) 
#>     0.007205798     0.071516058     0.093378021     0.104073418 
#> 
#> $ci
#>                 95% CI lower 95% CI upper
#> E(mean)         -0.081742193   0.11991564
#> E(acov)          0.325162832   0.37466046
#> E(acor)          0.230420612   0.29820852
#> var(mean)        0.755823564   1.01924534
#> var(acov)        0.020730771   0.04514670
#> var(acor)        0.001556578   0.02965636
#> cor(mean, acov) -0.091805035   0.19414210
#> cor(mean, acor) -0.231179834   0.13748321
#> cor(acov, acor) -0.675544697  -0.27027752
#> 
#> $quantity
#>                mean autocovariance autocorrelation
#>   [1,] -0.948988933     0.52520790     0.392006654
#>   [2,] -1.099806504     0.14305896     0.277908983
#>   [3,] -0.287040088     0.39069910     0.171940669
#>   [4,]  1.317953831     0.26341469    -0.129162722
#>   [5,] -0.524396036     0.26998981     0.172405487
#>   [6,]  0.146022981     0.27490946     0.156006208
#>   [7,]  0.287127515     0.33499206     0.380664562
#>   [8,]  1.457737676     0.99055509    -0.015157269
#>   [9,]  1.011291395     0.16831915     0.291522673
#>  [10,]  0.649885657     0.32692558     0.005988297
#>  [11,] -0.875714038     0.24654780     0.476975612
#>  [12,] -0.210114123     0.38142638     0.025418290
#>  [13,] -2.272722738     0.11841443     0.639582303
#>  [14,]  0.841495779     0.34834082     0.197824344
#>  [15,]  1.045816801     0.23803725    -0.071661462
#>  [16,]  1.369080817     0.20562337     0.252399751
#>  [17,]  0.658592968     1.04579719     0.054743824
#>  [18,] -0.535964072     0.19611925    -0.040272576
#>  [19,] -0.216926852     0.24731864     0.157918281
#>  [20,]  0.273799445     0.41456468     0.299953286
#>  [21,] -0.464845449     0.22672288     0.283951825
#>  [22,] -0.099667009     0.46224880     0.086363410
#>  [23,] -0.866447576     0.49491075     0.205070291
#>  [24,]  0.093774264     0.45212047    -0.248164224
#>  [25,] -0.612432765     0.27311014     0.037735858
#>  [26,]  0.492102751     0.36265749     0.100856908
#>  [27,] -0.573758484     0.38404866     0.050147487
#>  [28,]  1.148239234     0.11284478     0.453551630
#>  [29,]  0.179876876     0.51358710     0.452873973
#>  [30,] -1.065545588     0.13930410     0.326697888
#>  [31,] -1.758204003     0.54718858     0.294543806
#>  [32,]  0.009599268     0.44933273     0.332533036
#>  [33,]  0.310852854     0.65252331     0.329294186
#>  [34,]  1.841190747     0.42028383     0.157212728
#>  [35,]  0.440498825     0.14788481     0.339669161
#>  [36,] -0.857412232     0.31125988     0.243173260
#>  [37,] -0.623842068     0.44943587     0.175070394
#>  [38,] -0.029473700     0.09540119     0.534374599
#>  [39,] -0.573262631     0.30807631     0.111761328
#>  [40,] -0.678182097     0.30487893     0.241773154
#>  [41,]  0.983238189     0.26216258     0.435373548
#>  [42,]  1.347124579     0.30514299     0.227935732
#>  [43,]  0.219206467     0.16482715     0.443100167
#>  [44,] -0.219659291     0.24044917     0.489057794
#>  [45,] -0.435116715     0.32961701     0.387400498
#>  [46,] -1.273353188     0.14922881     0.173708708
#>  [47,]  2.207985257     0.11915848     0.536360351
#>  [48,]  0.277577885     0.21803721     0.287589397
#>  [49,]  0.372053928     0.51386090     0.157490821
#>  [50,]  0.718625297     0.31995878     0.089768816
#>  [51,] -0.660092194     0.18440146     0.348599340
#>  [52,]  1.171604112     0.31579882    -0.066987920
#>  [53,] -0.937086789     0.18722974     0.418396169
#>  [54,] -0.458849713     0.13959524     0.388540967
#>  [55,]  0.637037765     0.47761170     0.139426463
#>  [56,]  0.305928651     0.56407358     0.492878586
#>  [57,] -1.503028221     0.12423752     0.275085291
#>  [58,] -0.466684686     0.32107257    -0.156473712
#>  [59,]  0.058123378     0.20466123     0.364660250
#>  [60,] -0.511840760     0.22423225     0.361098620
#>  [61,] -0.302211011     0.36705839     0.347875992
#>  [62,]  0.914513828     0.38046699    -0.037609648
#>  [63,] -0.074587060     0.13553544     0.666948588
#>  [64,] -1.034139926     0.20901636     0.298371178
#>  [65,] -0.516954177     0.40095392    -0.074099144
#>  [66,] -0.637895278     0.41582645     0.238248718
#>  [67,]  0.125782648     0.44541755    -0.257086926
#>  [68,]  1.488713216     0.18188233    -0.052890432
#>  [69,]  0.506030356     0.14221227     0.375814093
#>  [70,]  1.294061300     0.20048005     0.473057622
#>  [71,] -0.482444195     0.22721331     0.014943955
#>  [72,]  0.268376929     0.44128015    -0.143425744
#>  [73,] -0.202273839     0.04280411     0.328820533
#>  [74,]  0.757803645     0.47250293     0.381639762
#>  [75,] -0.589953392     0.88043069     0.044026654
#>  [76,] -1.104247037     0.06563963     0.717635752
#>  [77,]  2.029655074     0.40970998     0.330313034
#>  [78,]  0.412118975     0.11506235     0.386460414
#>  [79,]  0.306911440     0.22666725     0.204124368
#>  [80,]  0.576264253     0.20853141    -0.009975075
#>  [81,] -0.563191899     0.29890583     0.353091750
#>  [82,]  0.091957046     0.26209551     0.168180317
#>  [83,] -0.140553037     0.19629898     0.343553212
#>  [84,]  1.738489779     0.25850666     0.223375324
#>  [85,]  0.335161745     0.56868811     0.090377767
#>  [86,] -1.728020079     0.15562137     0.283082764
#>  [87,] -1.858378301     0.31705946     0.248904061
#>  [88,] -1.392644472     0.07756735     0.576615749
#>  [89,] -0.551521007     0.36609690     0.087535062
#>  [90,] -0.553771070     0.54339428     0.199007047
#>  [91,]  1.581460861     0.34583313     0.281883174
#>  [92,]  0.949782521     0.04979902     0.471867614
#>  [93,] -0.434756126     0.41060889    -0.013213707
#>  [94,]  0.892848825     0.36334075     0.119825815
#>  [95,] -0.520062934     0.55244405     0.161573632
#>  [96,]  0.468437707     0.42973360     0.138122247
#>  [97,]  2.148880876     0.62298113     0.298187375
#>  [98,]  0.985899259     0.11534181     0.357243852
#>  [99,] -1.121434483     0.18037160     0.422793493
#> [100,] -0.583447974     0.57118224    -0.191437839
#> [101,] -0.166094836     0.42259439     0.305569800
#> [102,]  0.148231168     0.34038902     0.157625567
#> [103,]  1.352327899     0.17028177     0.400209111
#> [104,]  1.067831548     0.08050487     0.256652049
#> [105,] -0.117432108     0.48937798     0.458344565
#> [106,] -1.897053089     0.37917669     0.526195213
#> [107,] -1.190159017     0.44861657     0.148741818
#> [108,] -0.914013630     0.22473463     0.366445871
#> [109,] -1.798351731     0.23335184     0.362332908
#> [110,] -1.486542948     0.45605916    -0.074706133
#> [111,]  0.504612990     0.31222475     0.411174287
#> [112,] -1.484490177     0.23405874     0.439049435
#> [113,] -0.631013157     0.18796649     0.613982945
#> [114,]  1.564019878     0.24095692    -0.024870847
#> [115,] -0.533969033     0.36177737     0.348524288
#> [116,]  1.227509813     0.18970743     0.375031329
#> [117,]  1.309110882     0.48401670     0.213758703
#> [118,] -1.412704827     0.10385390     0.256078900
#> [119,] -1.590566689     0.40130096    -0.061129820
#> [120,] -0.260993114     0.16891470     0.117845632
#> [121,]  1.038271998     0.28171146     0.035759823
#> [122,]  1.293490560     0.23188830     0.296486086
#> [123,] -1.888996204     0.30459183    -0.071626156
#> [124,]  0.784401646     0.21022324     0.286733856
#> [125,] -0.994637914     0.12491291     0.410043239
#> [126,]  1.337417609     0.32774582     0.388920229
#> [127,]  1.291457580     0.18748182     0.063563742
#> [128,] -0.610865657     0.29066539     0.227631315
#> [129,]  0.125588818     0.52397155    -0.243494603
#> [130,]  0.184261804     0.23785467     0.315818475
#> [131,] -2.227223785     0.24559736     0.600687074
#> [132,] -1.179466464     0.33795333     0.208365057
#> [133,]  0.096833419     0.73218005     0.373335690
#> [134,] -1.121822533     0.18551852     0.225304689
#> [135,]  0.483796334     0.20023350     0.482309174
#> [136,] -0.606523332     0.95477871     0.370181445
#> [137,] -1.508722645     0.10027650     0.423181712
#> [138,]  1.224047844     0.42133416     0.460611708
#> [139,]  1.273125947     0.19493105     0.161967125
#> [140,] -0.086835747     0.52234267     0.017642220
#> [141,]  0.132186636     0.24874812     0.145698449
#> [142,]  0.622666787     0.32441842     0.273166655
#> [143,] -0.135934166     0.53362719     0.105126210
#> [144,] -0.822911090     0.64162249     0.446164059
#> [145,]  0.354207839     0.23476104     0.337362612
#> [146,] -0.257306607     0.18037701     0.366598042
#> [147,] -0.011092087     0.22606537     0.275193895
#> [148,] -0.012187287     0.31678736     0.415040824
#> [149,]  0.134586347     0.25650213     0.341166307
#> [150,]  0.277314160     0.34195219     0.458974714
#> [151,] -0.688706110     0.32806633     0.507936965
#> [152,] -0.930814004     0.39805588     0.109334416
#> [153,] -0.589572959     0.81216776    -0.237936785
#> [154,]  0.103884890     0.14278383     0.170096764
#> [155,]  1.579660340     0.38199320     0.236120109
#> [156,]  0.350549013     0.31863723     0.209342919
#> [157,]  2.087710853     0.11679404     0.490465660
#> [158,]  1.076551756     0.18809317     0.253174708
#> [159,]  0.882485501     0.47789861     0.129718625
#> [160,]  1.383473775     0.18137653     0.164111289
#> [161,]  0.135752234     0.36921992     0.003576812
#> [162,]  1.686378148     0.12610942     0.311254288
#> [163,] -0.405506470     0.59328290     0.132512482
#> [164,]  1.216045339     0.37565266     0.437911784
#> [165,]  0.298315941     0.21828376     0.385959930
#> [166,] -1.313399614     0.15140275     0.360807643
#> [167,]  1.360085370     0.24928663     0.149510489
#> [168,] -1.214063964     0.18726492     0.201034463
#> [169,]  0.191854010     0.25585128     0.337973939
#> [170,] -1.522970000     0.14215334     0.349431864
#> [171,] -0.191531065     0.77113734     0.026333537
#> [172,] -0.920247019     0.30947462     0.209334227
#> [173,] -0.336067359     0.37651333     0.130000948
#> [174,]  0.590331440     0.16031813     0.437024134
#> [175,]  0.010884586     0.31058496     0.639722889
#> [176,] -0.244116721     0.24877339     0.148031427
#> [177,]  0.398823591     0.54961908     0.236895757
#> [178,]  0.258498104     0.78453103     0.258535014
#> [179,] -0.268447732     0.20618592     0.625560953
#> [180,] -0.030320049     0.20806735     0.143393798
#> [181,] -0.317466894     0.20834322    -0.035476597
#> [182,]  0.155278453     0.46152052    -0.018958166
#> [183,] -0.814125221     0.59389714     0.239965245
#> [184,]  1.466733728     0.32926284     0.181309083
#> [185,]  1.509419595     0.11029948     0.414309355
#> [186,]  0.750041949     0.58385426     0.135703322
#> [187,] -0.199653012     0.12757090     0.455205897
#> [188,]  0.092323447     0.30801057    -0.068864242
#> [189,]  0.944763838     0.23897203     0.397249291
#> [190,] -1.378846757     0.23994334     0.055030408
#> [191,]  0.161750478     0.73428529     0.131050439
#> [192,] -1.119299957     0.14609106     0.336577598
#> [193,]  1.001743501     0.20286712    -0.153155769
#> [194,] -0.579133118     0.62223022    -0.157999072
#> [195,]  0.658906641     0.28494357     0.122491566
#> [196,]  0.875041078     0.37211527     0.227118730
#> [197,] -0.166131405     0.43854419     0.092217956
#> [198,]  0.896170788     0.13089766     0.134261525
#> [199,]  0.844465851     0.28705715     0.181637402
#> [200,] -0.371580090     0.14958882     0.017022874
#> [201,] -0.677447496     0.16129734     0.445818355
#> [202,]  1.623966558     0.32887900     0.434491285
#> [203,]  1.257377484     0.47527343     0.177497528
#> [204,] -0.181242379     0.17077467     0.236352473
#> [205,]  0.313551009     0.48715209     0.006337225
#> [206,] -0.420787179     0.10410573     0.294236180
#> [207,]  1.754229127     1.18549382     0.215112965
#> [208,] -0.062010981     0.45928264     0.089406071
#> [209,]  0.112322000     0.32020642     0.365835249
#> [210,] -0.878145569     0.32779133     0.160026564
#> [211,]  0.053369694     0.14916390     0.477559197
#> [212,] -0.206678998     0.68631867     0.021027691
#> [213,]  0.866778497     0.32950048     0.470906132
#> [214,] -1.149532679     0.29619617     0.238188008
#> [215,] -0.675320572     0.37305721     0.334206347
#> [216,] -0.845636888     0.55678959     0.439038796
#> [217,] -1.343863847     0.23921184     0.391952035
#> [218,] -0.478723920     0.16262291     0.348047300
#> [219,] -0.776911123     0.33477889     0.201145245
#> [220,]  1.027660011     0.29004604    -0.200053935
#> [221,] -0.855678681     0.18424277     0.252525301
#> [222,]  0.348352421     0.51794705     0.066542998
#> [223,] -0.553988478     0.32571899     0.654963814
#> [224,] -0.610615592     0.52219897     0.325239147
#> [225,] -1.136755215     0.53332935     0.338657026
#> [226,] -0.965261604     0.43494582     0.054684085
#> [227,]  0.255603769     0.77373202     0.363251888
#> [228,]  1.016931693     0.51134107    -0.040454535
#> [229,]  1.838702368     0.39836752     0.177224107
#> [230,] -0.008693336     0.10158042     0.155079767
#> [231,] -1.025006224     0.44521293     0.257634872
#> [232,]  1.048687786     0.25471614     0.112089317
#> [233,]  1.628969971     0.60553625     0.176247079
#> [234,] -0.420341534     0.33538677     0.055054518
#> [235,]  0.470444670     0.55998970     0.260446111
#> [236,] -0.446578771     0.29647918     0.200898653
#> [237,]  1.249011637     0.21463182     0.023866784
#> [238,]  0.484692807     0.47669030     0.360584501
#> [239,] -0.234948981     0.47546005     0.217587090
#> [240,]  0.175401160     0.03881722     0.610714567
#> [241,]  0.776929138     0.55724832     0.408180194
#> [242,]  1.790720215     0.45706258     0.092907365
#> [243,] -0.145792006     0.56201192     0.429651783
#> [244,] -0.622966603     0.07459995     0.037456475
#> [245,]  0.303228633     0.35054068     0.091435058
#> [246,] -0.455925726     0.37303329     0.020783927
#> [247,] -1.337192277     0.30128604     0.108483598
#> [248,] -1.984998023     0.67728586    -0.174439887
#> [249,]  0.473354559     0.39978321     0.378586481
#> [250,] -0.390954283     0.16559152     0.432860650
#> [251,] -0.676315452     1.03179759     0.095755625
#> [252,] -0.072182002     0.33040852     0.071345724
#> [253,] -0.269178163     0.26005244     0.277017173
#> [254,] -0.354820509     0.05584811     0.283850444
#> [255,] -1.769232227     0.10793970     0.498819229
#> [256,]  1.373443108     0.69200230     0.087926647
#> [257,] -0.641765866     0.17825419     0.361556517
#> [258,]  2.306278557     0.21862507     0.262319827
#> [259,] -0.010381489     0.11191915     0.072640519
#> [260,] -0.909931347     0.20035008     0.300794665
#> [261,]  0.486365973     0.16538758     0.087991096
#> [262,] -0.907540383     0.13541093    -0.030048919
#> [263,]  0.223355800     0.70190739     0.375248179
#> [264,]  0.988630758     0.59703442     0.141650612
#> [265,] -0.306492628     0.24059146     0.349248289
#> [266,]  1.124436479     0.16407889     0.329373888
#> [267,]  0.093163062     0.27317223     0.368630440
#> [268,]  0.069654807     0.23400179     0.203143831
#> [269,]  0.445738487     0.31765850     0.287530686
#> [270,] -1.172101200     0.26027110     0.002175614
#> [271,]  0.587927308     0.22772970     0.401172721
#> [272,]  0.104463511     0.16110953     0.274577610
#> [273,]  0.015071057     0.31335298     0.107808189
#> [274,] -0.456436276     0.59385072     0.342617998
#> [275,] -0.616563726     0.53797038     0.382307609
#> [276,] -0.429098726     0.30724633    -0.001486262
#> [277,] -2.236333136     0.59398400     0.388960467
#> [278,]  2.118571845     0.34326449     0.298842977
#> [279,]  0.013189103     0.20537088     0.538242377
#> [280,] -0.916738120     0.16560531     0.375631955
#> [281,] -0.306325819     0.40701367     0.247393810
#> [282,]  0.163442359     0.19986455     0.775098246
#> [283,] -0.711448177     0.39360661     0.104951378
#> [284,] -0.834933289     0.34031665     0.089208924
#> [285,]  0.791952258     0.27297928     0.133771159
#> [286,]  0.194461730     0.13012055     0.246095900
#> [287,]  0.054142312     0.24848648     0.299336962
#> [288,]  1.437805386     0.25288298     0.375517294
#> [289,] -0.036561022     0.25731733     0.158192259
#> [290,] -0.294304182     0.88704923     0.058119807
#> [291,]  0.567265203     0.26472782     0.325518477
#> [292,] -1.386596358     0.32468474     0.042972465
#> [293,] -1.054596898     0.25364500     0.365864787
#> [294,] -0.046520357     0.37561513    -0.088237168
#> [295,]  1.269921905     0.37387834     0.430612049
#> [296,]  2.196134926     0.42536790     0.222019883
#> [297,]  0.227461510     0.46722021     0.300263470
#> [298,]  0.091670528     0.56565272     0.137791217
#> [299,] -1.005410447     0.82752660     0.342256760
#> [300,] -1.200960534     0.36232168     0.205498110
#> 
#> $acov_order
#> [1] 0
#> 
#> $acor_order
#> [1] 1
#> 
#> $N
#> [1] 300
#> 
#> $S
#> [1] 50
#> 
#> $R
#> [1] 1000
#>